% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_OCcutline.R
\name{add_OCcutline}
\alias{add_OCcutline}
\title{Add Cutline to Plot in Optimal Classification (OC) Analysis}
\usage{
add_OCcutline(cutData, lwd = 2)
}
\arguments{
\item{cutData}{A numeric vector containing the parameters needed to compute the cutline. This typically includes coefficients and midpoint values from the OC analysis.}

\item{lwd}{A numeric value specifying the line width of the cutline. The default value is \code{2}.}
}
\description{
The \code{add_OCcutline} function calculates and adds a cutline to a plot based on the results of an Optimal Classification (OC) analysis. The cutline is determined by the slope and intercept derived from the provided \code{cutData}.
}
\details{
The function calculates the slope and intercept of the cutline based on the values provided in \code{cutData}. If the slope is not defined (i.e., it is \code{NA}), the function will attempt to plot a vertical line at the specified intercept. If the slope is defined, the function calculates the points where the line intersects the unit circle and plots the resulting line segment.

This function is intended for internal use within the package and is primarily designed to assist with visualizing OC analysis results.
}
\keyword{internal}
