% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{ANES1968}
\alias{ANES1968}
\title{ANES 1968 Feeling Thermometers and Voting Data}
\format{
A dataframe with several variables:
\describe{
\item{vote.turnout}{Binary variable indicating whether the respondent reported voting in the 1968 election (1 = voted, 0 = did not vote).}
\item{presidential.vote}{Categorical variable representing the respondent's reported presidential vote choice (1 = Humphrey, 2 = Nixon, 3 = Wallace).}
\item{T}{A matrix of feeling thermometers ranging from 0 ("very cold or unfavorable") to 100 ("very warm or favorable").}
}
}
\source{
American National Election Studies (ANES) 1968 dataset.
}
\description{
The \code{ANES1968} dataset includes feeling thermometers data and information on whether respondents reported voting in the 1968 elections, as well as their reported presidential vote choice.
}
\details{
The feeling thermometers measure respondents' attitudes towards various political figures and entities, with higher values indicating warmer or more favorable feelings. Values of 98, 99, and 100 were recoded in the original ANES file as 97, allowing these values to represent refused or missing responses.

The \code{ANES1968} dataset is a valuable resource for analyzing voter behavior and sentiment during the 1968 U.S. presidential election.
}
\examples{
\dontrun{
data(ANES1968)
summary(ANES1968)
}
}
\keyword{datasets}
