% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{ANES2004}
\alias{ANES2004}
\title{ANES 2004 Issue Scales Data}
\format{
A dataframe with several variables representing different political issues:
\describe{
\item{libcon}{Liberal–Conservative scale, ranging from 1 (left) to 7 (right).}
\item{diplomacy}{Diplomacy–Military Force scale, ranging from 1 (left) to 7 (right).}
\item{iraqwar}{Bush’s Handling of Iraq War scale, ranging from 1 (right) to 4 (left).}
\item{govtspend}{Government Spending/Services scale, ranging from 1 (right) to 7 (left).}
\item{defense}{Defense Spending scale, ranging from 1 (left) to 7 (right).}
\item{bushtaxcuts}{Bush Tax Cuts scale, ranging from 1 (right) to 4 (left).}
\item{healthinsurance}{Government Health Insurance scale, ranging from 1 (left) to 7 (right).}
\item{govtjobs}{Guaranteed Jobs scale, ranging from 1 (left) to 7 (right).}
\item{aidblacks}{Government Aid to Blacks scale, ranging from 1 (left) to 7 (right).}
\item{govtfundsabortion}{Government Abortion Funding scale, ranging from 1 (left) to 4 (right).}
\item{partialbirthabortion}{Partial-Birth Abortion Ban scale, ranging from 1 (right) to 4 (left).}
\item{environmentjobs}{Environment–Jobs scale, ranging from 1 (left) to 7 (right).}
\item{deathpenalty}{Death Penalty scale, ranging from 1 (right) to 4 (left).}
\item{gunregulations}{Gun Regulations scale, ranging from 1 (left) to 5 (right).}
\item{womenrole}{Women’s Role scale, ranging from 1 (left) to 7 (right).}
\item{gaymarriage}{Gay Marriage scale, ranging from 1 (left) to 3 (right).}
}
}
\source{
American National Election Studies (ANES) 2004 dataset.
}
\description{
The \code{ANES2004} dataset includes responses to various issue scales from the 2004 American National Election Studies (ANES). The dataset contains respondents' positions on several key political issues, measured on scales with varying ranges.
}
\details{
The \code{ANES2004} dataset captures respondents' views on a variety of political and social issues during the 2004 U.S. election period. The scales vary in their directionality, with some scales placing liberal positions on the left and others placing conservative positions on the right.

The dataset is useful for analyzing public opinion on key issues and understanding the political landscape during the 2004 election.
}
\examples{
\dontrun{
data(ANES2004)
summary(ANES2004)
}
}
\keyword{datasets}
