\name{asmcjr-package}
\alias{asmcjr-package}
\alias{asmcjr}
\docType{package}
\title{
\packageTitle{asmcjr}
}
\description{
\packageDescription{asmcjr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{asmcjr}
\packageIndices{asmcjr}
The \code{asmcjr} package supports the book - Analyzing Spatial Models of Choice and Judgment with R (2014, CRC Press).  Most of the functions are wrappers and other helper functions to facilitate estimation and examination of the scaling models discussed in the book.
}
\author{
\packageAuthor{asmcjr}

Maintainer: \packageMaintainer{asmcjr}
}
\references{
David Armstrong, Ryan Bakker, Royce Carroll, Christopher Hare, Keith Poole and Howard Rosenthal.  (2014)  Analyzing Spatial Models of Choice and Judgment with R.  CRC Press.
}
\keyword{Measurement}
