% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAM.R
\name{BAM}
\alias{BAM}
\title{Bayesian Aldrich-McKelvey Scaling}
\usage{
BAM(
  data,
  polarity,
  zhatSave = TRUE,
  abSave = FALSE,
  resp.idealpts = FALSE,
  n.sample = 2500,
  ...
)
}
\arguments{
\item{data}{A dataset prepared using the \code{\link{bamPrep}} function.}

\item{polarity}{Integer specifying the column number of the stimulus used to set the polarity.}

\item{zhatSave}{Logical, whether to save the posterior samples of \code{zhat} (default is TRUE).}

\item{abSave}{Logical, whether to save the posterior samples of \code{a} and \code{b} parameters (default is FALSE).}

\item{resp.idealpts}{Logical, whether to compute and save respondent ideal points (default is FALSE).}

\item{n.sample}{Integer, the number of MCMC samples to draw (default is 2500).}

\item{...}{Additional arguments passed to \code{\link[rjags]{jags.model}}.}
}
\value{
A list containing posterior samples and summaries for the estimated parameters.
}
\description{
This function performs Bayesian Aldrich-McKelvey (BAM) scaling using JAGS
to estimate ideal points and optionally save intermediate parameters.
}
\examples{
\donttest{
data(bamdata)
bam.france <- BAM(bamdata, polarity=2, n.adapt=2500, n.sample=5000,
                  zhatSave=TRUE, abSave=TRUE, resp.idealpts=TRUE)
}

}
