% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamPrep.R
\name{bamPrep}
\alias{bamPrep}
\title{Prepare Data for BAM (Bayesian Aldrich-McKelvey) Scaling}
\usage{
bamPrep(x, nmin = 1, missing = NULL, self = 1, midpt = NULL)
}
\arguments{
\item{x}{A numeric matrix or dataframe containing the data to be preprocessed.}

\item{nmin}{Integer, the minimum number of non-missing values required per row (default is 1).}

\item{missing}{A vector of values that should be treated as missing (e.g., \code{c(77, 88, 99)}).}

\item{self}{Integer, the column number representing self-placement (default is 1).}

\item{midpt}{Numeric, the midpoint used for centering the data. If NULL, the midpoint is calculated as the mean of the minimum and maximum values.}
}
\value{
A list with two components: \code{stims} (a matrix of stimuli placements) and \code{self} (a vector of self-placements).
}
\description{
This function preprocesses data for use in Bayesian Aldrich-McKelvey (BAM) scaling. It handles missing values, centers the data, and separates the self-placement from the stimuli placements.
}
\examples{
\dontrun{
# Aldrich-Mckelvey scaling, with cutoff of 5 or more
data(franceEES2009)
bamdata <- bamPrep(franceEES2009, missing=c(77,88,89), self=1)
}
}
