% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.blackbox.R
\name{boot.blackbox}
\alias{boot.blackbox}
\title{Bootstrap Analysis for Blackbox Model}
\usage{
boot.blackbox(
  data,
  missing,
  dims = 1,
  minscale,
  verbose = FALSE,
  posStimulus = 1,
  R = 100
)
}
\arguments{
\item{data}{A dataframe containing the data to be bootstrapped.}

\item{missing}{Vector of missing values.}

\item{dims}{Integer, the number of dimensions to be considered in the analysis.}

\item{minscale}{Minimum scale value.}

\item{verbose}{Logical, whether to print detailed information (default is FALSE).}

\item{posStimulus}{Integer, the position of the stimulus to check for sign correction (default is 1).}

\item{R}{Integer, the number of bootstrap replicates (default is 100).}
}
\value{
An array of bootstrapped results with class \code{"bootbb"}.
}
\description{
This function performs a bootstrap analysis on the Blackbox model.
}
\examples{
\dontrun{
data(issues.sweden)
boot.blackbox(issues.sweden, missing=8, dims=3, minscale=5, verbose=FALSE, posStimulus=13)}
}
