% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.blackbox_transpose.R
\name{boot.blackbox_transpose}
\alias{boot.blackbox_transpose}
\title{Bootstrap Analysis for the Blackbox Transpose Model}
\usage{
boot.blackbox_transpose(
  data,
  missing,
  dims = 1,
  minscale,
  verbose = FALSE,
  posStimulus = 1,
  R = 100
)
}
\arguments{
\item{data}{A matrix containing the input data for the analysis.}

\item{missing}{Numeric vector indicating which values are treated as missing.}

\item{dims}{Integer, the number of dimensions to analyze (default is 1).}

\item{minscale}{Numeric, the minimum scale value used in the analysis.}

\item{verbose}{Logical, whether to print detailed information during execution (default is FALSE).}

\item{posStimulus}{Integer, the index of the stimulus used to check and correct the sign (default is 1).}

\item{R}{Integer, the number of bootstrap replicates (default is 100).}
}
\value{
An array containing the bootstrapped results.
}
\description{
This function performs a bootstrap analysis on the Blackbox Transpose model, generating bootstrap replicates of the ideal points for stimuli.
}
\examples{
\dontrun{
data(issues.sweden)
result <- boot.blackbox_transpose(issues.sweden, missing = 8, 
dims = 2, minscale = 5, verbose = FALSE, posStimulus = 13, R = 100)
}
}
