% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffStims.R
\name{diffStims}
\alias{diffStims}
\title{Compute the Probabilities of Differences Between Stimuli}
\usage{
diffStims(x, stims, digits = 3, ...)
}
\arguments{
\item{x}{A matrix, data frame, or \code{mcmc.list} containing MCMC or resampled values. The rows represent different samples,
and the columns represent different stimuli or variables.}

\item{stims}{A vector of stimuli to compare. This can be a numeric vector of column indices or a character vector of column names.}

\item{digits}{An integer specifying the number of decimal places to display in the output. Default is 3.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A data frame with two columns: \code{Comparison}, which contains the comparison labels (e.g., "Pr(stim2 > stim1)"),
and \code{Probability}, which contains the probabilities formatted to the specified number of digits.
}
\description{
This function calculates the probabilities that one stimulus is greater than another based on MCMC or resampled values.
It can handle input data in the form of a matrix, data frame, or \code{mcmc.list}. The function computes pairwise comparisons
between the specified stimuli.
}
\examples{
\dontrun{
data(bam.france)
diffStims(bam.france$zhat, stims=c(3,4))
}
}
