% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{end.mcmc.list}
\alias{end.mcmc.list}
\title{Get the End Time of an MCMC List}
\usage{
\method{end}{mcmc.list}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mcmc.list}, which is a list of MCMC chains.}

\item{...}{Additional arguments (currently not used).}
}
\value{
The result of calling the \code{end} function on the first element of the MCMC list.
}
\description{
This function returns the end time or the last element of the first chain in an MCMC list.
}
\examples{
\donttest{
end_time <- end.mcmc.list(mcmc_list)
}
}
