% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geweke.ggplot.R
\name{geweke.ggplot}
\alias{geweke.ggplot}
\title{Geweke Diagnostic Plot using ggplot2}
\usage{
geweke.ggplot(x, frac1 = 0.1, frac2 = 0.5, nbins = 20, pvalue = 0.05)
}
\arguments{
\item{x}{An object of class \code{mcmc} or \code{mcmc.list}. This object represents the MCMC chains for which the Geweke diagnostic will be calculated.}

\item{frac1}{A numeric value specifying the fraction of the first segment of the chain to use in the Geweke diagnostic. Defaults to \code{0.1}.}

\item{frac2}{A numeric value specifying the fraction of the last segment of the chain to use in the Geweke diagnostic. Defaults to \code{0.5}.}

\item{nbins}{An integer specifying the number of segments (bins) into which the chain should be divided for the diagnostic calculation. Defaults to \code{20}.}

\item{pvalue}{A numeric value specifying the p-value for the confidence limits in the Z-score plot. Defaults to \code{0.05}.}
}
\value{
A \code{ggplot} object that displays the Geweke diagnostic Z-scores for each segment of the MCMC chains. The plot includes horizontal lines indicating the confidence limits.
}
\description{
The \code{geweke.ggplot} function creates a Geweke diagnostic plot for MCMC chains using \code{ggplot2}. It calculates the Geweke diagnostic for segments of the MCMC chain and visualizes the Z-scores over the first iteration in each segment.
}
\details{
The function splits the MCMC chain into segments, computes the Geweke diagnostic for each segment, and then plots the Z-scores. The Z-scores are compared against the standard normal quantiles, and segments outside the confidence limits indicate potential non-convergence.
}
\examples{
\dontrun{
library(coda)
data(line)
geweke_plot <- geweke.ggplot(line)
print(geweke_plot)
}

}
