% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gray.palette}
\alias{gray.palette}
\title{Generate a Gray Color Palette}
\usage{
gray.palette(n, lower = 0.3, upper = 0.7)
}
\arguments{
\item{n}{Integer, the number of gray shades to generate.}

\item{lower}{Numeric, the lower bound for the gray scale (default is 0.3).}

\item{upper}{Numeric, the upper bound for the gray scale (default is 0.7).}
}
\value{
A vector of gray colors.
}
\description{
This function generates a palette of gray colors.
}
\examples{
gray.palette(5)
}
