% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{issues}
\alias{issues}
\title{Issues Matrix from CDS2000 Dataset}
\format{
A numeric matrix with rows corresponding to observations and columns representing different issues.
}
\source{
Extracted from the \code{CDS2000} dataset.
}
\description{
This object, \code{issues}, is a matrix extracted from the \code{CDS2000} dataset. It contains selected columns that represent various issues.
}
\examples{
\dontrun{
data(issues)
}
}
\keyword{datasets}
