% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{legis_7th_Taiwan}
\alias{legis_7th_Taiwan}
\title{7th Legislative Yuan Roll Call Data from Taiwan (2008-2012)}
\format{
A data frame with 113 legislators and 1228 variables:
\describe{
\item{party}{Character. The political party of each legislator}
\item{7-1}{Roll call vote (numeric)}
\item{7-10}{Roll call vote (numeric)}
\item{7-100}{Roll call vote (numeric)}
\item{7-1000}{Roll call vote (numeric)}
\item{7-1001}{Roll call vote (numeric)}
\item{7-1002}{Roll call vote (numeric)}
\item{7-1003}{Roll call vote (numeric)}
\item{7-1004}{Roll call vote (numeric)}
\item{7-1005}{Roll call vote (numeric)}
\item{7-1006}{Roll call vote (numeric)}
\item{7-1007}{Roll call vote (numeric)}
\item{7-1008}{Roll call vote (numeric)}
\item{7-1009}{Roll call vote (numeric)}
\item{7-101}{Roll call vote (numeric)}
\item{7-1010}{Roll call vote (numeric)}
\item{7-1011}{Roll call vote (numeric)}
\item{7-1012}{Roll call vote (numeric)}
\item{7-1013}{Roll call vote (numeric)}
\item{7-1014}{Roll call vote (numeric)}
\item{7-1015}{Roll call vote (numeric)}
\item{7-1016}{Roll call vote (numeric)}
\item{7-1017}{Roll call vote (numeric)}
\item{7-1018}{Roll call vote (numeric)}
\item{7-1019}{Roll call vote (numeric)}
\item{7-102}{Roll call vote (numeric)}
\item{7-1020}{Roll call vote (numeric)}
\item{7-1021}{Roll call vote (numeric)}
\item{7-1022}{Roll call vote (numeric)}
\item{7-1023}{Roll call vote (numeric)}
\item{7-1024}{Roll call vote (numeric)}
\item{7-1025}{Roll call vote (numeric)}
\item{7-1026}{Roll call vote (numeric)}
\item{7-1027}{Roll call vote (numeric)}
\item{7-1028}{Roll call vote (numeric)}
\item{7-1029}{Roll call vote (numeric)}
\item{7-103}{Roll call vote (numeric)}
\item{7-1030}{Roll call vote (numeric)}
\item{7-1031}{Roll call vote (numeric)}
\item{7-1032}{Roll call vote (numeric)}
\item{7-1033}{Roll call vote (numeric)}
\item{7-1034}{Roll call vote (numeric)}
\item{7-1035}{Roll call vote (numeric)}
\item{7-1036}{Roll call vote (numeric)}
\item{7-1037}{Roll call vote (numeric)}
\item{7-1038}{Roll call vote (numeric)}
\item{7-1039}{Roll call vote (numeric)}
\item{7-104}{Roll call vote (numeric)}
\item{7-1040}{Roll call vote (numeric)}
\item{7-1041}{Roll call vote (numeric)}
\item{7-1042}{Roll call vote (numeric)}
\item{7-1043}{Roll call vote (numeric)}
\item{7-1044}{Roll call vote (numeric)}
\item{7-1045}{Roll call vote (numeric)}
\item{7-1046}{Roll call vote (numeric)}
\item{7-1047}{Roll call vote (numeric)}
\item{7-1048}{Roll call vote (numeric)}
\item{7-1049}{Roll call vote (numeric)}
\item{7-105}{Roll call vote (numeric)}
\item{7-1050}{Roll call vote (numeric)}
\item{7-1051}{Roll call vote (numeric)}
\item{7-1052}{Roll call vote (numeric)}
\item{7-1053}{Roll call vote (numeric)}
\item{7-1054}{Roll call vote (numeric)}
\item{7-1055}{Roll call vote (numeric)}
\item{7-1056}{Roll call vote (numeric)}
\item{7-1057}{Roll call vote (numeric)}
\item{7-1058}{Roll call vote (numeric)}
\item{7-1059}{Roll call vote (numeric)}
\item{7-106}{Roll call vote (numeric)}
\item{7-1060}{Roll call vote (numeric)}
\item{7-1061}{Roll call vote (numeric)}
\item{7-1062}{Roll call vote (numeric)}
\item{7-1063}{Roll call vote (numeric)}
\item{7-1064}{Roll call vote (numeric)}
\item{7-1065}{Roll call vote (numeric)}
\item{7-1066}{Roll call vote (numeric)}
\item{7-1067}{Roll call vote (numeric)}
\item{7-1068}{Roll call vote (numeric)}
\item{7-1069}{Roll call vote (numeric)}
\item{7-107}{Roll call vote (numeric)}
\item{7-1070}{Roll call vote (numeric)}
\item{7-1071}{Roll call vote (numeric)}
\item{7-1072}{Roll call vote (numeric)}
\item{7-1073}{Roll call vote (numeric)}
\item{7-1074}{Roll call vote (numeric)}
\item{7-1075}{Roll call vote (numeric)}
\item{7-1076}{Roll call vote (numeric)}
\item{7-1077}{Roll call vote (numeric)}
\item{7-1078}{Roll call vote (numeric)}
\item{7-1079}{Roll call vote (numeric)}
\item{7-108}{Roll call vote (numeric)}
\item{7-1080}{Roll call vote (numeric)}
\item{7-1081}{Roll call vote (numeric)}
\item{7-1082}{Roll call vote (numeric)}
\item{7-1083}{Roll call vote (numeric)}
\item{7-1084}{Roll call vote (numeric)}
\item{7-1085}{Roll call vote (numeric)}
\item{7-1086}{Roll call vote (numeric)}
\item{7-1087}{Roll call vote (numeric)}
\item{7-1088}{Roll call vote (numeric)}
\item{7-1089}{Roll call vote (numeric)}
\item{7-109}{Roll call vote (numeric)}
\item{7-1090}{Roll call vote (numeric)}
\item{7-1091}{Roll call vote (numeric)}
\item{7-1092}{Roll call vote (numeric)}
\item{7-1093}{Roll call vote (numeric)}
\item{7-1094}{Roll call vote (numeric)}
\item{7-1095}{Roll call vote (numeric)}
\item{7-1096}{Roll call vote (numeric)}
\item{7-1097}{Roll call vote (numeric)}
\item{7-1098}{Roll call vote (numeric)}
\item{7-1099}{Roll call vote (numeric)}
\item{7-11}{Roll call vote (numeric)}
\item{7-110}{Roll call vote (numeric)}
\item{7-1100}{Roll call vote (numeric)}
\item{7-1101}{Roll call vote (numeric)}
\item{7-1102}{Roll call vote (numeric)}
\item{7-1103}{Roll call vote (numeric)}
\item{7-1104}{Roll call vote (numeric)}
\item{7-1105}{Roll call vote (numeric)}
\item{7-1106}{Roll call vote (numeric)}
\item{7-1107}{Roll call vote (numeric)}
\item{7-1108}{Roll call vote (numeric)}
\item{7-1109}{Roll call vote (numeric)}
\item{7-111}{Roll call vote (numeric)}
\item{7-1110}{Roll call vote (numeric)}
\item{7-1111}{Roll call vote (numeric)}
\item{7-1112}{Roll call vote (numeric)}
\item{7-1113}{Roll call vote (numeric)}
\item{7-1114}{Roll call vote (numeric)}
\item{7-1115}{Roll call vote (numeric)}
\item{7-1116}{Roll call vote (numeric)}
\item{7-1117}{Roll call vote (numeric)}
\item{7-1118}{Roll call vote (numeric)}
\item{7-1119}{Roll call vote (numeric)}
\item{7-112}{Roll call vote (numeric)}
\item{7-1120}{Roll call vote (numeric)}
\item{7-1121}{Roll call vote (numeric)}
\item{7-1122}{Roll call vote (numeric)}
\item{7-1123}{Roll call vote (numeric)}
\item{7-1124}{Roll call vote (numeric)}
\item{7-1125}{Roll call vote (numeric)}
\item{7-1126}{Roll call vote (numeric)}
\item{7-1127}{Roll call vote (numeric)}
\item{7-1128}{Roll call vote (numeric)}
\item{7-1129}{Roll call vote (numeric)}
\item{7-113}{Roll call vote (numeric)}
\item{7-1130}{Roll call vote (numeric)}
\item{7-1131}{Roll call vote (numeric)}
\item{7-1132}{Roll call vote (numeric)}
\item{7-1133}{Roll call vote (numeric)}
\item{7-1134}{Roll call vote (numeric)}
\item{7-1135}{Roll call vote (numeric)}
\item{7-1136}{Roll call vote (numeric)}
\item{7-1137}{Roll call vote (numeric)}
\item{7-1138}{Roll call vote (numeric)}
\item{7-1139}{Roll call vote (numeric)}
\item{7-114}{Roll call vote (numeric)}
\item{7-1140}{Roll call vote (numeric)}
\item{7-1141}{Roll call vote (numeric)}
\item{7-1142}{Roll call vote (numeric)}
\item{7-1143}{Roll call vote (numeric)}
\item{7-1144}{Roll call vote (numeric)}
\item{7-1145}{Roll call vote (numeric)}
\item{7-1146}{Roll call vote (numeric)}
\item{7-1147}{Roll call vote (numeric)}
\item{7-1148}{Roll call vote (numeric)}
\item{7-1149}{Roll call vote (numeric)}
\item{7-115}{Roll call vote (numeric)}
\item{7-1150}{Roll call vote (numeric)}
\item{7-1151}{Roll call vote (numeric)}
\item{7-1152}{Roll call vote (numeric)}
\item{7-1153}{Roll call vote (numeric)}
\item{7-1154}{Roll call vote (numeric)}
\item{7-1155}{Roll call vote (numeric)}
\item{7-1156}{Roll call vote (numeric)}
\item{7-1157}{Roll call vote (numeric)}
\item{7-1158}{Roll call vote (numeric)}
\item{7-1159}{Roll call vote (numeric)}
\item{7-116}{Roll call vote (numeric)}
\item{7-1160}{Roll call vote (numeric)}
\item{7-1161}{Roll call vote (numeric)}
\item{7-1162}{Roll call vote (numeric)}
\item{7-1163}{Roll call vote (numeric)}
\item{7-1164}{Roll call vote (numeric)}
\item{7-1165}{Roll call vote (numeric)}
\item{7-1166}{Roll call vote (numeric)}
\item{7-1167}{Roll call vote (numeric)}
\item{7-1168}{Roll call vote (numeric)}
\item{7-1169}{Roll call vote (numeric)}
\item{7-117}{Roll call vote (numeric)}
\item{7-1170}{Roll call vote (numeric)}
\item{7-1171}{Roll call vote (numeric)}
\item{7-1172}{Roll call vote (numeric)}
\item{7-1173}{Roll call vote (numeric)}
\item{7-1174}{Roll call vote (numeric)}
\item{7-1175}{Roll call vote (numeric)}
\item{7-1176}{Roll call vote (numeric)}
\item{7-1177}{Roll call vote (numeric)}
\item{7-1178}{Roll call vote (numeric)}
\item{7-1179}{Roll call vote (numeric)}
\item{7-118}{Roll call vote (numeric)}
\item{7-1180}{Roll call vote (numeric)}
\item{7-1181}{Roll call vote (numeric)}
\item{7-1182}{Roll call vote (numeric)}
\item{7-1183}{Roll call vote (numeric)}
\item{7-1184}{Roll call vote (numeric)}
\item{7-1185}{Roll call vote (numeric)}
\item{7-1186}{Roll call vote (numeric)}
\item{7-1187}{Roll call vote (numeric)}
\item{7-1188}{Roll call vote (numeric)}
\item{7-1189}{Roll call vote (numeric)}
\item{7-119}{Roll call vote (numeric)}
\item{7-1190}{Roll call vote (numeric)}
\item{7-1191}{Roll call vote (numeric)}
\item{7-1192}{Roll call vote (numeric)}
\item{7-1193}{Roll call vote (numeric)}
\item{7-1194}{Roll call vote (numeric)}
\item{7-1195}{Roll call vote (numeric)}
\item{7-1196}{Roll call vote (numeric)}
\item{7-1197}{Roll call vote (numeric)}
\item{7-1198}{Roll call vote (numeric)}
\item{7-1199}{Roll call vote (numeric)}
\item{7-12}{Roll call vote (numeric)}
\item{7-120}{Roll call vote (numeric)}
\item{7-1200}{Roll call vote (numeric)}
\item{7-1201}{Roll call vote (numeric)}
\item{7-1202}{Roll call vote (numeric)}
\item{7-1203}{Roll call vote (numeric)}
\item{7-1204}{Roll call vote (numeric)}
\item{7-1205}{Roll call vote (numeric)}
\item{7-1206}{Roll call vote (numeric)}
\item{7-1207}{Roll call vote (numeric)}
\item{7-1208}{Roll call vote (numeric)}
\item{7-1209}{Roll call vote (numeric)}
\item{7-121}{Roll call vote (numeric)}
\item{7-1210}{Roll call vote (numeric)}
\item{7-1211}{Roll call vote (numeric)}
\item{7-1212}{Roll call vote (numeric)}
\item{7-1213}{Roll call vote (numeric)}
\item{7-1214}{Roll call vote (numeric)}
\item{7-1215}{Roll call vote (numeric)}
\item{7-1216}{Roll call vote (numeric)}
\item{7-1217}{Roll call vote (numeric)}
\item{7-1218}{Roll call vote (numeric)}
\item{7-1219}{Roll call vote (numeric)}
\item{7-122}{Roll call vote (numeric)}
\item{7-1220}{Roll call vote (numeric)}
\item{7-1221}{Roll call vote (numeric)}
\item{7-1223}{Roll call vote (numeric)}
\item{7-1224}{Roll call vote (numeric)}
\item{7-1225}{Roll call vote (numeric)}
\item{7-1226}{Roll call vote (numeric)}
\item{7-123}{Roll call vote (numeric)}
\item{7-124}{Roll call vote (numeric)}
\item{7-125}{Roll call vote (numeric)}
\item{7-126}{Roll call vote (numeric)}
\item{7-127}{Roll call vote (numeric)}
\item{7-128}{Roll call vote (numeric)}
\item{7-129}{Roll call vote (numeric)}
\item{7-13}{Roll call vote (numeric)}
\item{7-130}{Roll call vote (numeric)}
\item{7-131}{Roll call vote (numeric)}
\item{7-132}{Roll call vote (numeric)}
\item{7-133}{Roll call vote (numeric)}
\item{7-134}{Roll call vote (numeric)}
\item{7-135}{Roll call vote (numeric)}
\item{7-136}{Roll call vote (numeric)}
\item{7-137}{Roll call vote (numeric)}
\item{7-138}{Roll call vote (numeric)}
\item{7-139}{Roll call vote (numeric)}
\item{7-14}{Roll call vote (numeric)}
\item{7-140}{Roll call vote (numeric)}
\item{7-141}{Roll call vote (numeric)}
\item{7-142}{Roll call vote (numeric)}
\item{7-143}{Roll call vote (numeric)}
\item{7-144}{Roll call vote (numeric)}
\item{7-145}{Roll call vote (numeric)}
\item{7-146}{Roll call vote (numeric)}
\item{7-147}{Roll call vote (numeric)}
\item{7-148}{Roll call vote (numeric)}
\item{7-149}{Roll call vote (numeric)}
\item{7-15}{Roll call vote (numeric)}
\item{7-150}{Roll call vote (numeric)}
\item{7-151}{Roll call vote (numeric)}
\item{7-152}{Roll call vote (numeric)}
\item{7-153}{Roll call vote (numeric)}
\item{7-154}{Roll call vote (numeric)}
\item{7-155}{Roll call vote (numeric)}
\item{7-156}{Roll call vote (numeric)}
\item{7-157}{Roll call vote (numeric)}
\item{7-158}{Roll call vote (numeric)}
\item{7-159}{Roll call vote (numeric)}
\item{7-16}{Roll call vote (numeric)}
\item{7-160}{Roll call vote (numeric)}
\item{7-161}{Roll call vote (numeric)}
\item{7-162}{Roll call vote (numeric)}
\item{7-163}{Roll call vote (numeric)}
\item{7-164}{Roll call vote (numeric)}
\item{7-165}{Roll call vote (numeric)}
\item{7-166}{Roll call vote (numeric)}
\item{7-167}{Roll call vote (numeric)}
\item{7-168}{Roll call vote (numeric)}
\item{7-169}{Roll call vote (numeric)}
\item{7-17}{Roll call vote (numeric)}
\item{7-170}{Roll call vote (numeric)}
\item{7-171}{Roll call vote (numeric)}
\item{7-172}{Roll call vote (numeric)}
\item{7-173}{Roll call vote (numeric)}
\item{7-174}{Roll call vote (numeric)}
\item{7-175}{Roll call vote (numeric)}
\item{7-176}{Roll call vote (numeric)}
\item{7-177}{Roll call vote (numeric)}
\item{7-178}{Roll call vote (numeric)}
\item{7-179}{Roll call vote (numeric)}
\item{7-18}{Roll call vote (numeric)}
\item{7-180}{Roll call vote (numeric)}
\item{7-181}{Roll call vote (numeric)}
\item{7-182}{Roll call vote (numeric)}
\item{7-183}{Roll call vote (numeric)}
\item{7-184}{Roll call vote (numeric)}
\item{7-185}{Roll call vote (numeric)}
\item{7-186}{Roll call vote (numeric)}
\item{7-187}{Roll call vote (numeric)}
\item{7-188}{Roll call vote (numeric)}
\item{7-189}{Roll call vote (numeric)}
\item{7-19}{Roll call vote (numeric)}
\item{7-190}{Roll call vote (numeric)}
\item{7-191}{Roll call vote (numeric)}
\item{7-192}{Roll call vote (numeric)}
\item{7-193}{Roll call vote (numeric)}
\item{7-194}{Roll call vote (numeric)}
\item{7-195}{Roll call vote (numeric)}
\item{7-196}{Roll call vote (numeric)}
\item{7-197}{Roll call vote (numeric)}
\item{7-198}{Roll call vote (numeric)}
\item{7-199}{Roll call vote (numeric)}
\item{7-2}{Roll call vote (numeric)}
\item{7-20}{Roll call vote (numeric)}
\item{7-200}{Roll call vote (numeric)}
\item{7-201}{Roll call vote (numeric)}
\item{7-202}{Roll call vote (numeric)}
\item{7-203}{Roll call vote (numeric)}
\item{7-204}{Roll call vote (numeric)}
\item{7-205}{Roll call vote (numeric)}
\item{7-206}{Roll call vote (numeric)}
\item{7-207}{Roll call vote (numeric)}
\item{7-208}{Roll call vote (numeric)}
\item{7-209}{Roll call vote (numeric)}
\item{7-21}{Roll call vote (numeric)}
\item{7-210}{Roll call vote (numeric)}
\item{7-211}{Roll call vote (numeric)}
\item{7-212}{Roll call vote (numeric)}
\item{7-213}{Roll call vote (numeric)}
\item{7-214}{Roll call vote (numeric)}
\item{7-215}{Roll call vote (numeric)}
\item{7-216}{Roll call vote (numeric)}
\item{7-217}{Roll call vote (numeric)}
\item{7-218}{Roll call vote (numeric)}
\item{7-219}{Roll call vote (numeric)}
\item{7-22}{Roll call vote (numeric)}
\item{7-220}{Roll call vote (numeric)}
\item{7-221}{Roll call vote (numeric)}
\item{7-222}{Roll call vote (numeric)}
\item{7-223}{Roll call vote (numeric)}
\item{7-224}{Roll call vote (numeric)}
\item{7-225}{Roll call vote (numeric)}
\item{7-226}{Roll call vote (numeric)}
\item{7-227}{Roll call vote (numeric)}
\item{7-228}{Roll call vote (numeric)}
\item{7-229}{Roll call vote (numeric)}
\item{7-23}{Roll call vote (numeric)}
\item{7-230}{Roll call vote (numeric)}
\item{7-231}{Roll call vote (numeric)}
\item{7-232}{Roll call vote (numeric)}
\item{7-233}{Roll call vote (numeric)}
\item{7-234}{Roll call vote (numeric)}
\item{7-235}{Roll call vote (numeric)}
\item{7-236}{Roll call vote (numeric)}
\item{7-237}{Roll call vote (numeric)}
\item{7-238}{Roll call vote (numeric)}
\item{7-239}{Roll call vote (numeric)}
\item{7-24}{Roll call vote (numeric)}
\item{7-240}{Roll call vote (numeric)}
\item{7-241}{Roll call vote (numeric)}
\item{7-242}{Roll call vote (numeric)}
\item{7-243}{Roll call vote (numeric)}
\item{7-244}{Roll call vote (numeric)}
\item{7-245}{Roll call vote (numeric)}
\item{7-246}{Roll call vote (numeric)}
\item{7-247}{Roll call vote (numeric)}
\item{7-248}{Roll call vote (numeric)}
\item{7-249}{Roll call vote (numeric)}
\item{7-25}{Roll call vote (numeric)}
\item{7-250}{Roll call vote (numeric)}
\item{7-251}{Roll call vote (numeric)}
\item{7-252}{Roll call vote (numeric)}
\item{7-253}{Roll call vote (numeric)}
\item{7-254}{Roll call vote (numeric)}
\item{7-255}{Roll call vote (numeric)}
\item{7-256}{Roll call vote (numeric)}
\item{7-257}{Roll call vote (numeric)}
\item{7-258}{Roll call vote (numeric)}
\item{7-259}{Roll call vote (numeric)}
\item{7-26}{Roll call vote (numeric)}
\item{7-260}{Roll call vote (numeric)}
\item{7-261}{Roll call vote (numeric)}
\item{7-262}{Roll call vote (numeric)}
\item{7-263}{Roll call vote (numeric)}
\item{7-264}{Roll call vote (numeric)}
\item{7-265}{Roll call vote (numeric)}
\item{7-266}{Roll call vote (numeric)}
\item{7-267}{Roll call vote (numeric)}
\item{7-268}{Roll call vote (numeric)}
\item{7-269}{Roll call vote (numeric)}
\item{7-27}{Roll call vote (numeric)}
\item{7-270}{Roll call vote (numeric)}
\item{7-271}{Roll call vote (numeric)}
\item{7-272}{Roll call vote (numeric)}
\item{7-273}{Roll call vote (numeric)}
\item{7-274}{Roll call vote (numeric)}
\item{7-275}{Roll call vote (numeric)}
\item{7-276}{Roll call vote (numeric)}
\item{7-277}{Roll call vote (numeric)}
\item{7-278}{Roll call vote (numeric)}
\item{7-279}{Roll call vote (numeric)}
\item{7-28}{Roll call vote (numeric)}
\item{7-280}{Roll call vote (numeric)}
\item{7-281}{Roll call vote (numeric)}
\item{7-282}{Roll call vote (numeric)}
\item{7-283}{Roll call vote (numeric)}
\item{7-284}{Roll call vote (numeric)}
\item{7-285}{Roll call vote (numeric)}
\item{7-286}{Roll call vote (numeric)}
\item{7-287}{Roll call vote (numeric)}
\item{7-288}{Roll call vote (numeric)}
\item{7-289}{Roll call vote (numeric)}
\item{7-29}{Roll call vote (numeric)}
\item{7-290}{Roll call vote (numeric)}
\item{7-291}{Roll call vote (numeric)}
\item{7-292}{Roll call vote (numeric)}
\item{7-293}{Roll call vote (numeric)}
\item{7-294}{Roll call vote (numeric)}
\item{7-295}{Roll call vote (numeric)}
\item{7-296}{Roll call vote (numeric)}
\item{7-297}{Roll call vote (numeric)}
\item{7-298}{Roll call vote (numeric)}
\item{7-299}{Roll call vote (numeric)}
\item{7-3}{Roll call vote (numeric)}
\item{7-30}{Roll call vote (numeric)}
\item{7-300}{Roll call vote (numeric)}
\item{7-301}{Roll call vote (numeric)}
\item{7-302}{Roll call vote (numeric)}
\item{7-303}{Roll call vote (numeric)}
\item{7-304}{Roll call vote (numeric)}
\item{7-305}{Roll call vote (numeric)}
\item{7-306}{Roll call vote (numeric)}
\item{7-307}{Roll call vote (numeric)}
\item{7-308}{Roll call vote (numeric)}
\item{7-309}{Roll call vote (numeric)}
\item{7-31}{Roll call vote (numeric)}
\item{7-310}{Roll call vote (numeric)}
\item{7-311}{Roll call vote (numeric)}
\item{7-312}{Roll call vote (numeric)}
\item{7-313}{Roll call vote (numeric)}
\item{7-314}{Roll call vote (numeric)}
\item{7-315}{Roll call vote (numeric)}
\item{7-316}{Roll call vote (numeric)}
\item{7-317}{Roll call vote (numeric)}
\item{7-318}{Roll call vote (numeric)}
\item{7-319}{Roll call vote (numeric)}
\item{7-32}{Roll call vote (numeric)}
\item{7-320}{Roll call vote (numeric)}
\item{7-321}{Roll call vote (numeric)}
\item{7-322}{Roll call vote (numeric)}
\item{7-323}{Roll call vote (numeric)}
\item{7-324}{Roll call vote (numeric)}
\item{7-325}{Roll call vote (numeric)}
\item{7-326}{Roll call vote (numeric)}
\item{7-327}{Roll call vote (numeric)}
\item{7-328}{Roll call vote (numeric)}
\item{7-329}{Roll call vote (numeric)}
\item{7-33}{Roll call vote (numeric)}
\item{7-330}{Roll call vote (numeric)}
\item{7-331}{Roll call vote (numeric)}
\item{7-332}{Roll call vote (numeric)}
\item{7-333}{Roll call vote (numeric)}
\item{7-334}{Roll call vote (numeric)}
\item{7-335}{Roll call vote (numeric)}
\item{7-336}{Roll call vote (numeric)}
\item{7-337}{Roll call vote (numeric)}
\item{7-338}{Roll call vote (numeric)}
\item{7-339}{Roll call vote (numeric)}
\item{7-34}{Roll call vote (numeric)}
\item{7-340}{Roll call vote (numeric)}
\item{7-341}{Roll call vote (numeric)}
\item{7-342}{Roll call vote (numeric)}
\item{7-343}{Roll call vote (numeric)}
\item{7-344}{Roll call vote (numeric)}
\item{7-345}{Roll call vote (numeric)}
\item{7-346}{Roll call vote (numeric)}
\item{7-347}{Roll call vote (numeric)}
\item{7-348}{Roll call vote (numeric)}
\item{7-349}{Roll call vote (numeric)}
\item{7-35}{Roll call vote (numeric)}
\item{7-350}{Roll call vote (numeric)}
\item{7-351}{Roll call vote (numeric)}
\item{7-352}{Roll call vote (numeric)}
\item{7-353}{Roll call vote (numeric)}
\item{7-354}{Roll call vote (numeric)}
\item{7-355}{Roll call vote (numeric)}
\item{7-356}{Roll call vote (numeric)}
\item{7-357}{Roll call vote (numeric)}
\item{7-358}{Roll call vote (numeric)}
\item{7-359}{Roll call vote (numeric)}
\item{7-36}{Roll call vote (numeric)}
\item{7-360}{Roll call vote (numeric)}
\item{7-361}{Roll call vote (numeric)}
\item{7-362}{Roll call vote (numeric)}
\item{7-363}{Roll call vote (numeric)}
\item{7-364}{Roll call vote (numeric)}
\item{7-365}{Roll call vote (numeric)}
\item{7-366}{Roll call vote (numeric)}
\item{7-367}{Roll call vote (numeric)}
\item{7-368}{Roll call vote (numeric)}
\item{7-369}{Roll call vote (numeric)}
\item{7-37}{Roll call vote (numeric)}
\item{7-370}{Roll call vote (numeric)}
\item{7-371}{Roll call vote (numeric)}
\item{7-372}{Roll call vote (numeric)}
\item{7-373}{Roll call vote (numeric)}
\item{7-374}{Roll call vote (numeric)}
\item{7-375}{Roll call vote (numeric)}
\item{7-376}{Roll call vote (numeric)}
\item{7-377}{Roll call vote (numeric)}
\item{7-378}{Roll call vote (numeric)}
\item{7-379}{Roll call vote (numeric)}
\item{7-38}{Roll call vote (numeric)}
\item{7-380}{Roll call vote (numeric)}
\item{7-381}{Roll call vote (numeric)}
\item{7-382}{Roll call vote (numeric)}
\item{7-383}{Roll call vote (numeric)}
\item{7-384}{Roll call vote (numeric)}
\item{7-385}{Roll call vote (numeric)}
\item{7-386}{Roll call vote (numeric)}
\item{7-387}{Roll call vote (numeric)}
\item{7-388}{Roll call vote (numeric)}
\item{7-389}{Roll call vote (numeric)}
\item{7-39}{Roll call vote (numeric)}
\item{7-390}{Roll call vote (numeric)}
\item{7-391}{Roll call vote (numeric)}
\item{7-392}{Roll call vote (numeric)}
\item{7-393}{Roll call vote (numeric)}
\item{7-394}{Roll call vote (numeric)}
\item{7-395}{Roll call vote (numeric)}
\item{7-396}{Roll call vote (numeric)}
\item{7-397}{Roll call vote (numeric)}
\item{7-398}{Roll call vote (numeric)}
\item{7-399}{Roll call vote (numeric)}
\item{7-4}{Roll call vote (numeric)}
\item{7-40}{Roll call vote (numeric)}
\item{7-400}{Roll call vote (numeric)}
\item{7-401}{Roll call vote (numeric)}
\item{7-402}{Roll call vote (numeric)}
\item{7-403}{Roll call vote (numeric)}
\item{7-404}{Roll call vote (numeric)}
\item{7-405}{Roll call vote (numeric)}
\item{7-406}{Roll call vote (numeric)}
\item{7-407}{Roll call vote (numeric)}
\item{7-408}{Roll call vote (numeric)}
\item{7-409}{Roll call vote (numeric)}
\item{7-41}{Roll call vote (numeric)}
\item{7-410}{Roll call vote (numeric)}
\item{7-411}{Roll call vote (numeric)}
\item{7-412}{Roll call vote (numeric)}
\item{7-413}{Roll call vote (numeric)}
\item{7-414}{Roll call vote (numeric)}
\item{7-415}{Roll call vote (numeric)}
\item{7-416}{Roll call vote (numeric)}
\item{7-417}{Roll call vote (numeric)}
\item{7-418}{Roll call vote (numeric)}
\item{7-419}{Roll call vote (numeric)}
\item{7-42}{Roll call vote (numeric)}
\item{7-420}{Roll call vote (numeric)}
\item{7-421}{Roll call vote (numeric)}
\item{7-422}{Roll call vote (numeric)}
\item{7-423}{Roll call vote (numeric)}
\item{7-424}{Roll call vote (numeric)}
\item{7-425}{Roll call vote (numeric)}
\item{7-426}{Roll call vote (numeric)}
\item{7-427}{Roll call vote (numeric)}
\item{7-428}{Roll call vote (numeric)}
\item{7-429}{Roll call vote (numeric)}
\item{7-43}{Roll call vote (numeric)}
\item{7-430}{Roll call vote (numeric)}
\item{7-431}{Roll call vote (numeric)}
\item{7-432}{Roll call vote (numeric)}
\item{7-433}{Roll call vote (numeric)}
\item{7-434}{Roll call vote (numeric)}
\item{7-435}{Roll call vote (numeric)}
\item{7-436}{Roll call vote (numeric)}
\item{7-437}{Roll call vote (numeric)}
\item{7-438}{Roll call vote (numeric)}
\item{7-439}{Roll call vote (numeric)}
\item{7-44}{Roll call vote (numeric)}
\item{7-440}{Roll call vote (numeric)}
\item{7-441}{Roll call vote (numeric)}
\item{7-442}{Roll call vote (numeric)}
\item{7-443}{Roll call vote (numeric)}
\item{7-444}{Roll call vote (numeric)}
\item{7-445}{Roll call vote (numeric)}
\item{7-446}{Roll call vote (numeric)}
\item{7-447}{Roll call vote (numeric)}
\item{7-448}{Roll call vote (numeric)}
\item{7-449}{Roll call vote (numeric)}
\item{7-45}{Roll call vote (numeric)}
\item{7-450}{Roll call vote (numeric)}
\item{7-451}{Roll call vote (numeric)}
\item{7-452}{Roll call vote (numeric)}
\item{7-453}{Roll call vote (numeric)}
\item{7-454}{Roll call vote (numeric)}
\item{7-455}{Roll call vote (numeric)}
\item{7-456}{Roll call vote (numeric)}
\item{7-457}{Roll call vote (numeric)}
\item{7-458}{Roll call vote (numeric)}
\item{7-459}{Roll call vote (numeric)}
\item{7-46}{Roll call vote (numeric)}
\item{7-460}{Roll call vote (numeric)}
\item{7-461}{Roll call vote (numeric)}
\item{7-462}{Roll call vote (numeric)}
\item{7-463}{Roll call vote (numeric)}
\item{7-464}{Roll call vote (numeric)}
\item{7-465}{Roll call vote (numeric)}
\item{7-466}{Roll call vote (numeric)}
\item{7-467}{Roll call vote (numeric)}
\item{7-468}{Roll call vote (numeric)}
\item{7-469}{Roll call vote (numeric)}
\item{7-47}{Roll call vote (numeric)}
\item{7-470}{Roll call vote (numeric)}
\item{7-471}{Roll call vote (numeric)}
\item{7-472}{Roll call vote (numeric)}
\item{7-473}{Roll call vote (numeric)}
\item{7-474}{Roll call vote (numeric)}
\item{7-475}{Roll call vote (numeric)}
\item{7-476}{Roll call vote (numeric)}
\item{7-477}{Roll call vote (numeric)}
\item{7-478}{Roll call vote (numeric)}
\item{7-479}{Roll call vote (numeric)}
\item{7-48}{Roll call vote (numeric)}
\item{7-480}{Roll call vote (numeric)}
\item{7-481}{Roll call vote (numeric)}
\item{7-482}{Roll call vote (numeric)}
\item{7-483}{Roll call vote (numeric)}
\item{7-484}{Roll call vote (numeric)}
\item{7-485}{Roll call vote (numeric)}
\item{7-486}{Roll call vote (numeric)}
\item{7-487}{Roll call vote (numeric)}
\item{7-488}{Roll call vote (numeric)}
\item{7-489}{Roll call vote (numeric)}
\item{7-49}{Roll call vote (numeric)}
\item{7-490}{Roll call vote (numeric)}
\item{7-491}{Roll call vote (numeric)}
\item{7-492}{Roll call vote (numeric)}
\item{7-493}{Roll call vote (numeric)}
\item{7-494}{Roll call vote (numeric)}
\item{7-495}{Roll call vote (numeric)}
\item{7-496}{Roll call vote (numeric)}
\item{7-497}{Roll call vote (numeric)}
\item{7-498}{Roll call vote (numeric)}
\item{7-499}{Roll call vote (numeric)}
\item{7-5}{Roll call vote (numeric)}
\item{7-50}{Roll call vote (numeric)}
\item{7-500}{Roll call vote (numeric)}
\item{7-501}{Roll call vote (numeric)}
\item{7-502}{Roll call vote (numeric)}
\item{7-503}{Roll call vote (numeric)}
\item{7-504}{Roll call vote (numeric)}
\item{7-505}{Roll call vote (numeric)}
\item{7-506}{Roll call vote (numeric)}
\item{7-507}{Roll call vote (numeric)}
\item{7-508}{Roll call vote (numeric)}
\item{7-509}{Roll call vote (numeric)}
\item{7-51}{Roll call vote (numeric)}
\item{7-510}{Roll call vote (numeric)}
\item{7-511}{Roll call vote (numeric)}
\item{7-512}{Roll call vote (numeric)}
\item{7-513}{Roll call vote (numeric)}
\item{7-514}{Roll call vote (numeric)}
\item{7-515}{Roll call vote (numeric)}
\item{7-516}{Roll call vote (numeric)}
\item{7-517}{Roll call vote (numeric)}
\item{7-518}{Roll call vote (numeric)}
\item{7-519}{Roll call vote (numeric)}
\item{7-52}{Roll call vote (numeric)}
\item{7-520}{Roll call vote (numeric)}
\item{7-521}{Roll call vote (numeric)}
\item{7-522}{Roll call vote (numeric)}
\item{7-523}{Roll call vote (numeric)}
\item{7-524}{Roll call vote (numeric)}
\item{7-525}{Roll call vote (numeric)}
\item{7-526}{Roll call vote (numeric)}
\item{7-527}{Roll call vote (numeric)}
\item{7-528}{Roll call vote (numeric)}
\item{7-529}{Roll call vote (numeric)}
\item{7-53}{Roll call vote (numeric)}
\item{7-530}{Roll call vote (numeric)}
\item{7-531}{Roll call vote (numeric)}
\item{7-532}{Roll call vote (numeric)}
\item{7-533}{Roll call vote (numeric)}
\item{7-534}{Roll call vote (numeric)}
\item{7-535}{Roll call vote (numeric)}
\item{7-536}{Roll call vote (numeric)}
\item{7-537}{Roll call vote (numeric)}
\item{7-538}{Roll call vote (numeric)}
\item{7-539}{Roll call vote (numeric)}
\item{7-54}{Roll call vote (numeric)}
\item{7-540}{Roll call vote (numeric)}
\item{7-541}{Roll call vote (numeric)}
\item{7-542}{Roll call vote (numeric)}
\item{7-543}{Roll call vote (numeric)}
\item{7-544}{Roll call vote (numeric)}
\item{7-545}{Roll call vote (numeric)}
\item{7-546}{Roll call vote (numeric)}
\item{7-547}{Roll call vote (numeric)}
\item{7-548}{Roll call vote (numeric)}
\item{7-549}{Roll call vote (numeric)}
\item{7-55}{Roll call vote (numeric)}
\item{7-550}{Roll call vote (numeric)}
\item{7-551}{Roll call vote (numeric)}
\item{7-552}{Roll call vote (numeric)}
\item{7-553}{Roll call vote (numeric)}
\item{7-554}{Roll call vote (numeric)}
\item{7-555}{Roll call vote (numeric)}
\item{7-556}{Roll call vote (numeric)}
\item{7-557}{Roll call vote (numeric)}
\item{7-558}{Roll call vote (numeric)}
\item{7-559}{Roll call vote (numeric)}
\item{7-56}{Roll call vote (numeric)}
\item{7-560}{Roll call vote (numeric)}
\item{7-561}{Roll call vote (numeric)}
\item{7-562}{Roll call vote (numeric)}
\item{7-563}{Roll call vote (numeric)}
\item{7-564}{Roll call vote (numeric)}
\item{7-565}{Roll call vote (numeric)}
\item{7-566}{Roll call vote (numeric)}
\item{7-567}{Roll call vote (numeric)}
\item{7-568}{Roll call vote (numeric)}
\item{7-569}{Roll call vote (numeric)}
\item{7-57}{Roll call vote (numeric)}
\item{7-570}{Roll call vote (numeric)}
\item{7-571}{Roll call vote (numeric)}
\item{7-572}{Roll call vote (numeric)}
\item{7-573}{Roll call vote (numeric)}
\item{7-574}{Roll call vote (numeric)}
\item{7-575}{Roll call vote (numeric)}
\item{7-576}{Roll call vote (numeric)}
\item{7-577}{Roll call vote (numeric)}
\item{7-578}{Roll call vote (numeric)}
\item{7-579}{Roll call vote (numeric)}
\item{7-58}{Roll call vote (numeric)}
\item{7-580}{Roll call vote (numeric)}
\item{7-581}{Roll call vote (numeric)}
\item{7-582}{Roll call vote (numeric)}
\item{7-583}{Roll call vote (numeric)}
\item{7-584}{Roll call vote (numeric)}
\item{7-585}{Roll call vote (numeric)}
\item{7-586}{Roll call vote (numeric)}
\item{7-587}{Roll call vote (numeric)}
\item{7-588}{Roll call vote (numeric)}
\item{7-589}{Roll call vote (numeric)}
\item{7-59}{Roll call vote (numeric)}
\item{7-590}{Roll call vote (numeric)}
\item{7-591}{Roll call vote (numeric)}
\item{7-592}{Roll call vote (numeric)}
\item{7-593}{Roll call vote (numeric)}
\item{7-594}{Roll call vote (numeric)}
\item{7-595}{Roll call vote (numeric)}
\item{7-596}{Roll call vote (numeric)}
\item{7-597}{Roll call vote (numeric)}
\item{7-598}{Roll call vote (numeric)}
\item{7-599}{Roll call vote (numeric)}
\item{7-6}{Roll call vote (numeric)}
\item{7-60}{Roll call vote (numeric)}
\item{7-600}{Roll call vote (numeric)}
\item{7-601}{Roll call vote (numeric)}
\item{7-602}{Roll call vote (numeric)}
\item{7-603}{Roll call vote (numeric)}
\item{7-604}{Roll call vote (numeric)}
\item{7-605}{Roll call vote (numeric)}
\item{7-606}{Roll call vote (numeric)}
\item{7-607}{Roll call vote (numeric)}
\item{7-608}{Roll call vote (numeric)}
\item{7-609}{Roll call vote (numeric)}
\item{7-61}{Roll call vote (numeric)}
\item{7-610}{Roll call vote (numeric)}
\item{7-611}{Roll call vote (numeric)}
\item{7-612}{Roll call vote (numeric)}
\item{7-613}{Roll call vote (numeric)}
\item{7-614}{Roll call vote (numeric)}
\item{7-615}{Roll call vote (numeric)}
\item{7-616}{Roll call vote (numeric)}
\item{7-617}{Roll call vote (numeric)}
\item{7-618}{Roll call vote (numeric)}
\item{7-619}{Roll call vote (numeric)}
\item{7-62}{Roll call vote (numeric)}
\item{7-620}{Roll call vote (numeric)}
\item{7-621}{Roll call vote (numeric)}
\item{7-622}{Roll call vote (numeric)}
\item{7-623}{Roll call vote (numeric)}
\item{7-624}{Roll call vote (numeric)}
\item{7-625}{Roll call vote (numeric)}
\item{7-626}{Roll call vote (numeric)}
\item{7-627}{Roll call vote (numeric)}
\item{7-628}{Roll call vote (numeric)}
\item{7-629}{Roll call vote (numeric)}
\item{7-63}{Roll call vote (numeric)}
\item{7-630}{Roll call vote (numeric)}
\item{7-631}{Roll call vote (numeric)}
\item{7-632}{Roll call vote (numeric)}
\item{7-633}{Roll call vote (numeric)}
\item{7-634}{Roll call vote (numeric)}
\item{7-635}{Roll call vote (numeric)}
\item{7-636}{Roll call vote (numeric)}
\item{7-637}{Roll call vote (numeric)}
\item{7-638}{Roll call vote (numeric)}
\item{7-639}{Roll call vote (numeric)}
\item{7-64}{Roll call vote (numeric)}
\item{7-640}{Roll call vote (numeric)}
\item{7-641}{Roll call vote (numeric)}
\item{7-642}{Roll call vote (numeric)}
\item{7-643}{Roll call vote (numeric)}
\item{7-644}{Roll call vote (numeric)}
\item{7-645}{Roll call vote (numeric)}
\item{7-646}{Roll call vote (numeric)}
\item{7-647}{Roll call vote (numeric)}
\item{7-648}{Roll call vote (numeric)}
\item{7-649}{Roll call vote (numeric)}
\item{7-65}{Roll call vote (numeric)}
\item{7-650}{Roll call vote (numeric)}
\item{7-651}{Roll call vote (numeric)}
\item{7-652}{Roll call vote (numeric)}
\item{7-653}{Roll call vote (numeric)}
\item{7-654}{Roll call vote (numeric)}
\item{7-655}{Roll call vote (numeric)}
\item{7-656}{Roll call vote (numeric)}
\item{7-657}{Roll call vote (numeric)}
\item{7-658}{Roll call vote (numeric)}
\item{7-659}{Roll call vote (numeric)}
\item{7-66}{Roll call vote (numeric)}
\item{7-660}{Roll call vote (numeric)}
\item{7-661}{Roll call vote (numeric)}
\item{7-662}{Roll call vote (numeric)}
\item{7-663}{Roll call vote (numeric)}
\item{7-664}{Roll call vote (numeric)}
\item{7-665}{Roll call vote (numeric)}
\item{7-666}{Roll call vote (numeric)}
\item{7-667}{Roll call vote (numeric)}
\item{7-668}{Roll call vote (numeric)}
\item{7-669}{Roll call vote (numeric)}
\item{7-67}{Roll call vote (numeric)}
\item{7-670}{Roll call vote (numeric)}
\item{7-671}{Roll call vote (numeric)}
\item{7-672}{Roll call vote (numeric)}
\item{7-673}{Roll call vote (numeric)}
\item{7-674}{Roll call vote (numeric)}
\item{7-675}{Roll call vote (numeric)}
\item{7-676}{Roll call vote (numeric)}
\item{7-677}{Roll call vote (numeric)}
\item{7-678}{Roll call vote (numeric)}
\item{7-679}{Roll call vote (numeric)}
\item{7-68}{Roll call vote (numeric)}
\item{7-680}{Roll call vote (numeric)}
\item{7-681}{Roll call vote (numeric)}
\item{7-682}{Roll call vote (numeric)}
\item{7-683}{Roll call vote (numeric)}
\item{7-684}{Roll call vote (numeric)}
\item{7-685}{Roll call vote (numeric)}
\item{7-686}{Roll call vote (numeric)}
\item{7-687}{Roll call vote (numeric)}
\item{7-688}{Roll call vote (numeric)}
\item{7-689}{Roll call vote (numeric)}
\item{7-69}{Roll call vote (numeric)}
\item{7-690}{Roll call vote (numeric)}
\item{7-691}{Roll call vote (numeric)}
\item{7-692}{Roll call vote (numeric)}
\item{7-693}{Roll call vote (numeric)}
\item{7-694}{Roll call vote (numeric)}
\item{7-695}{Roll call vote (numeric)}
\item{7-696}{Roll call vote (numeric)}
\item{7-697}{Roll call vote (numeric)}
\item{7-698}{Roll call vote (numeric)}
\item{7-699}{Roll call vote (numeric)}
\item{7-7}{Roll call vote (numeric)}
\item{7-70}{Roll call vote (numeric)}
\item{7-700}{Roll call vote (numeric)}
\item{7-701}{Roll call vote (numeric)}
\item{7-702}{Roll call vote (numeric)}
\item{7-703}{Roll call vote (numeric)}
\item{7-704}{Roll call vote (numeric)}
\item{7-705}{Roll call vote (numeric)}
\item{7-706}{Roll call vote (numeric)}
\item{7-707}{Roll call vote (numeric)}
\item{7-708}{Roll call vote (numeric)}
\item{7-709}{Roll call vote (numeric)}
\item{7-71}{Roll call vote (numeric)}
\item{7-710}{Roll call vote (numeric)}
\item{7-711}{Roll call vote (numeric)}
\item{7-712}{Roll call vote (numeric)}
\item{7-713}{Roll call vote (numeric)}
\item{7-714}{Roll call vote (numeric)}
\item{7-715}{Roll call vote (numeric)}
\item{7-716}{Roll call vote (numeric)}
\item{7-717}{Roll call vote (numeric)}
\item{7-718}{Roll call vote (numeric)}
\item{7-719}{Roll call vote (numeric)}
\item{7-72}{Roll call vote (numeric)}
\item{7-720}{Roll call vote (numeric)}
\item{7-721}{Roll call vote (numeric)}
\item{7-722}{Roll call vote (numeric)}
\item{7-723}{Roll call vote (numeric)}
\item{7-724}{Roll call vote (numeric)}
\item{7-725}{Roll call vote (numeric)}
\item{7-726}{Roll call vote (numeric)}
\item{7-727}{Roll call vote (numeric)}
\item{7-728}{Roll call vote (numeric)}
\item{7-729}{Roll call vote (numeric)}
\item{7-73}{Roll call vote (numeric)}
\item{7-730}{Roll call vote (numeric)}
\item{7-731}{Roll call vote (numeric)}
\item{7-732}{Roll call vote (numeric)}
\item{7-733}{Roll call vote (numeric)}
\item{7-734}{Roll call vote (numeric)}
\item{7-735}{Roll call vote (numeric)}
\item{7-736}{Roll call vote (numeric)}
\item{7-737}{Roll call vote (numeric)}
\item{7-738}{Roll call vote (numeric)}
\item{7-739}{Roll call vote (numeric)}
\item{7-74}{Roll call vote (numeric)}
\item{7-740}{Roll call vote (numeric)}
\item{7-741}{Roll call vote (numeric)}
\item{7-742}{Roll call vote (numeric)}
\item{7-743}{Roll call vote (numeric)}
\item{7-744}{Roll call vote (numeric)}
\item{7-745}{Roll call vote (numeric)}
\item{7-746}{Roll call vote (numeric)}
\item{7-747}{Roll call vote (numeric)}
\item{7-748}{Roll call vote (numeric)}
\item{7-749}{Roll call vote (numeric)}
\item{7-75}{Roll call vote (numeric)}
\item{7-750}{Roll call vote (numeric)}
\item{7-751}{Roll call vote (numeric)}
\item{7-752}{Roll call vote (numeric)}
\item{7-753}{Roll call vote (numeric)}
\item{7-754}{Roll call vote (numeric)}
\item{7-755}{Roll call vote (numeric)}
\item{7-756}{Roll call vote (numeric)}
\item{7-757}{Roll call vote (numeric)}
\item{7-758}{Roll call vote (numeric)}
\item{7-759}{Roll call vote (numeric)}
\item{7-76}{Roll call vote (numeric)}
\item{7-760}{Roll call vote (numeric)}
\item{7-761}{Roll call vote (numeric)}
\item{7-762}{Roll call vote (numeric)}
\item{7-763}{Roll call vote (numeric)}
\item{7-764}{Roll call vote (numeric)}
\item{7-765}{Roll call vote (numeric)}
\item{7-766}{Roll call vote (numeric)}
\item{7-767}{Roll call vote (numeric)}
\item{7-768}{Roll call vote (numeric)}
\item{7-769}{Roll call vote (numeric)}
\item{7-77}{Roll call vote (numeric)}
\item{7-770}{Roll call vote (numeric)}
\item{7-771}{Roll call vote (numeric)}
\item{7-772}{Roll call vote (numeric)}
\item{7-773}{Roll call vote (numeric)}
\item{7-774}{Roll call vote (numeric)}
\item{7-775}{Roll call vote (numeric)}
\item{7-776}{Roll call vote (numeric)}
\item{7-777}{Roll call vote (numeric)}
\item{7-778}{Roll call vote (numeric)}
\item{7-779}{Roll call vote (numeric)}
\item{7-78}{Roll call vote (numeric)}
\item{7-780}{Roll call vote (numeric)}
\item{7-781}{Roll call vote (numeric)}
\item{7-782}{Roll call vote (numeric)}
\item{7-783}{Roll call vote (numeric)}
\item{7-784}{Roll call vote (numeric)}
\item{7-785}{Roll call vote (numeric)}
\item{7-786}{Roll call vote (numeric)}
\item{7-787}{Roll call vote (numeric)}
\item{7-788}{Roll call vote (numeric)}
\item{7-789}{Roll call vote (numeric)}
\item{7-79}{Roll call vote (numeric)}
\item{7-790}{Roll call vote (numeric)}
\item{7-791}{Roll call vote (numeric)}
\item{7-792}{Roll call vote (numeric)}
\item{7-793}{Roll call vote (numeric)}
\item{7-794}{Roll call vote (numeric)}
\item{7-795}{Roll call vote (numeric)}
\item{7-796}{Roll call vote (numeric)}
\item{7-797}{Roll call vote (numeric)}
\item{7-798}{Roll call vote (numeric)}
\item{7-799}{Roll call vote (numeric)}
\item{7-8}{Roll call vote (numeric)}
\item{7-80}{Roll call vote (numeric)}
\item{7-800}{Roll call vote (numeric)}
\item{7-801}{Roll call vote (numeric)}
\item{7-802}{Roll call vote (numeric)}
\item{7-803}{Roll call vote (numeric)}
\item{7-804}{Roll call vote (numeric)}
\item{7-805}{Roll call vote (numeric)}
\item{7-806}{Roll call vote (numeric)}
\item{7-807}{Roll call vote (numeric)}
\item{7-808}{Roll call vote (numeric)}
\item{7-809}{Roll call vote (numeric)}
\item{7-81}{Roll call vote (numeric)}
\item{7-810}{Roll call vote (numeric)}
\item{7-811}{Roll call vote (numeric)}
\item{7-812}{Roll call vote (numeric)}
\item{7-813}{Roll call vote (numeric)}
\item{7-814}{Roll call vote (numeric)}
\item{7-815}{Roll call vote (numeric)}
\item{7-816}{Roll call vote (numeric)}
\item{7-817}{Roll call vote (numeric)}
\item{7-818}{Roll call vote (numeric)}
\item{7-819}{Roll call vote (numeric)}
\item{7-82}{Roll call vote (numeric)}
\item{7-820}{Roll call vote (numeric)}
\item{7-821}{Roll call vote (numeric)}
\item{7-822}{Roll call vote (numeric)}
\item{7-823}{Roll call vote (numeric)}
\item{7-824}{Roll call vote (numeric)}
\item{7-825}{Roll call vote (numeric)}
\item{7-826}{Roll call vote (numeric)}
\item{7-827}{Roll call vote (numeric)}
\item{7-828}{Roll call vote (numeric)}
\item{7-829}{Roll call vote (numeric)}
\item{7-83}{Roll call vote (numeric)}
\item{7-830}{Roll call vote (numeric)}
\item{7-831}{Roll call vote (numeric)}
\item{7-832}{Roll call vote (numeric)}
\item{7-833}{Roll call vote (numeric)}
\item{7-834}{Roll call vote (numeric)}
\item{7-835}{Roll call vote (numeric)}
\item{7-836}{Roll call vote (numeric)}
\item{7-837}{Roll call vote (numeric)}
\item{7-838}{Roll call vote (numeric)}
\item{7-839}{Roll call vote (numeric)}
\item{7-84}{Roll call vote (numeric)}
\item{7-840}{Roll call vote (numeric)}
\item{7-841}{Roll call vote (numeric)}
\item{7-842}{Roll call vote (numeric)}
\item{7-843}{Roll call vote (numeric)}
\item{7-844}{Roll call vote (numeric)}
\item{7-845}{Roll call vote (numeric)}
\item{7-846}{Roll call vote (numeric)}
\item{7-847}{Roll call vote (numeric)}
\item{7-848}{Roll call vote (numeric)}
\item{7-849}{Roll call vote (numeric)}
\item{7-85}{Roll call vote (numeric)}
\item{7-850}{Roll call vote (numeric)}
\item{7-851}{Roll call vote (numeric)}
\item{7-852}{Roll call vote (numeric)}
\item{7-853}{Roll call vote (numeric)}
\item{7-854}{Roll call vote (numeric)}
\item{7-855}{Roll call vote (numeric)}
\item{7-856}{Roll call vote (numeric)}
\item{7-857}{Roll call vote (numeric)}
\item{7-858}{Roll call vote (numeric)}
\item{7-859}{Roll call vote (numeric)}
\item{7-86}{Roll call vote (numeric)}
\item{7-860}{Roll call vote (numeric)}
\item{7-861}{Roll call vote (numeric)}
\item{7-862}{Roll call vote (numeric)}
\item{7-863}{Roll call vote (numeric)}
\item{7-864}{Roll call vote (numeric)}
\item{7-865}{Roll call vote (numeric)}
\item{7-866}{Roll call vote (numeric)}
\item{7-867}{Roll call vote (numeric)}
\item{7-868}{Roll call vote (numeric)}
\item{7-869}{Roll call vote (numeric)}
\item{7-87}{Roll call vote (numeric)}
\item{7-870}{Roll call vote (numeric)}
\item{7-871}{Roll call vote (numeric)}
\item{7-872}{Roll call vote (numeric)}
\item{7-873}{Roll call vote (numeric)}
\item{7-874}{Roll call vote (numeric)}
\item{7-875}{Roll call vote (numeric)}
\item{7-876}{Roll call vote (numeric)}
\item{7-877}{Roll call vote (numeric)}
\item{7-878}{Roll call vote (numeric)}
\item{7-879}{Roll call vote (numeric)}
\item{7-88}{Roll call vote (numeric)}
\item{7-880}{Roll call vote (numeric)}
\item{7-881}{Roll call vote (numeric)}
\item{7-882}{Roll call vote (numeric)}
\item{7-883}{Roll call vote (numeric)}
\item{7-884}{Roll call vote (numeric)}
\item{7-885}{Roll call vote (numeric)}
\item{7-886}{Roll call vote (numeric)}
\item{7-887}{Roll call vote (numeric)}
\item{7-888}{Roll call vote (numeric)}
\item{7-889}{Roll call vote (numeric)}
\item{7-89}{Roll call vote (numeric)}
\item{7-890}{Roll call vote (numeric)}
\item{7-891}{Roll call vote (numeric)}
\item{7-892}{Roll call vote (numeric)}
\item{7-893}{Roll call vote (numeric)}
\item{7-894}{Roll call vote (numeric)}
\item{7-895}{Roll call vote (numeric)}
\item{7-896}{Roll call vote (numeric)}
\item{7-897}{Roll call vote (numeric)}
\item{7-898}{Roll call vote (numeric)}
\item{7-899}{Roll call vote (numeric)}
\item{7-9}{Roll call vote (numeric)}
\item{7-90}{Roll call vote (numeric)}
\item{7-900}{Roll call vote (numeric)}
\item{7-901}{Roll call vote (numeric)}
\item{7-902}{Roll call vote (numeric)}
\item{7-903}{Roll call vote (numeric)}
\item{7-904}{Roll call vote (numeric)}
\item{7-905}{Roll call vote (numeric)}
\item{7-906}{Roll call vote (numeric)}
\item{7-907}{Roll call vote (numeric)}
\item{7-908}{Roll call vote (numeric)}
\item{7-909}{Roll call vote (numeric)}
\item{7-91}{Roll call vote (numeric)}
\item{7-910}{Roll call vote (numeric)}
\item{7-911}{Roll call vote (numeric)}
\item{7-912}{Roll call vote (numeric)}
\item{7-913}{Roll call vote (numeric)}
\item{7-914}{Roll call vote (numeric)}
\item{7-915}{Roll call vote (numeric)}
\item{7-916}{Roll call vote (numeric)}
\item{7-917}{Roll call vote (numeric)}
\item{7-918}{Roll call vote (numeric)}
\item{7-919}{Roll call vote (numeric)}
\item{7-92}{Roll call vote (numeric)}
\item{7-920}{Roll call vote (numeric)}
\item{7-921}{Roll call vote (numeric)}
\item{7-922}{Roll call vote (numeric)}
\item{7-923}{Roll call vote (numeric)}
\item{7-924}{Roll call vote (numeric)}
\item{7-925}{Roll call vote (numeric)}
\item{7-926}{Roll call vote (numeric)}
\item{7-927}{Roll call vote (numeric)}
\item{7-928}{Roll call vote (numeric)}
\item{7-929}{Roll call vote (numeric)}
\item{7-93}{Roll call vote (numeric)}
\item{7-930}{Roll call vote (numeric)}
\item{7-931}{Roll call vote (numeric)}
\item{7-932}{Roll call vote (numeric)}
\item{7-933}{Roll call vote (numeric)}
\item{7-934}{Roll call vote (numeric)}
\item{7-935}{Roll call vote (numeric)}
\item{7-936}{Roll call vote (numeric)}
\item{7-937}{Roll call vote (numeric)}
\item{7-938}{Roll call vote (numeric)}
\item{7-939}{Roll call vote (numeric)}
\item{7-94}{Roll call vote (numeric)}
\item{7-940}{Roll call vote (numeric)}
\item{7-941}{Roll call vote (numeric)}
\item{7-942}{Roll call vote (numeric)}
\item{7-943}{Roll call vote (numeric)}
\item{7-944}{Roll call vote (numeric)}
\item{7-945}{Roll call vote (numeric)}
\item{7-946}{Roll call vote (numeric)}
\item{7-947}{Roll call vote (numeric)}
\item{7-948}{Roll call vote (numeric)}
\item{7-949}{Roll call vote (numeric)}
\item{7-95}{Roll call vote (numeric)}
\item{7-950}{Roll call vote (numeric)}
\item{7-951}{Roll call vote (numeric)}
\item{7-952}{Roll call vote (numeric)}
\item{7-953}{Roll call vote (numeric)}
\item{7-954}{Roll call vote (numeric)}
\item{7-955}{Roll call vote (numeric)}
\item{7-956}{Roll call vote (numeric)}
\item{7-957}{Roll call vote (numeric)}
\item{7-958}{Roll call vote (numeric)}
\item{7-959}{Roll call vote (numeric)}
\item{7-96}{Roll call vote (numeric)}
\item{7-960}{Roll call vote (numeric)}
\item{7-961}{Roll call vote (numeric)}
\item{7-962}{Roll call vote (numeric)}
\item{7-963}{Roll call vote (numeric)}
\item{7-964}{Roll call vote (numeric)}
\item{7-965}{Roll call vote (numeric)}
\item{7-966}{Roll call vote (numeric)}
\item{7-967}{Roll call vote (numeric)}
\item{7-968}{Roll call vote (numeric)}
\item{7-969}{Roll call vote (numeric)}
\item{7-97}{Roll call vote (numeric)}
\item{7-970}{Roll call vote (numeric)}
\item{7-971}{Roll call vote (numeric)}
\item{7-972}{Roll call vote (numeric)}
\item{7-973}{Roll call vote (numeric)}
\item{7-974}{Roll call vote (numeric)}
\item{7-975}{Roll call vote (numeric)}
\item{7-976}{Roll call vote (numeric)}
\item{7-977}{Roll call vote (numeric)}
\item{7-978}{Roll call vote (numeric)}
\item{7-979}{Roll call vote (numeric)}
\item{7-98}{Roll call vote (numeric)}
\item{7-980}{Roll call vote (numeric)}
\item{7-981}{Roll call vote (numeric)}
\item{7-982}{Roll call vote (numeric)}
\item{7-983}{Roll call vote (numeric)}
\item{7-984}{Roll call vote (numeric)}
\item{7-985}{Roll call vote (numeric)}
\item{7-986}{Roll call vote (numeric)}
\item{7-987}{Roll call vote (numeric)}
\item{7-988}{Roll call vote (numeric)}
\item{7-989}{Roll call vote (numeric)}
\item{7-99}{Roll call vote (numeric)}
\item{7-990}{Roll call vote (numeric)}
\item{7-991}{Roll call vote (numeric)}
\item{7-992}{Roll call vote (numeric)}
\item{7-993}{Roll call vote (numeric)}
\item{7-994}{Roll call vote (numeric)}
\item{7-995}{Roll call vote (numeric)}
\item{7-996}{Roll call vote (numeric)}
\item{7-997}{Roll call vote (numeric)}
\item{7-998}{Roll call vote (numeric)}
\item{7-999}{Roll call vote (numeric)}
}
}
\source{
Yen-Chieh Liao (2024). Electoral Reform and Fragmented Polarization:
New Evidence from Taiwan Legislative Roll Call. Legislative Studies Quarterly.
\doi{10.1111/lsq.12459}
}
\usage{
data(legis_7th_Taiwan)
}
\description{
This dataset contains roll call voting data from the 7th Legislative Yuan
(National Congress) of Taiwan. The dataset includes the names of legislators
and their corresponding votes on various bills.
}
\details{
The data captures the legislative behavior during the 7th session of the
Legislative Yuan of Taiwan, providing valuable insights into the political
dynamics and decision-making processes. Each row represents one legislator,
and each vote column (7-1 through 7-1226) represents a specific bill or
motion voted upon.
}
\examples{
\dontrun{
data(legis_7th_Taiwan)

# View structure
str(legis_7th_Taiwan)

# First few legislators and votes
head(legis_7th_Taiwan[, c(1:5, 1227:1228)])

# Summary by party
table(legis_7th_Taiwan$party)
}

}
\keyword{datasets}
