% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCutlineAngles.R
\name{makeCutlineAngles}
\alias{makeCutlineAngles}
\title{Calculate Cutline Angles}
\usage{
makeCutlineAngles(obj)
}
\arguments{
\item{obj}{An object of class \code{wnominate}, which can either be a \code{nomObject} or \code{OCobject}. This object contains roll call data and other relevant information needed to calculate the cutline angles.}
}
\value{
A data frame with three columns:
\describe{
\item{angle}{The calculated cutline angles in degrees.}
\item{N1W}{The adjusted first component of the normal vector to the cutline.}
\item{N2W}{The adjusted second component of the normal vector to the cutline.}
}
}
\description{
The \code{makeCutlineAngles} function calculates the cutline angles for a given object of class \code{wnominate}, which can either be a \code{nomObject} or \code{OCobject}. The function processes the roll call data and computes the angles associated with the cutlines.
}
\details{
The function differentiates between objects of class \code{nomObject} and \code{OCobject} to perform appropriate calculations. It uses the roll call data within the object to compute the normal vectors and their corresponding angles.
}
\examples{
\dontrun{
# Assuming `wnominate_obj` is a valid wnominate object of class `nomObject` or `OCobject`
result <- makeCutlineAngles(wnominate_obj)
print(result)
}

}
\keyword{datasets}
