% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{mexicoCSES2000}
\alias{mexicoCSES2000}
\title{Mexican Political Party Positions on Left-Right Scale (2000 & 2006)}
\format{
A data frame with the following variables representing major political parties in Mexico:
\describe{
\item{PAN}{Left-right placement of the National Action Party (Partido Acción Nacional).}
\item{PRI}{Left-right placement of the Institutional Revolutionary Party (Partido Revolucionario Institucional).}
\item{PRD}{Left-right placement of the Party of the Democratic Revolution (Partido de la Revolución Democrática).}
\item{PT}{Left-right placement of the Labor Party (Partido del Trabajo).}
\item{Greens}{Left-right placement of the Green Party (Partido Verde Ecologista de México).}
\item{PARM}{Left-right placement of the Authentic Party of the Mexican Revolution (Partido Auténtico de la Revolución Mexicana).}
}
}
\source{
Data from the 2000 and 2006 Mexican modules of the Comparative Study of Electoral Systems (CSES).
}
\usage{
data(mexicoCSES2000)
}
\description{
The \code{mexicoCSES2000} dataset contains data from the 2000 and 2006 Mexican modules of the Comparative Study of Electoral Systems (CSES).
In these surveys, Mexican citizens were asked to place the major political parties on an 11-point left-right scale.
}
\details{
The data were collected as part of the 2000 and 2006 Mexican modules of the Comparative Study of Electoral Systems (CSES).
In these surveys, respondents were asked to rate the major political parties in Mexico on an 11-point left-right ideological scale.
}
\examples{
\dontrun{
data(mexicoCSES2000)
}

}
\keyword{datasets}
