% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{mexicoCSES2006}
\alias{mexicoCSES2006}
\title{Mexican Module of the Comparative Study of Electoral Systems (CSES) 2000 and 2006}
\format{
A data frame with rows representing respondents and 8 columns representing different political parties:
\describe{
\item{PAN}{Numeric, placement of the PAN (National Action Party) on the left-right scale.}
\item{PRD}{Numeric, placement of the PRD (Party of the Democratic Revolution) on the left-right scale.}
\item{PRI}{Numeric, placement of the PRI (Institutional Revolutionary Party) on the left-right scale.}
\item{Greens}{Numeric, placement of the Green Party on the left-right scale.}
\item{PT}{Numeric, placement of the PT (Labor Party) on the left-right scale.}
\item{Convergencia}{Numeric, placement of the Convergencia (Convergence) party on the left-right scale.}
\item{Nueva Alianza}{Numeric, placement of the Nueva Alianza (New Alliance) party on the left-right scale.}
\item{PSD}{Numeric, placement of the PSD (Social Democratic Party) on the left-right scale.}
}
}
\source{
Comparative Study of Electoral Systems (CSES), Mexican modules for 2000 and 2006.
}
\usage{
data(mexicoCSES2006)
}
\description{
The \code{mexicoCSES2006} dataset contains data from the 2000 and 2006 Mexican modules of the Comparative Study of Electoral Systems (CSES). In these surveys, Mexican citizens were asked to place the major political parties on an 11-point left-right scale.
}
\details{
The dataset includes responses from the 2000 and 2006 Mexican modules of the Comparative Study of Electoral Systems (CSES), where respondents were asked to place major political parties on an 11-point left-right scale.
}
\examples{
\dontrun{
data(mexicoCSES2006)
}

}
\keyword{datasets}
