% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nation}
\alias{nation}
\title{Nation Similarity Ratings Dataset}
\format{
A 12x12 matrix with the following column and row names representing nations:
\describe{
\item{Brazil}{Average similarity ratings involving Brazil.}
\item{Congo}{Average similarity ratings involving Congo.}
\item{Cuba}{Average similarity ratings involving Cuba.}
\item{Egypt}{Average similarity ratings involving Egypt.}
\item{France}{Average similarity ratings involving France.}
\item{India}{Average similarity ratings involving India.}
\item{Israel}{Average similarity ratings involving Israel.}
\item{Japan}{Average similarity ratings involving Japan.}
\item{China}{Average similarity ratings involving China.}
\item{USSR}{Average similarity ratings involving the Soviet Union (USSR).}
\item{USA}{Average similarity ratings involving the United States of America (USA).}
\item{Yugoslavia}{Average similarity ratings involving Yugoslavia.}
}
}
\source{
Wish, M. (1971). Individual differences in perceptions and preferences among nations. \emph{In Proceedings of the 79th Annual Convention of the American Psychological Association}.
}
\usage{
data(nations)
}
\description{
The \code{nation} dataset contains similarity ratings between twelve nations, as collected by Wish (1971).
In 1968, Wish asked 18 students in his psychological measurement class to rate the perceived similarity between each pair of twelve nations using a 9-point scale,
where '1' indicates "very different" and '9' indicates "very similar". The data in this dataset represent the average similarity ratings between these nations.
}
\details{
The \code{nations} dataset was constructed by averaging the similarity ratings provided by 18 students in Wish's 1968 psychological measurement class.
The dataset is a symmetric matrix where both the rows and columns represent the twelve nations, and each cell contains the average similarity rating
between the corresponding pair of nations.
}
\examples{
\dontrun{
data(nations)
}

}
\keyword{datasets}
