% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{nes1968_urbanunrest}
\alias{nes1968_urbanunrest}
\title{Urban Unrest Issue Scales from the 1968 National Election Study (NES)}
\format{
A data frame with rows representing respondents and the following columns:
\describe{
\item{vote.choice}{Categorical, respondent's reported vote choice in the 1968 presidential election. This variable indicates which of the three major presidential candidates (Humphrey, Nixon, or Wallace) the respondent voted for, if any.}
\item{johnson}{Numeric, placement of President Lyndon Johnson on the Vietnam War scale (1 to 7).}
\item{humphrey}{Numeric, placement of Democratic candidate Hubert Humphrey on the Vietnam War scale (1 to 7).}
\item{nixon}{Numeric, placement of Republican candidate Richard Nixon on the Vietnam War scale (1 to 7).}
\item{wallace}{Numeric, placement of American Independent candidate George Wallace on the Vietnam War scale (1 to 7).}
\item{self}{Numeric, respondent's self-placement on the Vietnam War scale (1 to 7).}
}
}
\source{
1968 American National Election Study (ANES).
}
\usage{
data(nes1968_urbanunrest)
}
\description{
The \code{nes1968_urbanunrest} dataset contains responses from the 1968 National Election Study (NES) where respondents were asked to place themselves, President Lyndon Johnson, and the three major presidential candidates—Democrat Hubert Humphrey, Republican Richard Nixon, and American Independent George Wallace—on two seven-point issue scales regarding urban unrest.
}
\details{
This dataset is part of the 1968 National Election Study (NES). Respondents were asked to place themselves and key political figures on two seven-point scales relating to urban unrest. These scales measure opinions on how the issue of urban unrest should be addressed.
}
\examples{
\dontrun{
data(nes1968_urbanunrest)
}

}
\keyword{datasets}
