% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{original}
\alias{original}
\title{Transposed Rankings Data using Blackbox Method}
\format{
A list object containing the results of the \code{blackbox_transpose} function.
}
\usage{
data(original)
}
\description{
The \code{original} object is created by applying the \code{blackbox_transpose} function to a dataset of rankings.
This function is used to perform multidimensional scaling on a set of rankings, handling missing values
and using specified dimensions and scaling parameters.
}
\details{
The \code{original} object was created using the following call:
\preformatted{
original <- blackbox_transpose(rankings,
                               missing = c(77, 88, 89),
                               dims = 3,
                               minscale = 5,
                               verbose = FALSE)
}
\itemize{
\item \strong{rankings}: A dataset containing the rankings to be analyzed.
\item \strong{missing}: A vector of values that are treated as missing data.
\item \strong{dims}: The number of dimensions to be used in the multidimensional scaling.
\item \strong{minscale}: The minimum scale to be applied in the scaling process.
\item \strong{verbose}: A logical flag indicating whether to print detailed output during the process.
}

The \code{blackbox_transpose} function is typically used to analyze and visualize multidimensional data, allowing researchers
to understand the underlying structure of ranked data by transforming it into a lower-dimensional space.
}
\examples{
\dontrun{
data(original)
}

}
\keyword{datasets}
