% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.aldmck_ci.R
\name{plot.aldmck_ci}
\alias{plot.aldmck_ci}
\title{Plot Confidence Intervals for Aldmck CI}
\usage{
\method{plot}{aldmck_ci}(x, ...)
}
\arguments{
\item{x}{An object of class \code{aldmck_ci} containing ideal points and confidence intervals.}

\item{...}{Additional arguments passed to \code{ggplot2} functions.}
}
\value{
A ggplot2 object representing the confidence intervals for the ideal points of stimuli.
}
\description{
This function creates a plot of the ideal points and their confidence intervals for an object of class \code{aldmck_ci}.
}
\examples{
\dontrun{
data(result.france)
plot.aldmck_ci(result.france$sumstats)
}
}
