% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mlsmu6.R
\name{plot.mlsmu6}
\alias{plot.mlsmu6}
\title{Plot Method for \code{mlsmu6} Class Objects}
\usage{
\method{plot}{mlsmu6}(x, ..., selected.stims = NULL, ind.id.size = 3, stim.id.size = 6)
}
\arguments{
\item{x}{An object of class \code{mlsmu6}, typically the output from the \code{mlsmu6} function.}

\item{...}{Additional arguments passed to the plot method (currently not used).}

\item{selected.stims}{A vector of stimulus names to highlight in the plot. If \code{NULL}, all stimuli are plotted.}

\item{ind.id.size}{Numeric value specifying the text size for individual labels in the plot. Default is 3.}

\item{stim.id.size}{Numeric value specifying the text size for stimulus labels in the plot. Default is 6.}
}
\value{
A \code{ggplot} object representing the 2D plot of the estimated positions.
}
\description{
This function creates a 2D plot of the results from the \code{mlsmu6} function. It visualizes the estimated positions of both individuals and stimuli in a lower-dimensional space.
}
\details{
The plot shows individuals and stimuli in a 2D space defined by the first two dimensions (\code{Dim1} and \code{Dim2}). Individuals are represented with smaller labels, while stimuli are larger, making them more prominent. If \code{selected.stims} is specified, only the selected stimuli will be highlighted, while the rest will be omitted from the plot.
}
\examples{
\dontrun{
# Assuming `result` is the output from `mlsmu6` function:
data(mlsmu6_out)
plot.mlsmu6(mlsmu6_out) 
}

}
