% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_blackbox.R
\name{plot_blackbox}
\alias{plot_blackbox}
\title{Plot Blackbox Results with Optional Grouping and Issue Vectors}
\usage{
plot_blackbox(
  result,
  dims,
  whichRes = NULL,
  groupVar = NULL,
  issueVector = NULL,
  data = NULL,
  missing = NULL,
  rug = FALSE,
  xlab = NULL,
  main = NULL,
  ylab = NULL,
  nudgeX = NULL,
  nudgeY = NULL,
  ...
)
}
\arguments{
\item{result}{A list object containing the results of the blackbox analysis. This object should include individual-level data for each dimension.}

\item{dims}{A numeric vector of length 2 indicating the dimensions to plot.}

\item{whichRes}{An optional integer specifying which result set to use from the \code{result$individuals} list. If \code{NULL}, the highest dimension is used.}

\item{groupVar}{An optional vector for grouping the data points. If \code{NULL}, no grouping is applied.}

\item{issueVector}{A vector indicating which issues to include as vectors in the plot. This can be a character vector of column names or a numeric vector of column indices.}

\item{data}{A data frame containing the original data used in the analysis. Required if \code{issueVector} is specified.}

\item{missing}{A vector of values to be treated as missing in the data.}

\item{rug}{Logical, if \code{TRUE}, adds marginal rugs to the plot.}

\item{xlab}{A character string for the x-axis label. If \code{NULL}, the label is generated automatically.}

\item{main}{A character string for the plot title. If \code{NULL}, no title is added.}

\item{ylab}{A character string for the y-axis label. If \code{NULL}, the label is generated automatically.}

\item{nudgeX}{A numeric vector for nudging the x positions of the issue vector labels. Defaults to \code{NULL}.}

\item{nudgeY}{A numeric vector for nudging the y positions of the issue vector labels. Defaults to \code{NULL}.}

\item{...}{Additional arguments passed to \code{ggplot2} functions.}
}
\value{
A \code{ggplot2} object representing the blackbox analysis plot.
}
\description{
This function creates a scatter plot of the results from a blackbox analysis. It allows for optional grouping of data points and the inclusion of issue vectors based on ordinal regression.
}
\examples{
\dontrun{
# CDS2000 is a dataset containing U.S. Congressional data,
# where the first column represents party affiliation: 1 for Democrats, 2 for Republicans.

# Recode the party variable: 1 = 'Democrat', 2 = 'Republican', others as NA
party <- car::recode(CDS2000[,1],
                     "1='Democrat'; 2='Republican'; else=NA",
                     as.factor=TRUE)

# Create a Blackbox plot to compare the distribution of different parties in a two-dimensional space
plot_blackbox(result.repdem, dims=c(1, 2), groupVar=party,
              xlab="First Dimension\n(Left-Right)",
              ylab="Second Dimension") +
  theme(legend.position="bottom", aspect.ratio=1) +
  guides(shape=guide_legend(override.aes=list(size=4))) +
  labs(colour="Party")
}
}
