% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_oc_coords.R
\name{plot_oc_coords}
\alias{plot_oc_coords}
\title{Plot OC Coordinates with Rotation}
\usage{
plot_oc_coords(
  obj,
  shapeVar = NULL,
  dropNV = FALSE,
  ptSize = 4,
  rotMat = diag(2)
)
}
\arguments{
\item{obj}{A list object containing the OC results. Typically, this comes from a function like \code{oc}.}

\item{shapeVar}{Optional. A variable to define the shape and color of points in the plot, typically a factor.}

\item{dropNV}{Logical. If \code{TRUE}, rows with missing values in the \code{shapeVar} will be dropped from the plot.}

\item{ptSize}{Numeric. The size of the points in the plot.}

\item{rotMat}{A 2x2 rotation matrix. Defaults to the identity matrix \code{diag(2)}, which means no rotation.}
}
\value{
A ggplot object representing the OC coordinates plot.
}
\description{
This function plots the OC (Optimal Classification) coordinates after applying an optional rotation matrix.
}
\examples{
\dontrun{
# Example usage:
# Assume `rc` is an object containing legislative data
result <- oc(rc, dims=2, minvotes=20, lop=0.025, polarity=c(2,2), verbose=FALSE)

# Define a 45-degree rotation matrix
deg2rad <- function(x) x * pi / 180
rad45 <- deg2rad(45)
A <- matrix(c(cos(rad45), -sin(rad45), sin(rad45), cos(rad45)), nrow=2, ncol=2, byrow=TRUE)

# Plot OC coordinates without rotation
plot1 <- plot_oc_coords(result)
print(plot1)

# Plot OC coordinates with a 45-degree rotation
plot2 <- plot_oc_coords(result, rotMat=A)
print(plot2)
}

}
