% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_OCcutlines2.R
\name{plot_OCcutlines2}
\alias{plot_OCcutlines2}
\title{Plot Optimal Classification Cutlines in 2D}
\usage{
plot_OCcutlines2(
  x,
  main.title = "Cutting Lines",
  d1.title = "First Dimension",
  d2.title = "Second Dimension",
  lines = 50,
  dims = c(1, 2),
  lwd = 2,
  ...
)
}
\arguments{
\item{x}{An object of class \code{OCobject}. This object should contain the results of an Optimal Classification analysis, including roll call data and norm vectors for plotting cutlines.}

\item{main.title}{A character string specifying the title of the plot. Defaults to \code{"Cutting Lines"}.}

\item{d1.title}{A character string specifying the label for the first dimension (x-axis). Defaults to \code{"First Dimension"}.}

\item{d2.title}{A character string specifying the label for the second dimension (y-axis). Defaults to \code{"Second Dimension"}.}

\item{lines}{An integer specifying the number of cutlines to plot. If \code{lines} is a vector, it selects specific cutlines by their row indices. If \code{lines} is a single integer, it randomly selects that many cutlines to plot. Defaults to \code{50}.}

\item{dims}{A numeric vector of length 2 specifying the dimensions to be plotted. Defaults to \code{c(1, 2)}.}

\item{lwd}{A numeric value specifying the line width of the cutlines. Defaults to \code{2}.}

\item{...}{Additional graphical parameters passed to the \code{symbols} function.}
}
\value{
This function does not return a value but produces a object for plot.
}
\description{
The \code{plot_OCcutlines2} function plots cutlines for a 2D Optimal Classification (OC) object. It visualizes the decision boundaries (cutlines) in a two-dimensional space based on the OC analysis results.
}
\details{
The function checks that the input object \code{x} is of class \code{OCobject} and that it contains two dimensions for plotting. If the \code{lines} parameter is a single integer, it will randomly select that many cutlines to display. If \code{lines} is a vector, it will use those specific rows from the roll call data to plot the cutlines.#' @importFrom yourpackage add_OCcutline
}
\examples{
\dontrun{
# Assuming `oc_result` is an OCobject with 2D results
plot_OCcutlines2(oc_result, lines = 10)
}

}
