% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_resphist.R
\name{plot_resphist}
\alias{plot_resphist}
\title{Plot Respondent Ideal Points Histogram}
\usage{
plot_resphist(
  result,
  groupVar = NULL,
  facetVar = NULL,
  addStim = FALSE,
  scaleDensity = TRUE,
  weights = c("all", "positive", "negative"),
  xlab = NULL,
  main = NULL,
  ylab = NULL,
  whichRes = NULL,
  dim = NULL,
  ...
)
}
\arguments{
\item{result}{An object of class "aldmck" or "blackbox" containing respondent data.}

\item{groupVar}{Optional grouping variable.}

\item{facetVar}{Optional faceting variable.}

\item{addStim}{Logical, whether to add stimuli points to the plot.}

\item{scaleDensity}{Logical, whether to scale the density by group proportions.}

\item{weights}{Character, specifying which weights to use ("all", "positive", "negative").}

\item{xlab}{Character, label for the x-axis.}

\item{main}{Character, main title for the plot.}

\item{ylab}{Character, label for the y-axis.}

\item{whichRes}{Optional, which respondent data to use.}

\item{dim}{Optional, dimension for "blackbox" objects.}

\item{...}{Additional arguments passed to ggplot2 functions.}
}
\value{
A ggplot2 object
}
\description{
This function generates a histogram plot of respondent ideal points, optionally grouped or faceted by variables.
}
\examples{
\dontrun{
data(result.france)
plot_resphist(result.france, addStim = TRUE, xlab = "Left-Right") +
  theme(legend.position = "bottom", aspect.ratio = 1) +
  guides(shape = guide_legend(override.aes = list(size = 4), nrow = 3)) +
  labs(shape = "Party", colour = "Party")
}
}
