% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rollcall.R
\name{plot_rollcall}
\alias{plot_rollcall}
\title{Plot Roll Call Results}
\usage{
plot_rollcall(
  obj,
  data,
  gdat,
  rcnum,
  shapeVar = NULL,
  dropNV = FALSE,
  onlyErrors = FALSE,
  ptSize = 4
)
}
\arguments{
\item{obj}{An object of class \code{nomObject} containing the results of a NOMINATE analysis, including legislators' ideal points and roll call data.}

\item{data}{A data frame containing the roll call data, including vote outcomes and codes for "yea" and "nay" votes.}

\item{gdat}{A data frame containing additional data related to the legislators. (This parameter is currently not used in the function but can be included for future extensions.)}

\item{rcnum}{An integer specifying the roll call number to be plotted.}

\item{shapeVar}{An optional variable used to differentiate groups of legislators by shape in the plot. If provided, the points will be shaped according to this variable.}

\item{dropNV}{A logical value indicating whether to drop non-voters (NA values) from the plot. Default is \code{FALSE}.}

\item{onlyErrors}{A logical value indicating whether to plot only the legislators who made voting errors. Default is \code{FALSE}.}

\item{ptSize}{A numeric value specifying the size of the points in the plot. Default is \code{4}.}
}
\value{
A \code{ggplot} object showing the roll call results, including the voting patterns of legislators.
}
\description{
This function creates a scatter plot of legislators' ideal points for a given roll call vote. It allows for the visualization of voting patterns, highlighting errors, and optionally adding shape variation for different groups of legislators.
}
\examples{
\dontrun{
g <- plot_rollcall(result, hr108, gdat = NULL, rcnum = 528)
print(g)
}
}
