% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.aldmck_ci.R
\name{print.aldmck_ci}
\alias{print.aldmck_ci}
\title{Print Aldmck Confidence Intervals}
\usage{
\method{print}{aldmck_ci}(x, ..., digits = 3)
}
\arguments{
\item{x}{An object of class \code{aldmck_ci} containing ideal points and confidence intervals.}

\item{...}{Additional arguments passed to other methods (currently not used).}

\item{digits}{Integer, the number of decimal places to display (default is 3).}
}
\value{
The function prints the formatted \code{aldmck_ci} object as a data frame.
}
\description{
This function formats and prints an \code{aldmck_ci} object, displaying the ideal points and their associated confidence intervals with specified precision.
}
\examples{
\dontrun{
data(result.france)
print.aldmck_ci(result.france$sumstats, digits = 2)
}
}
