% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{rankings}
\alias{rankings}
\title{Rankings Data from France EES 2009}
\format{
A numeric matrix with rows representing respondents and columns representing the rankings of different political parties.
}
\source{
France 2009 European Election Study (EES)
}
\description{
This dataset, \code{rankings}, contains the rankings data extracted from the France 2009 European Election Study (EES). It is a matrix of numeric rankings for various political parties.
}
\details{
The \code{rankings} matrix was created from the \code{franceEES2009} dataset by selecting columns 2 to 9, which correspond to the respondents' rankings of different political parties. The data was then converted to numeric mode for further analysis.

The following code was used to create the \code{rankings} object:

\code{
rankings <- as.matrix(franceEES2009[,2:9])
mode(rankings) <- "numeric"
}

This dataset is used as an input for the \code{blackbox_transpose} function to perform dimensional analysis.
}
\keyword{datasets}
