% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.errors.R
\name{rc.errors}
\alias{rc.errors}
\title{Calculate Roll Call Errors and Proportional Reduction in Error (PRE)}
\usage{
rc.errors(obj, data, rcnum, rotMat = diag(2))
}
\arguments{
\item{obj}{An object of class \code{nomObject} or \code{OCobject} containing the results of a NOMINATE or Optimal Classification analysis.}

\item{data}{A data frame containing the roll call data, including vote outcomes and codes for "yea" and "nay" votes.}

\item{rcnum}{An integer specifying the roll call number for which errors and PRE should be calculated.}

\item{rotMat}{A 2x2 rotation matrix used to rotate the coordinates in the case of an \code{OCobject}. The default is the identity matrix.}
}
\value{
A list with the following components:
\itemize{
\item \code{tot.errors}: A named vector with the total number of "yea" and "nay" errors.
\item \code{PRE}: The Proportional Reduction in Error statistic, calculated as the proportion of errors reduced relative to the worst-case scenario.
\item \code{errors}: A matrix indicating which legislators made errors in their votes, with columns for "yea" and "nay" errors.
}
}
\description{
This function calculates the roll call errors and the Proportional Reduction in Error (PRE) statistic for a given roll call vote. The function can handle both "nomObject" and "OCobject" classes, typically representing NOMINATE or Optimal Classification results.
}
\examples{
\dontrun{
data(hr108)
library(wnominate)
result <- wnominate(hr108, ubeta=15, uweights=0.5, dims=2,
minvotes=20, lop=0.025, trials=1, polarity=c(1,5), verbose=FALSE)
rc.errors(result, hr108, 528)[c("tot.errors", "PRE")]
}
}
