% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{rc_ep}
\alias{rc_ep}
\title{Roll Call Data for U.S. Congress}
\format{
An object of class \code{"rollcall"} containing roll call vote data.
}
\source{
Poole, K. T., & Rosenthal, H. (Various Years). House and Senate Roll Call Data. Retrieved from \url{http://www.voteview.com}.
}
\usage{
data(rc_ep)
}
\description{
The \code{rc_ep} object is a roll call dataset compiled by Poole and Rosenthal. This dataset represents roll call votes in the U.S. Congress and has been processed into an object of class \code{rollcall()} for analysis.
}
\details{
Poole and Rosenthal have compiled House and Senate roll call datasets covering the history of the U.S. Congress. These datasets are maintained at \url{http://www.voteview.com}. The \code{rc_ep} object represents a specific subset of these data and is formatted as a \code{rollcall} object, suitable for various forms of legislative analysis.
}
\examples{
\dontrun{
data(rc_ep)
}

}
\keyword{datasets}
