% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{result.repdem}
\alias{result.repdem}
\title{Blackbox Analysis Results for Issues Matrix}
\format{
An object of class \code{blackbox} containing the results of the dimensional analysis, including the identified dimensions and other related statistics.
}
\value{
The \code{result.repdem} object contains the extracted dimensions, along with other statistics generated by the \code{blackbox} function.
}
\description{
The object \code{result.repdem} contains the results of applying the \code{blackbox} function to the \code{issues} matrix. This analysis was performed to extract dimensions that represent the underlying structure of the issues, specifically for a dataset containing political data.
}
\examples{
\dontrun{
data(result.repdem)
}
}
\keyword{datasets}
