% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{result_2006}
\alias{result_2006}
\title{Blackbox Transpose Analysis Result for Mexico CSES 2006}
\format{
A list containing the results of the Blackbox transpose analysis, including estimated dimensions, scaling factors, and other relevant metrics.
}
\source{
Generated by the \code{blackbox_transpose} function applied to the \code{mexicoCSES2006} dataset.
}
\usage{
result_2006
}
\description{
The \code{result_2006} object contains the results of a Blackbox transpose analysis performed on the \code{mexicoCSES2006} dataset. This analysis was carried out to examine the dimensional structure of the dataset with specific handling for missing data, scaling, and dimensionality reduction.
}
\details{
The \code{result_2006} object captures the outputs from the Blackbox transpose analysis, where the \code{mexicoCSES2006} dataset was analyzed across 3 dimensions with a minimum scaling factor of 5. Missing data were coded as \code{99}, and the analysis was performed with verbose output to provide detailed information on the process.
}
\keyword{datasets}
