% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sub.europe}
\alias{sub.europe}
\title{CHES EU Dataset: Party Means and Standard Deviations}
\format{
A data frame with the following variables:
\describe{
\item{party}{The name or identifier of the political party.}
\item{mean}{The mean placement of the party as calculated from the responses of the experts.}
\item{sd}{The standard deviation of the party placements across the experts.}
\item{country}{The country where the party is based.}
\item{vignette}{Indicator of whether the party is one of the vignette parties (1) or an actual party (0).}
}
}
\source{
Data from the 2010 wave of the Chapel Hill Expert Survey (CHES).
}
\usage{
data(ches_eu)
}
\description{
The \code{ches_eu} dataset contains data used to calculate party means and standard deviations across all CHES experts.
The dataset is based on the 2010 wave of the Chapel Hill Expert Survey (CHES), which includes 118 parties and 224 experts
from 14 member countries of the European Union. Over 160 experts placed all 3 vignette parties, and between 8 and 17 experts
placed each of the actual parties.
}
\details{
The dataset was used in the analysis presented in the book chapter "2.5 Using Anchoring Vignettes" (pages 58-60).
The data helps in understanding how experts from the Chapel Hill Expert Survey (CHES) placed different political parties
on various dimensions, providing a basis for calculating mean positions and standard deviations.

The 2010 wave of the CHES consists of 118 parties and 224 experts across 14 European Union member countries. Over 160 experts
placed all three vignette parties, while between 8 and 17 experts placed each of the actual parties.
}
\examples{
\dontrun{
data(ches_eu)
}

}
\keyword{datasets}
