% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{Sweden2010}
\alias{Sweden2010}
\title{2010 Swedish Parliamentary Candidate Survey}
\format{
A data frame with the following variables:
\describe{
\item{id}{Unique identifier for each candidate.}
\item{elected}{Indicator of whether the candidate was elected (1) or not (0).}
\item{party.name}{Name of the political party the candidate belongs to.}
\item{party.code}{Numeric code representing the political party.}
\item{govt.party}{Indicator of whether the candidate's party is part of the government coalition (1) or not (0).}
\item{left.right.self.fivept}{Left-right self-placement on a 1-5 scale.}
\item{congestion.taxes}{Opinion on congestion taxes.}
\item{highspeed.trains}{Opinion on high-speed trains.}
\item{hunt.wolves}{Opinion on hunting wolves.}
\item{nuclear.power}{Opinion on nuclear power.}
\item{gasoline.taxes}{Opinion on gasoline taxes.}
\item{museum.fees}{Opinion on museum fees.}
\item{online.piracy}{Opinion on online piracy.}
\item{state.TV}{Opinion on state TV.}
\item{refugee.cities}{Opinion on establishing refugee cities.}
\item{asylum.seekers}{Opinion on asylum seekers.}
\item{refugee.healthcare}{Opinion on healthcare for refugees.}
\item{teacher.veils}{Opinion on teachers wearing veils.}
\item{paternal.leave}{Opinion on paternal leave.}
\item{affirmative.action.universities}{Opinion on affirmative action in universities.}
\item{child.raising.allowance}{Opinion on child-raising allowance.}
\item{property.taxes.wealthy}{Opinion on property taxes for the wealthy.}
\item{wealth.tax}{Opinion on wealth tax.}
\item{tax.wealthy}{Opinion on taxing the wealthy.}
\item{tax.pensions}{Opinion on taxing pensions.}
\item{household.services.deduction}{Opinion on household services deduction.}
\item{work.income.tax}{Opinion on work income tax.}
\item{sex.purchase}{Opinion on purchasing sex.}
\item{DUI.penalty}{Opinion on penalties for driving under the influence (DUI).}
\item{criminal.sentences}{Opinion on criminal sentences.}
\item{wiretaps}{Opinion on wiretapping.}
\item{retirement.age}{Opinion on retirement age.}
\item{health.insurance.time}{Opinion on health insurance time limits.}
\item{dental.insurance}{Opinion on dental insurance.}
\item{competition.public.sector}{Opinion on competition in the public sector.}
\item{mandatory.unemployment.insurance}{Opinion on mandatory unemployment insurance.}
\item{municipal.home.care}{Opinion on municipal home care services.}
\item{circumcision}{Opinion on circumcision.}
\item{private.healthcare.profits}{Opinion on profits in private healthcare.}
\item{govt.alcohol.monopoly}{Opinion on the government's alcohol monopoly.}
\item{employment.protection}{Opinion on employment protection.}
\item{sell.public.corporations}{Opinion on selling public corporations.}
\item{Aghanistan.withdrawal}{Opinion on withdrawal from Afghanistan.}
\item{exporting.arms}{Opinion on exporting arms.}
\item{aid.undemocratic.countries}{Opinion on aid to undemocratic countries.}
\item{compulsory.military.service}{Opinion on compulsory military service.}
\item{leave.EU}{Opinion on leaving the European Union.}
\item{transfer.students}{Opinion on transferring students.}
\item{local.control.education}{Opinion on local control of education.}
\item{student.grades}{Opinion on student grading systems.}
\item{number.private.schools}{Opinion on the number of private schools.}
\item{university.eligibility}{Opinion on university eligibility.}
\item{criminalize.racist.organizations}{Opinion on criminalizing racist organizations.}
\item{abolish.monarchy}{Opinion on abolishing the monarchy.}
\item{ballot.order}{Opinion on the order of ballots.}
\item{referenda.elections}{Opinion on referenda during elections.}
}
}
\source{
Data collected by the Swedish public broadcasting network Sveriges Television (SVT).
}
\usage{
data(Sweden2010)
}
\description{
The \code{Sweden2010} dataset contains data from the 2010 Swedish Parliamentary Candidate Survey, conducted by the Swedish
public broadcasting network Sveriges Television (SVT). The survey targeted all 5,627 parliamentary candidates, with
completed interviews from 2,830 candidates, including 289 of the 349 candidates who were elected.
}
\details{
Candidates were asked 50 Likert-type questions, using a 4-point scale (from "strongly disagree" to "strongly agree") to register
their opinions on a series of policy statements. Most issue scales focus on economic/social welfare issues, but questions related
to foreign policy, social/cultural matters, law and order, immigration, and environmental issues are also included.
Missing responses are coded as 8.
}
\examples{
\dontrun{
data(Sweden2010)
}

}
\keyword{datasets}
