\name{uk}
\alias{uk}
\docType{data}
\title{UK Data from the MARPOR project}
\description{
A subset of the MARPOR data for Great Britain.
}
\usage{data(uk)}
\format{
A data frame with the following variables: 
\describe{
    \item{per105}{military -}
    \item{per107}{internationalism +}
    \item{per204}{constitution -}
    \item{per302}{centralisation}
    \item{per406}{protectionism +}
    \item{per504}{welfare +}
    \item{per506}{education +}
    \item{per602}{national way of life -}
    \item{per604}{traditional morality -}
    \item{per607}{multiculturalism +}
    \item{per701}{labour groups +}
    \item{per403}{market regulation}
    \item{per404}{economic planning}
    \item{per412}{controlled economy}
    \item{per413}{nationalisation}
    \item{per104}{military +}
    \item{per109}{internationalism -}
    \item{per203}{constitution +}
    \item{per301}{decentralisation}
    \item{per407}{protectionism -}
    \item{per505}{welfare -}
    \item{per507}{education -}
    \item{per601}{national way of life +}
    \item{per603}{traditional morality +}
    \item{per608}{multiculturalism -}
    \item{per702}{labour groups -}
    \item{per401}{free market economy}
    \item{per402}{incentives}
    \item{per414}{economic orthodoxy}
    \item{edate}{election date}
    \item{date}{election date (YYYYMM)}
    \item{party}{party abbreviation}
    \item{partyname}{party name}
    \item{rile}{right-left ideological index}
    \item{total}{total number of statements in manifesto}
    \item{elecnum}{Election numer (consecutive integers)}
    }
}
\source{
\url{https://manifesto-project.wzb.eu/}
}
\keyword{datasets}
