% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anes.input}
\alias{anes.input}
\title{ANES Input Data}
\format{
A numeric matrix with 12 columns, representing selected variables from the ANES 1968 dataset.
}
\source{
The data comes from the American National Election Study (ANES) 1968, available in the \code{asmcjr} package.
}
\description{
The \code{anes.input} object is a subset of the American National Election Study (ANES) 1968 dataset. It contains selected variables used as input for further analysis.
}
\details{
The \code{anes.input} object is created by loading the \code{ANES1968} dataset from the \code{asmcjr} package and selecting the first 12 columns. The data is then converted to a matrix format for analysis.

The steps to create \code{anes.input} are as follows:
\preformatted{
data(ANES1968)
anes.input <- ANES1968[, 1:12]
anes.input <- as.matrix(anes.input)
}
}
\examples{
\dontrun{
data(anes.input)
}
}
