\name{ANES1968}
\alias{ANES1968}
\docType{data}
\title{1968 American National Election Study
}
\description{
These data are a subset of the feeling thermometers along with vote turnout and vote intention from the 1968 American National Election Study.  
}
\usage{data(ANES1968)}
\format{
  A data frame with 1673 observations on the following 14 variables.
  \describe{
    \item{\code{Wallace}}{Respondent feeling thermometer score George Wallace}
    \item{\code{Humphrey}}{Respondent feeling thermometer score Hubert Humphrey}
    \item{\code{Nixon}}{Respondent feeling thermometer score Richard Nixon}
    \item{\code{McCarthy}}{Respondent feeling thermometer score Eugene McCarthy}
    \item{\code{Reagan}}{Respondent feeling thermometer score Ronald Reagan}
    \item{\code{Rockefeller}}{Respondent feeling thermometer score Nelson Rockefeller}
    \item{\code{Johnson}}{Respondent feeling thermometer score Lyndon Johnson}
    \item{\code{Romney}}{Respondent feeling thermometer score George Romney}
    \item{\code{Kennedy}}{Respondent feeling thermometer score Robert Kennedy}
    \item{\code{Muskie}}{Respondent feeling thermometer score Edmund Muskie}
    \item{\code{Agnew}}{Respondent feeling thermometer score Spiro Agnew}
    \item{\code{LeMay}}{Respondent feeling thermometer score Curtis LeMay}
    \item{\code{vote.turnout}}{Respondent reported vote turnout 0=no, 1=yes}
    \item{\code{presidential.vote}}{Respondent reported presidential vote}
  }
}
\source{
\url{http://www.electionstudies.org/studypages/1968prepost/1968prepost.htm}
}
\references{
The Political Behavior Program of the Survey Research Center, Institute for Social Research, University of Michigan. NATIONAL ELECTION STUDIES, 1968: PRE-/POST-ELECTION STUDY [dataset]. Ann Arbor, MI: University of Michigan, Center for Political Studies [producer and distributor], 1999.
}
\examples{
data(ANES1968)
}
\keyword{datasets}
