% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ANES2004_OOC}
\alias{ANES2004_OOC}
\title{2004 American National Election Study (ANES) Data}
\format{
A data frame with rows representing respondents and columns representing their policy preferences on various issues.
}
\source{
The data was collected as part of the 2004 American National Election Study (ANES).
}
\usage{
data(ANES2004_OOC)
}
\description{
The \code{ANES2004_OOC} dataset contains data from the 2004 American National Election Study (ANES). The 2004 ANES asked respondents about their policy preferences on issues ranging from diplomacy and defense spending to government spending and abortion.
}
\details{
This dataset is part of the 2004 American National Election Study (ANES), which surveyed respondents on a wide range of political and social issues. The data includes responses related to diplomacy, defense spending, government spending, and abortion, among others.
}
\examples{
\dontrun{
data(ANES2004_OOC)
}

}
\keyword{datasets}
