% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bam.france}
\alias{bam.france}
\title{BAM Analysis of French Political Data}
\format{
A list of class \code{BAM} containing the following components:
\describe{
\item{polarity}{The polarity of the analysis, set to 2 in this case. This indicates the polarity constraint applied during the scaling.}
\item{n.adapt}{The number of iterations used for adaptation, which was 2500 in this case.}
\item{n.sample}{The number of MCMC samples collected, which was 5000 in this case.}
\item{zhat}{A logical value indicating whether the ideal points should be adjusted for mean-zero scaling (\code{zhat=TRUE}).}
\item{ab}{A logical value indicating whether the response ideal points are used (\code{ab=TRUE}).}
\item{resp.idealpts}{A logical value indicating whether to estimate respondent ideal points (\code{resp.idealpts=TRUE}).}
\item{data}{The original data used in the analysis.}
\item{idealpoints}{The estimated ideal points from the BAM analysis.}
\item{posteriors}{Posterior distributions of the ideal points and other parameters.}
\item{convergence}{Convergence diagnostics for the MCMC chains.}
\item{other_components}{Additional components that store the results and diagnostics of the BAM analysis.}
}
}
\source{
The BAM model was applied to a dataset \code{bamdata} with specific settings to generate \code{bam.france}.
}
\description{
The object \code{bam.france} is a result of applying the \code{BAM()} function to a dataset (\code{bamdata}) with specific parameters
to analyze French political data. This object contains the Bayesian Aldrich-McKelvey scaling results.
}
\details{
The \code{bam.france} object was created using the \code{BAM()} function with the following parameters:
\code{bam.france <- BAM(bamdata, polarity=2, n.adapt=2500, n.sample=5000, zhat=TRUE, ab=TRUE, resp.idealpts=TRUE)}

This analysis uses a Bayesian Aldrich-McKelvey scaling model to estimate ideal points for French political data,
capturing the political preferences and scaling them accordingly.
}
\seealso{
\code{\link[BAM]{BAM}} for more details on the BAM function and its parameters.
}
\keyword{datasets}
