% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bamdata}
\alias{bamdata}
\title{BAM Data from France EES 2009}
\format{
A list of class \code{bamPrep} with two components:
\describe{
\item{stims}{A matrix of stimuli placements (excluding self-placement) with missing values handled.}
\item{self}{A vector of self-placement values.}
}
}
\source{
France 2009 European Election Study (EES)
}
\description{
This dataset, \code{bamdata}, was prepared using the \code{bamPrep} function on the France 2009 European Election Study (EES) data. It is used for Bayesian Aldrich-McKelvey (BAM) scaling.
}
\details{
The \code{bamdata} object was created by applying the \code{bamPrep} function to the \code{franceEES2009} dataset with specified missing values (\code{c(77, 88, 89)}), focusing on self-placement (\code{self = 1}), and requiring a minimum of 5 non-missing values per respondent.
}
