\name{binary.comparisons}
\alias{binary.comparisons}
\title{Calculates a matrix of binary comparisons of stimuli}
\description{
Calculates a matrix of binary comparisons of stimuli
}
\usage{
binary.comparisons(obj)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{obj}{An observations by stimuli matrix of data}
}
\value{
An observations-by-# comparisons matrix of comparisons. 
}
