\name{candidatetherms2008}
\alias{candidatetherms2008}
\docType{data}
\title{2008 American National Election Study 
}
\description{
Feeling thermometers for presidential candidates from the 2008 American National Election Study
}
\usage{data(candidatetherms2008)}
\format{
	A 2322x9 matrix of feeling thermometers for McCain, Bush, Obama, Biden, Palin, H. Clinton, B. Clinton, Dem Party and Rep Party, respectively
}
\source{
\url{https://electionstudies.org/data-center/2008-time-series-study/}
}
\references{
The American National Election Studies (ANES; www.electionstudies.org). The ANES 2008 Time Series Study [dataset]. Stanford University and the University of Michigan [producers].
}
\examples{
data(candidatetherms2008)
}
\keyword{datasets}
