% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{candidatetherms2008}
\alias{candidatetherms2008}
\title{Candidate Favorability Ratings from the 2008 ANES}
\format{
A matrix with rows representing respondents and columns representing the following nine political stimuli:
\describe{
\item{mccain}{Favorability rating for Sen. John McCain.}
\item{bush}{Favorability rating for Pres. George W. Bush.}
\item{obama}{Favorability rating for Sen. Barack Obama.}
\item{biden}{Favorability rating for Sen. Joe Biden.}
\item{palin}{Favorability rating for Gov. Sarah Palin.}
\item{hclinton}{Favorability rating for Sen. Hillary Clinton.}
\item{bclinton}{Favorability rating for former Pres. Bill Clinton.}
\item{demparty}{Favorability rating for the Democratic Party.}
\item{repparty}{Favorability rating for the Republican Party.}
}
}
\source{
Data from the 2008 American National Election Study (ANES).
}
\usage{
data(candidatetherms2008)
}
\description{
The \code{candidatetherms2008} dataset contains favorability ratings from the 2008 American National Election Study (ANES).
Respondents were asked to rate their favorability towards nine political figures and parties on a 0-100 scale.
}
\details{
The 2008 ANES asked respondents to rate their favorability towards nine political figures and parties on a scale from 0 to 100,
where 0 represents the least favorable and 100 represents the most favorable. Missing values are coded as \code{NA}.
}
\examples{
\dontrun{
data(candidatetherms2008)
}

}
\keyword{datasets}
