% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CDS2000}
\alias{CDS2000}
\title{2000 Convention Delegate Study (CDS)}
\format{
A data frame with the following 14 variables:
\describe{
\item{Party}{The party affiliation of the delegate (Democratic or Republican).}
\item{Preferred Presidential Nominee}{The delegate's preferred presidential nominee.}
\item{Race}{The race of the delegate.}
\item{Religious Tradition}{The religious tradition of the delegate.}
\item{Lib-Con}{Liberal-Conservative self-placement scale.}
\item{Abortion}{Policy preference on abortion.}
\item{Govt Services}{Policy preference on government services.}
\item{Defense Spending}{Policy preference on defense spending.}
\item{Aid to Blacks}{Policy preference on aid to Black Americans.}
\item{Health Insurance}{Policy preference on health insurance.}
\item{Protect Homosexuals}{Policy preference on protecting homosexual rights.}
\item{Affirmative Action}{Policy preference on affirmative action.}
\item{Surplus for Tax Cuts}{Policy preference on using budget surplus for tax cuts.}
\item{Free Trade}{Policy preference on free trade.}
}
}
\source{
Data collected from the 2000 Convention Delegate Study (CDS). Studies that have analyzed CDS data include:
Stone, W. J., & Abramowitz, A. I. (1983), Layman, G. C. (2001) and Layman, G. C., et al. (2010).
}
\usage{
data(CDS2000)
}
\description{
The \code{CDS2000} dataset contains data from the 2000 Convention Delegate Study (CDS), which interviewed delegates to the
Republican and Democratic National Conventions. The survey included a battery of issue scales on which delegates were
asked to place their policy preferences and those of major political figures (e.g., Al Gore and George W. Bush).
}
\details{
In 2000, the CDS received completed questionnaires from 1,907 delegates to the Democratic National Convention and 985
delegates to the Republican National Convention. This dataset has been used in various studies analyzing delegate behavior
and policy preferences, including works by Stone and Abramowitz (1983), Layman (2001), and Layman et al. (2010).
}
\examples{
\dontrun{
data(CDS2000)
}
}
\keyword{datasets}
