\name{ches_eu}
\alias{ches_eu}
\alias{sub.europe}
\docType{data}
\title{Chapel Hill Expert Study data}
\description{
A subset of data from the 2010 Chapel Hill Expert Survey
}
\usage{data(ches_eu)}
\format{
A dataset with 118 variables and 224 rows.  The rows relate to experts who place some of the 118 parties represented in the dataset.  See codebook at the link below for a full listing of all parties.
}
\source{
\url{https://www.chesdata.eu/our-surveys/}
}
\references{
Ryan Bakker, Catherine de Vries, Erica Edwards, Liesbet Hooghe, Seth Jolly, Gary Marks, Jonathan Polk, Jan Rovny, Marco Steenbergen, Milada Vachudova (2015), "Measuring Party Positions in Europe: The Chapel Hill Expert Survey Trend File, 1999-2010," Party Politics, 21.1: 143-152.
}
\keyword{datasets}
