% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{ches_eu}
\alias{ches_eu}
\title{CHES EU Dataset: Party Positions from Chapel Hill Expert Survey}
\format{
A data frame with 118 variables:
\describe{
\item{1001}{Variable 1001}
\item{1002}{Variable 1002}
\item{1003}{Variable 1003}
\item{1004}{Variable 1004}
\item{1005}{Variable 1005}
\item{1014}{Variable 1014}
\item{1016}{Variable 1016}
\item{1017}{Variable 1017}
\item{1018}{Variable 1018}
\item{1019}{Variable 1019}
\item{102}{Variable 102}
\item{103}{Variable 103}
\item{104}{Variable 104}
\item{105}{Variable 105}
\item{106}{Variable 106}
\item{107}{Variable 107}
\item{108}{Variable 108}
\item{109}{Variable 109}
\item{110}{Variable 110}
\item{1101}{Variable 1101}
\item{1102}{Variable 1102}
\item{1104}{Variable 1104}
\item{1105}{Variable 1105}
\item{1106}{Variable 1106}
\item{1107}{Variable 1107}
\item{1108}{Variable 1108}
\item{1109}{Variable 1109}
\item{112}{Variable 112}
\item{113}{Variable 113}
\item{114}{Variable 114}
\item{115}{Variable 115}
\item{116}{Variable 116}
\item{1301}{Variable 1301}
\item{1302}{Variable 1302}
\item{1303}{Variable 1303}
\item{1304}{Variable 1304}
\item{1307}{Variable 1307}
\item{1401}{Variable 1401}
\item{1402}{Variable 1402}
\item{1403}{Variable 1403}
\item{1404}{Variable 1404}
\item{1405}{Variable 1405}
\item{1406}{Variable 1406}
\item{1408}{Variable 1408}
\item{1409}{Variable 1409}
\item{1601}{Variable 1601}
\item{1602}{Variable 1602}
\item{1603}{Variable 1603}
\item{1604}{Variable 1604}
\item{1605}{Variable 1605}
\item{1606}{Variable 1606}
\item{1607}{Variable 1607}
\item{1609}{Variable 1609}
\item{1610}{Variable 1610}
\item{2001}{Variable 2001}
\item{2002}{Variable 2002}
\item{2003}{Variable 2003}
\item{2004}{Variable 2004}
\item{2007}{Variable 2007}
\item{2008}{Variable 2008}
\item{2009}{Variable 2009}
\item{201}{Variable 201}
\item{2010}{Variable 2010}
\item{202}{Variable 202}
\item{203}{Variable 203}
\item{206}{Variable 206}
\item{2101}{Variable 2101}
\item{2102}{Variable 2102}
\item{2103}{Variable 2103}
\item{2104}{Variable 2104}
\item{2107}{Variable 2107}
\item{2109}{Variable 2109}
\item{211}{Variable 211}
\item{2110}{Variable 2110}
\item{213}{Variable 213}
\item{215}{Variable 215}
\item{218}{Variable 218}
\item{2201}{Variable 2201}
\item{2202}{Variable 2202}
\item{2203}{Variable 2203}
\item{2204}{Variable 2204}
\item{2206}{Variable 2206}
\item{2207}{Variable 2207}
\item{2301}{Variable 2301}
\item{2302}{Variable 2302}
\item{2303}{Variable 2303}
\item{2304}{Variable 2304}
\item{2307}{Variable 2307}
\item{2308}{Variable 2308}
\item{2309}{Variable 2309}
\item{2601}{Variable 2601}
\item{2603}{Variable 2603}
\item{2604}{Variable 2604}
\item{2605}{Variable 2605}
\item{2606}{Variable 2606}
\item{2607}{Variable 2607}
\item{2611}{Variable 2611}
\item{2612}{Variable 2612}
\item{301}{Variable 301}
\item{302}{Variable 302}
\item{303}{Variable 303}
\item{304}{Variable 304}
\item{306}{Variable 306}
\item{308}{Variable 308}
\item{501}{Variable 501}
\item{502}{Variable 502}
\item{504}{Variable 504}
\item{505}{Variable 505}
\item{506}{Variable 506}
\item{507}{Variable 507}
\item{511}{Variable 511}
\item{513}{Variable 513}
\item{517}{Variable 517}
\item{520}{Variable 520}
\item{523}{Variable 523}
\item{a}{Variable a}
\item{b}{Variable b}
\item{c}{Variable c}
}
}
\source{
Bakker, R., De Vries, C., Edwards, E., Hooghe, L., Jolly, S., Marks, G.,
Polk, J., Rovny, J., Steenbergen, M., & Vachudova, M. A. (2015).
Measuring party positions in Europe: The Chapel Hill expert survey trend
file, 1999-2010. \emph{Party Politics}, 21(1), 143-152.
\doi{10.1177/1354068812462931}

CHES data: \url{https://www.chesdata.eu/}
}
\usage{
data(ches_eu)
}
\description{
Party means and standard deviations from the 2010 Chapel Hill Expert Survey
(CHES). Contains expert placements of 118 political parties from 14 EU
member countries, including three anchoring vignette parties for scale
calibration.
}
\details{
The 2010 Chapel Hill Expert Survey (CHES) included 224 experts evaluating
parties from 14 EU member countries. Each actual party was evaluated by
8-17 experts, while all three vignette parties were evaluated by over 160
experts. Vignette parties are hypothetical parties used as anchoring points
to calibrate expert placements and adjust for differential item functioning
(DIF).

This dataset is used to demonstrate anchoring vignette methods in spatial
analysis, as discussed in Section 2.5 (pages 58-60) of the accompanying
textbook. The vignettes allow researchers to adjust for systematic
differences in how experts use the rating scales.
}
\examples{
\dontrun{
data(ches_eu)

# Dataset structure
str(ches_eu)
dim(ches_eu)  # 121 parties (118 actual + 3 vignettes)

# Separate vignettes from actual parties
vignettes <- ches_eu[ches_eu$vignette == 1, ]
actual_parties <- ches_eu[ches_eu$vignette == 0, ]

# Summary statistics
summary(ches_eu$mean)
summary(ches_eu$sd)

# Parties by country
table(actual_parties$country)

# Compare uncertainty: actual parties vs. vignettes
boxplot(sd ~ vignette, data = ches_eu,
        names = c("Actual Parties", "Vignettes"),
        ylab = "Standard Deviation",
        main = "Expert Agreement: Actual vs. Vignette Parties")

# Plot party positions
plot(ches_eu$mean, ches_eu$sd,
     col = ifelse(ches_eu$vignette == 1, "red", "blue"),
     pch = ifelse(ches_eu$vignette == 1, 17, 16),
     xlab = "Mean Position", ylab = "Standard Deviation",
     main = "Party Positions from CHES 2010")
legend("topright", legend = c("Actual", "Vignette"),
       col = c("blue", "red"), pch = c(16, 17))
}

}
\references{
Bakker, R., et al. (2015). Measuring party positions in Europe: The Chapel
Hill expert survey trend file, 1999-2010. \emph{Party Politics}, 21(1),
143-152.

King, G., & Wand, J. (2007). Comparing Incomparable Survey Responses:
Evaluating and Selecting Anchoring Vignettes. \emph{Political Analysis},
15(1), 46-66.
}
\keyword{datasets}
