\name{doubleCenter}
\alias{doubleCenter}
\title{Calculates a double-centered matrix}
\description{
Calculates a double-centered matrix
}
\usage{
doubleCenter(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{A \code{matrix} to be double centered.}
}
\value{
A double-centered matrix
}
