\name{france4}
\alias{france4}
\docType{data}
\title{Fourth French Republic Roll Call Data
}
\description{
Roll Call data from the fourth French republic (1946-1958).
}
\usage{data(france4)}
\format{
  A data frame with 1566 observations on the following 894 variables.  See source for full coding details. 
}
\source{
\url{http://www9.georgetown.edu/faculty/ev42/france.htm}
}
\references{
Howard Rosenthal and Erik Voeten. 2004. \sQuote{Analyzing Roll Calls with Perfect Spatial Voting: France 1946-1958.} American Journal of Political Science 48(3): 620-32.

Duncan MacRae Jr. 1967. Parliament, Parties, and Society in France 1946-1958. New York: St. Martin's Press.
}
\examples{
data(france4)
}
\keyword{datasets}
