% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{france4}
\alias{france4}
\title{Roll Call Data from the French Fourth Republic}
\format{
A data frame with the following variables:
\describe{
\item{NAME}{Name of the deputy (legislator).}
\item{MID}{Deputy ID, which remains constant even if the deputy switches party.}
\item{CASEID}{Deputy ID, which is unique for each party affiliation if the deputy switches party.}
\item{PAR}{Party affiliation of the deputy.}
\item{PARSEQ}{Sequence of party affiliation for deputies who switched parties.}
\item{VOTE1}{Result of the first roll call vote (and so on for subsequent votes).}
\item{...}{Other variables representing additional roll call votes.}
}
}
\source{
Data from Rosenthal and Voeten (2004). The original dataset and documentation can be found online at
\url{http://www9.georgetown.edu/faculty/ev42/france.htm}.
}
\usage{
data(france4)
}
\description{
The \code{france4} dataset contains roll call data from the French Fourth Republic, as analyzed by Rosenthal and Voeten (2004).
This dataset was used to estimate a party-switcher model where a separate ideal point is estimated each time a legislator
changes party affiliation. The dataset includes legislator-specific variables and roll call votes. Rosenthal and Voeten (2004)
found that the latent ideological space remained stable over the course of the French Fourth Republic, so the roll call data
is not segmented by legislative session.
}
\details{
The dataset includes 1,416 separate ideal points, as each party-switching deputy has a separate entry for each party affiliation.
The first five columns of the dataset are legislator-specific variables: NAME (deputy name), MID (constant deputy ID), CASEID
(unique ID for each party affiliation), PAR (party affiliation), and PARSEQ (sequence number of party affiliations for party-switching deputies).
The remaining columns represent the roll call votes.

The dataset and its extensive documentation were originally made available online by Rosenthal and Voeten (2004).
For more information and access to the original data, visit \url{http://www9.georgetown.edu/faculty/ev42/france.htm}.
}
\examples{
\dontrun{
data(france4)
}

}
\keyword{datasets}
