\name{franceEES2009}
\alias{franceEES2009}
\alias{french.parties.individuals}
\docType{data}
\title{2009 European Election Survey (Voters, France)}
\description{
A subset of the 2009 European Election Survey for French voters. 
}
\usage{data(franceEES2009)}
\format{
A 1000x9 matrix of respondent placements of themselves and 8 political parties (Extreme Left, Communist, Socialist, Green, UDF, UMP, National Front and Left, respectively) on the left-right scale (0=extreme left, 10=extreme right, missing values=77,88,89).  
}
\source{
\url{https://dbk.gesis.org/dbksearch/sdesc2.asp?no=5055&db=e&doi=10.4232/1.11760}
}
\references{
Egmond, Marcel van; Brug, Wouter van der; Hobolt, Sara; Franklin, Mark; Sapir, Eliyahu V. (2013): European Parliament Election Study 2009, Voter Study. GESIS Data Archive, Cologne. ZA5055 Data file Version 1.1.0, doi:10.4232/1.11760}
\examples{
data(franceEES2009)
}
\keyword{datasets}
