% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{franceEES2009}
\alias{franceEES2009}
\title{French Module of the 2009 European Election Study (EES)}
\format{
A data frame with 1,000 rows and 9 columns:
\describe{
\item{self}{Numeric, self-placement on the left-right scale.}
\item{Extreme Left}{Numeric, placement of the "Extreme Left" party on the left-right scale.}
\item{Communist}{Numeric, placement of the "Communist" party on the left-right scale.}
\item{Socialist}{Numeric, placement of the "Socialist" party on the left-right scale.}
\item{Greens}{Numeric, placement of the "Greens" party on the left-right scale.}
\item{UDF (Bayrou)}{Numeric, placement of the "UDF (Bayrou)" party on the left-right scale.}
\item{UMP (Sarkozy)}{Numeric, placement of the "UMP (Sarkozy)" party on the left-right scale.}
\item{National Front}{Numeric, placement of the "National Front" party on the left-right scale.}
\item{Left Party}{Numeric, placement of the "Left Party" on the left-right scale.}
}
}
\source{
European Election Study (EES) 2009, French module.
}
\usage{
data(franceEES2009)
}
\description{
The \code{franceEES2009} dataset contains data from the French module of the 2009 European Election Study (EES). The EES surveyed 1,000 French citizens, asking them to place themselves and eight major political parties on a 0-10 left-right scale (0 representing the most left-wing position, 10 representing the most right-wing position).
}
\details{
Responses are coded as 77, 88, or 89 if the respondent refuses to answer or does not know the party or where to place it. The dataset is useful for analyzing the political landscape in France during 2009, especially in terms of how citizens and parties are positioned on the left-right spectrum.
}
\examples{
\dontrun{
data(franceEES2009)
}

}
\keyword{datasets}
