% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{french.parties.individuals}
\alias{french.parties.individuals}
\title{French Party Placement Data from the 2009 European Election Study (EES)}
\format{
A matrix with 1,000 rows (representing respondents) and 8 columns (representing political parties):
\describe{
\item{extremeleft}{Placement of the "Extreme Left" party.}
\item{communist}{Placement of the Communist Party.}
\item{socialist}{Placement of the Socialist Party.}
\item{greens}{Placement of the Greens Party.}
\item{udfbayrou}{Placement of the Union for French Democracy (Bayrou's party).}
\item{umpsarkozy}{Placement of the Union for a Popular Movement (Sarkozy's party).}
\item{nationalfront}{Placement of the National Front party.}
\item{leftparty}{Placement of the Left Party.}
}
}
\source{
Data from the French module of the 2009 European Election Study (EES).
}
\usage{
data(french.parties.individuals)
}
\description{
The \code{french.parties.individuals} dataset contains party placement data from the French module of the 2009 European Election Study (EES).
This dataset was used in Chapter 2 to directly scale respondents’ placements of eight major political parties on a ten-point left-right ideological scale.
}
\details{
The French module of the 2009 EES asked 1,000 respondents to place eight major political parties on a ten-point left-right ideological scale.
This dataset, \code{french.parties.individuals}, contains those placements. In Chapter 2, these placements were scaled directly to compare the results
obtained from scaling different forms of data from the same set of observations.

In addition to directly scaling respondents' placements, each set of placements can be arranged into a similarities matrix for further analysis.
}
\examples{
\dontrun{
data(french.parties.individuals)
}

}
\keyword{datasets}
