\name{gray.palette}
\alias{gray.palette}
\title{Gray color rampo}
\description{Produces a color ramp of grays.}
\usage{
gray.palette(n, lower = 0.3, upper = 0.7)
}
\arguments{
  \item{n}{Number of colors desired.}
  \item{lower}{Smallest color value.}
  \item{upper}{Largest color value.}
}
\value{
A \code{n}-length vector of gray colors.
}
