\name{interest1981}
\alias{interest1981}
\docType{data}
\title{
1981 Interest Group Rating Scores
}
\description{
Subset of Keith Poole's Interest Group data (1959-1981) 
}
\usage{data(interest1981)}
\format{
  A data frame with 100 observations on the following 49 variables.
  \describe{
    \item{\code{year}}{Year of observation (only 1981)}
    \item{\code{congress}}{Congress number (only 97)}
    \item{\code{session}}{Session number (only 1)}
    \item{\code{senate}}{Senate number (only 1)}
    \item{\code{prid}}{ICPSR ID Corrected by Poole and Rosenthal}
    \item{\code{party}}{Party Code (1=Democrat, 2=Republican)}
    \item{\code{state}}{ICPSR State Code}
    \item{\code{name}}{Legislator Name}
    \item{\code{ada}}{Americans for Democrat Action Rating Score}
    \item{\code{cope}}{Committee on Political Action (AFL-CIO) Rating Score}
    \item{\code{nfu}}{National Farmer's Union Rating Score}
    \item{\code{ccus}}{Chamber of Commerce of the United States Rating Score}
    \item{\code{aca}}{Americans for Constitutional Action Rating Score}
    \item{\code{cc}}{Conservative Coalition (CQ) Rating Score}
    \item{\code{ll}}{Liberty Lobby Rating Score}
    \item{\code{asc}}{American Security Council Rating Score}
    \item{\code{ripon}}{Ripon Society Rating Score}
    \item{\code{uaw}}{United Auto Workers Rating Score}
    \item{\code{nea}}{National Education Association Rating Score}
    \item{\code{lwv}}{League of Women Voters Rating Score}
    \item{\code{ntu}}{National Taxpayers' Union Rating Score}
    \item{\code{lcv}}{League of Conservation VotersRating Score}
    \item{\code{cfa}}{Consumer Federation of America Rating Score}
    \item{\code{nfo}}{National Farmers Organization Rating Score}
    \item{\code{aft}}{American Federation of Teachers Rating Score}
    \item{\code{nfib}}{National Federation of Independent Busioness Rating Score}
    \item{\code{cfnfmp}}{Coalition for a New Foreign and Military Policy Rating Score}
    \item{\code{nasc}}{National Alliance of Senior Citizens Rating Score}
    \item{\code{afbf}}{American Farm Bureau Federation Rating Score}
    \item{\code{ccause}}{Common Cause Rating Score}
    \item{\code{bfw}}{Bread for the World Rating Score}
    \item{\code{umw}}{United Mine Workers Rating Score}
    \item{\code{bctd}}{Building and Construction Trade Department (AFL-CIO) Rating Score}
    \item{\code{cv}}{Christian Voice Rating Score}
    \item{\code{aclu}}{American Civil Liberties Union Rating Score}
    \item{\code{csfca}}{Committee for the Survival of a Free Congress: All Issues Rating Score}
    \item{\code{cvvf}}{Christian Voters Victory Fund Rating Score}
    \item{\code{reagan}}{Reagan Support Score (CQ) Rating Score}
    \item{\code{stateid}}{Numeric State ID}
    \item{\code{district}}{Numeric vector of zeros}
    \item{\code{surname}}{Legislator Surname and first initial}
    \item{\code{dw1}}{First dimension DW Nominate Score}
    \item{\code{dw2}}{Second Dimension DW Nominate Score}
    \item{\code{se1}}{SE of DW Nominate First Dimension}
    \item{\code{se2}}{SE of DW Nominate Second Dimension}
    \item{\code{logl}}{logl }
    \item{\code{choices}}{choices }
    \item{\code{errors}}{errors }
    \item{\code{gmp}}{gmp }
  }
}
\source{
\url{http://www.voteview.com/dwnl_5.htm}
}
\references{
Include Reference
}
\keyword{datasets}
