% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interest1981}
\alias{interest1981}
\title{Interest Group Ratings of Members of Congress (1959-1981)}
\source{
The data were compiled by Keith Poole and have been analyzed in several studies:
Poole, K. T. (1981, 1984, 1990) and Poole, K. T., & Daniels, P. (1985).
}
\usage{
data(interest1981)
}
\description{
The \code{interest1981} dataset is a subset of a much larger dataset compiled by Keith Poole, containing nearly 200,000
interest group ratings of members of Congress between 1959 and 1981. These data have been extensively analyzed
by Poole (1981, 1984, 1990) and Poole and Daniels (1985). The dataset has been used to perform the MLSMU6
unfolding procedure in two dimensions, following the methodologies established in the mentioned studies.
}
\details{
This dataset contains interest group ratings that were used to perform the MLSMU6 unfolding procedure in two dimensions,
as described in the works of Poole (1981, 1984, 1990) and Poole and Daniels (1985).
}
\examples{
data(interest1981)

}
\keyword{datasets}
