\name{mexicoCSES2000}
\alias{mexicoCSES2000}
\docType{data}
\title{
Respondent Party Placements from the 2000 Mexico CSES dataset
}
\description{
A numeric matrix giving respondent placements on the left-right scale for PAN, PRI, PRD, PT, Greens and PARM, respectively.}
\usage{data(mexicoCSES2000)}
\format{
A numeric matrix with the following columns:
\describe{
   \item{\code{PAN}}{Respondent Left-Right rating of Partido Accion Nacional (National Action Party)}
   \item{\code{PRI}}{Respondent Left-Right rating of Partido Revolucionario Institucional (Institutional Revolutionary Party)}
   \item{\code{PRD}}{Respondent Left-Right rating of Partido de la Revolucion Democratica (Party of the Democratic Revolution)}
   \item{\code{PT}}{Respondent Left-Right rating of Partido del Trabajo (Labor Party)}
   \item{\code{Greens}}{Respondent Left-Right rating of Partido Verde Ecologista de Mexico (Ecologist Green Party of Mexico)}
   \item{\code{PRAM}}{Respondent Left-Right rating of Partido AUtentico de la REvolucion Mexicana (Authentic Party of the Mexican Revolution)}
   }
}
\source{
The Comparative Study of Electoral Systems (\url{www.cses.org}). CSES MODULE 1 FULL RELEASE [dataset]. Ann Arbor, MI: University of Michigan, Center for Political Studies [producer and distributor]. August 4, 2003 version. doi:10.7804/cses.module1.2003-08-04
}
\keyword{datasets}
