\name{mexicoCSES2006}
\alias{mexicoCSES2006}
\docType{data}
\title{
Respondent Party Placements from the 2006 Mexico CSES dataset
}
\description{
A numeric matrix giving respondent placements on the left-right scale for PAN, PRD, PRI, Greens, PT, Convergencia, Nueva Alianza and PSD, respectively.}
\usage{data(mexicoCSES2000)}
\format{
A numeric matrix with the following columns:
\describe{
   \item{\code{PAN}}{Respondent Left-Right rating of Partido Accion Nacional (National Action Party)}
   \item{\code{PRD}}{Respondent Left-Right rating of Partido de la Revolucion Democratica (Party of the Democratic Revolution)}
   \item{\code{PRI}}{Respondent Left-Right rating of Partido Revolucionario Institucional (Institutional Revolutionary Party)}
   \item{\code{Greens}}{Respondent Left-Right rating of Partido Verde Ecologista de Mexico (Ecologist Green Party of Mexico)}
   \item{\code{PT}}{Respondent Left-Right rating of Partido del Trabajo (Labor Party)}
   \item{\code{Convergencia}}{Respondent Left-Right rating Convergencia por la Democracia (Convergence for Democracy)}
   \item{\code{Nueva Allianza}}{Respondent Left-Right rating Nueva Allianza (New Alliance)}
   \item{\code{PSD}}{Respondent Left-Right rating Partido Socialdemocrata (Social Democrats)}
   }
}
\source{
The Comparative Study of Electoral Systems (\url{www.cses.org}). CSES MODULE 3 FULL RELEASE [dataset]. March 27, 2013 version. doi:10.7804/cses.module3.2013-03-27}
\keyword{datasets}
