% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mlsmu6_out}
\alias{mlsmu6_out}
\title{MDS Solution from mlsmu6}
\format{
A matrix or data frame, depending on the structure of the output from \code{mlsmu6}, typically containing coordinates in the reduced dimensional space.
}
\source{
Generated using the \code{mlsmu6} function on the \code{interest1981} dataset.
}
\description{
The \code{mlsmu6_out} dataset contains the multidimensional scaling (MDS) solution generated by applying the \code{mlsmu6} function to a subset of the \code{interest1981} dataset. The MDS solution is computed using two dimensions and a cutoff value of 5, with data grouped by political party labels.
}
\details{
The \code{mlsmu6_out} object is generated by running the following command:
\preformatted{
mlsmu6_out <- mlsmu6(input = interest1981[, 9:38], ndim = 2, cutoff = 5,
                     id = factor(interest1981$party, labels = c("D", "R")))
}
The \code{mlsmu6} function applies multidimensional scaling to reduce the dimensionality of the input data, with the number of dimensions set to 2 and a cutoff value of 5. The input data is a subset of the \code{interest1981} dataset, specifically columns 9 to 38, grouped by the political party (\code{party}) variable.
}
\examples{
\dontrun{
data(mlsmu6_out)
}
}
\seealso{
\code{\link{mlsmu6}}, \code{\link{interest1981}}
}
\keyword{datasets}
