\name{nations}
\alias{nations}
\docType{data}
\title{
Similarities of Nations (1968)
}
\description{
Average similarity ratings of 12 countries from 18 students in Myron Wish's Psychological Measurement class in 1968. 
}
\usage{data(nations)}
\format{
A numeric matrix where each entry is the average similarity (on a 1=very different to 9=very similar scale).  The columns and rows refer to the following countries (in this order): Brazil, Congo, Cuba, Egypt, France, India, Israel, Japan, China, USSR, USA, Yugoslavia
}
\source{
Wish, Myron. 1971. \sQuote{Individual Differences in Perceptions and Preferences Among Nations}.  In Attitude Research Reaches New Heights, ed. Charls W. King and Douglas J. Tigert.  Chicago: American Marketing Association, pp. 312-328. 
}
\keyword{datasets}
