\name{nes1968_urbanunrest}
\alias{nes1968_urbanunrest}
\docType{data}
\title{
Issue scale data from the 1968 American National Election Study regarding Urban Unrest
}
\description{
Respondent placements of themselves and political figures on a 1-7 scale of urban unrest.  Endpoints of the scale are \sQuote{Solve Problems of Poverty and Unemployment} and \sQuote{Use All Available Force}.  See \sQuote{Details} for full question wording. 
}
\usage{data(nes1968_urbanunrest)}
\format{
A numeric matrix of respondent placements with the following column names: 
\describe{
   \item{\code{vote.choice}}{}
   \item{\code{Johson}}{Respondent placement of Lyndon Johnson (President)}
   \item{\code{Humphrey}}{Respondent placement of Hubert Humphrey (D Presidential Candidate)}
   \item{\code{Nixon}}{Respondent placement of Richard Nixon (R Presidential Candidate)}
   \item{\code{Wallace}}{Respondent placement of George Wallace (I Presidential Candidate)}
   \item{\code{self}}{Respondent placement of Self}
}
}
\details{
The question asked in the ANES was as follows: 

There is much discussion about the best way to deal with the problem of urban unrest and rioting. Some say it is more important to use all available force to maintain law and order -- no matter what results. Others say it is more important to correct the problems of poverty and unemployment that give rise to the disturbances. And, of course, other people have opinions in between. Suppose the people who stress the use of force are at one end of this scale -- at point number 7. And suppose the people who stress doing more about the problems of poverty and unemployment are at the other end -- at point number 1. Where would you place ... on this scale?
}
\source{
The Political Behavior Program of the Survey Research Center, Institute for Social Research, University of Michigan. NATIONAL ELECTION STUDIES, 1968: PRE-/POST-ELECTION STUDY [dataset]. Ann Arbor, MI: University of Michigan, Center for Political Studies [producer and distributor], 1999.
}
\keyword{datasets}
