\name{plot.aldmck_ci}
\alias{plot.aldmck_ci}
\title{Plotting method for }
\description{Produces a plot of the stimulus locations with confidence/credible intervals.}

\usage{
\method{plot}{aldmck_ci}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{An object of class \code{aldmck_ci} produce either by \code{\link{boot.aldmck}} or \code{\link{BAM}}.}
  \item{\dots}{Other arguments to be passed down, currently not implemented.}
}
\value{
A \code{ggplot}.
}
