\name{presvote2008}
\alias{presvote2008}
\docType{data}
\title{
Presidential Vote in 2008
}
\description{
Presidential Vote in 2008 from the American National Election Study.
}
\usage{data(presvote2008)}
\format{
A numeric vector providing respondent vote for president in 2008 (0=McCain [R], 1=Obama [D])
}
\source{
The American National Election Studies (ANES; \url{www.electionstudies.org}). The ANES 2008 Time Series Study [dataset]. Stanford University and the University of Michigan [producers].
}
\keyword{datasets}
