% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{presvote2008}
\alias{presvote2008}
\title{2008 U.S. Presidential Vote Data}
\usage{
data(presvote2008)
}
\description{
The \code{presvote2008} dataset contains data on the voting behavior in the 2008 U.S. Presidential election, where a vote
for John McCain is coded as 0 and a vote for Barack Obama is coded as 1.
}
\details{
This dataset provides a binary coding for the voting choice in the 2008 U.S. Presidential election, which can be used for various
analyses, such as logistic regression, to study voting behavior and preferences.
}
\examples{
\dontrun{
data(presvote2008)
}

}
\keyword{datasets}
