\name{print.aldmck_ci}
\alias{print.aldmck_ci}
\title{printing method for Aldrich-McKelvey Scaling REsults}
\description{Produces a printed output of the stimulus locations with confidence/credible intervals.}

\usage{
\method{plot}{aldmck_ci}(x, ..., digits=3)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{An object of class \code{aldmck_ci} produce either by \code{\link{boot.aldmck}} or \code{\link{BAM}}.}
  \item{\dots}{Other arguments to be passed down, currently not implemented.}
  \item{digits}{Number of digits to print}
}
