% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{rcv_ep1}
\alias{rcv_ep1}
\title{Roll Call Voting Data from the First European Parliament (1979-1984)
#' @description Roll call voting data from the first elected European Parliament (1979-1984),
assembled by Hix, Noury, and Roland (2006). Contains voting records of
Members of the European Parliament (MEPs) on 886 roll call votes.}
\format{
A data frame with 410 MEPs (rows) and 891 columns:

A data frame with 548 rows and 891 columns:
\describe{
\item{EPG}{European Parliament Group (character)}
\item{MEPID}{Member of European Parliament ID (numeric)}
\item{MEPNAME}{Member of European Parliament name (character)}
\item{MS}{Member State (character)}
\item{NP}{National Party (character)}
\item{V1}{Roll call vote (numeric)}
\item{V2}{Roll call vote (numeric)}
\item{V3}{Roll call vote (numeric)}
\item{V4}{Roll call vote (numeric)}
\item{V5}{Roll call vote (numeric)}
\item{V6}{Roll call vote (numeric)}
\item{V7}{Roll call vote (numeric)}
\item{V8}{Roll call vote (numeric)}
\item{V9}{Roll call vote (numeric)}
\item{V10}{Roll call vote (numeric)}
\item{V11}{Roll call vote (numeric)}
\item{V12}{Roll call vote (numeric)}
\item{V13}{Roll call vote (numeric)}
\item{V14}{Roll call vote (numeric)}
\item{V15}{Roll call vote (numeric)}
\item{V16}{Roll call vote (numeric)}
\item{V17}{Roll call vote (numeric)}
\item{V18}{Roll call vote (numeric)}
\item{V19}{Roll call vote (numeric)}
\item{V20}{Roll call vote (numeric)}
\item{V21}{Roll call vote (numeric)}
\item{V22}{Roll call vote (numeric)}
\item{V23}{Roll call vote (numeric)}
\item{V24}{Roll call vote (numeric)}
\item{V25}{Roll call vote (numeric)}
\item{V26}{Roll call vote (numeric)}
\item{V27}{Roll call vote (numeric)}
\item{V28}{Roll call vote (numeric)}
\item{V29}{Roll call vote (numeric)}
\item{V30}{Roll call vote (numeric)}
\item{V31}{Roll call vote (numeric)}
\item{V32}{Roll call vote (numeric)}
\item{V33}{Roll call vote (numeric)}
\item{V34}{Roll call vote (numeric)}
\item{V35}{Roll call vote (numeric)}
\item{V36}{Roll call vote (numeric)}
\item{V37}{Roll call vote (numeric)}
\item{V38}{Roll call vote (numeric)}
\item{V39}{Roll call vote (numeric)}
\item{V40}{Roll call vote (numeric)}
\item{V41}{Roll call vote (numeric)}
\item{V42}{Roll call vote (numeric)}
\item{V43}{Roll call vote (numeric)}
\item{V44}{Roll call vote (numeric)}
\item{V45}{Roll call vote (numeric)}
\item{V46}{Roll call vote (numeric)}
\item{V47}{Roll call vote (numeric)}
\item{V48}{Roll call vote (numeric)}
\item{V49}{Roll call vote (numeric)}
\item{V50}{Roll call vote (numeric)}
\item{V51}{Roll call vote (numeric)}
\item{V52}{Roll call vote (numeric)}
\item{V53}{Roll call vote (numeric)}
\item{V54}{Roll call vote (numeric)}
\item{V55}{Roll call vote (numeric)}
\item{V56}{Roll call vote (numeric)}
\item{V57}{Roll call vote (numeric)}
\item{V58}{Roll call vote (numeric)}
\item{V59}{Roll call vote (numeric)}
\item{V60}{Roll call vote (numeric)}
\item{V61}{Roll call vote (numeric)}
\item{V62}{Roll call vote (numeric)}
\item{V63}{Roll call vote (numeric)}
\item{V64}{Roll call vote (numeric)}
\item{V65}{Roll call vote (numeric)}
\item{V66}{Roll call vote (numeric)}
\item{V67}{Roll call vote (numeric)}
\item{V68}{Roll call vote (numeric)}
\item{V69}{Roll call vote (numeric)}
\item{V70}{Roll call vote (numeric)}
\item{V71}{Roll call vote (numeric)}
\item{V72}{Roll call vote (numeric)}
\item{V73}{Roll call vote (numeric)}
\item{V74}{Roll call vote (numeric)}
\item{V75}{Roll call vote (numeric)}
\item{V76}{Roll call vote (numeric)}
\item{V77}{Roll call vote (numeric)}
\item{V78}{Roll call vote (numeric)}
\item{V79}{Roll call vote (numeric)}
\item{V80}{Roll call vote (numeric)}
\item{V81}{Roll call vote (numeric)}
\item{V82}{Roll call vote (numeric)}
\item{V83}{Roll call vote (numeric)}
\item{V84}{Roll call vote (numeric)}
\item{V85}{Roll call vote (numeric)}
\item{V86}{Roll call vote (numeric)}
\item{V87}{Roll call vote (numeric)}
\item{V88}{Roll call vote (numeric)}
\item{V89}{Roll call vote (numeric)}
\item{V90}{Roll call vote (numeric)}
\item{V91}{Roll call vote (numeric)}
\item{V92}{Roll call vote (numeric)}
\item{V93}{Roll call vote (numeric)}
\item{V94}{Roll call vote (numeric)}
\item{V95}{Roll call vote (numeric)}
\item{V96}{Roll call vote (numeric)}
\item{V97}{Roll call vote (numeric)}
\item{V98}{Roll call vote (numeric)}
\item{V99}{Roll call vote (numeric)}
\item{V100}{Roll call vote (numeric)}
\item{V101}{Roll call vote (numeric)}
\item{V102}{Roll call vote (numeric)}
\item{V103}{Roll call vote (numeric)}
\item{V104}{Roll call vote (numeric)}
\item{V105}{Roll call vote (numeric)}
\item{V106}{Roll call vote (numeric)}
\item{V107}{Roll call vote (numeric)}
\item{V108}{Roll call vote (numeric)}
\item{V109}{Roll call vote (numeric)}
\item{V110}{Roll call vote (numeric)}
\item{V111}{Roll call vote (numeric)}
\item{V112}{Roll call vote (numeric)}
\item{V113}{Roll call vote (numeric)}
\item{V114}{Roll call vote (numeric)}
\item{V115}{Roll call vote (numeric)}
\item{V116}{Roll call vote (numeric)}
\item{V117}{Roll call vote (numeric)}
\item{V118}{Roll call vote (numeric)}
\item{V119}{Roll call vote (numeric)}
\item{V120}{Roll call vote (numeric)}
\item{V121}{Roll call vote (numeric)}
\item{V122}{Roll call vote (numeric)}
\item{V123}{Roll call vote (numeric)}
\item{V124}{Roll call vote (numeric)}
\item{V125}{Roll call vote (numeric)}
\item{V126}{Roll call vote (numeric)}
\item{V127}{Roll call vote (numeric)}
\item{V128}{Roll call vote (numeric)}
\item{V129}{Roll call vote (numeric)}
\item{V130}{Roll call vote (numeric)}
\item{V131}{Roll call vote (numeric)}
\item{V132}{Roll call vote (numeric)}
\item{V133}{Roll call vote (numeric)}
\item{V134}{Roll call vote (numeric)}
\item{V135}{Roll call vote (numeric)}
\item{V136}{Roll call vote (numeric)}
\item{V137}{Roll call vote (numeric)}
\item{V138}{Roll call vote (numeric)}
\item{V139}{Roll call vote (numeric)}
\item{V140}{Roll call vote (numeric)}
\item{V141}{Roll call vote (numeric)}
\item{V142}{Roll call vote (numeric)}
\item{V143}{Roll call vote (numeric)}
\item{V144}{Roll call vote (numeric)}
\item{V145}{Roll call vote (numeric)}
\item{V146}{Roll call vote (numeric)}
\item{V147}{Roll call vote (numeric)}
\item{V148}{Roll call vote (numeric)}
\item{V149}{Roll call vote (numeric)}
\item{V150}{Roll call vote (numeric)}
\item{V151}{Roll call vote (numeric)}
\item{V152}{Roll call vote (numeric)}
\item{V153}{Roll call vote (numeric)}
\item{V154}{Roll call vote (numeric)}
\item{V155}{Roll call vote (numeric)}
\item{V156}{Roll call vote (numeric)}
\item{V157}{Roll call vote (numeric)}
\item{V158}{Roll call vote (numeric)}
\item{V159}{Roll call vote (numeric)}
\item{V160}{Roll call vote (numeric)}
\item{V161}{Roll call vote (numeric)}
\item{V162}{Roll call vote (numeric)}
\item{V163}{Roll call vote (numeric)}
\item{V164}{Roll call vote (numeric)}
\item{V165}{Roll call vote (numeric)}
\item{V166}{Roll call vote (numeric)}
\item{V167}{Roll call vote (numeric)}
\item{V168}{Roll call vote (numeric)}
\item{V169}{Roll call vote (numeric)}
\item{V170}{Roll call vote (numeric)}
\item{V171}{Roll call vote (numeric)}
\item{V172}{Roll call vote (numeric)}
\item{V173}{Roll call vote (numeric)}
\item{V174}{Roll call vote (numeric)}
\item{V175}{Roll call vote (numeric)}
\item{V176}{Roll call vote (numeric)}
\item{V177}{Roll call vote (numeric)}
\item{V178}{Roll call vote (numeric)}
\item{V179}{Roll call vote (numeric)}
\item{V180}{Roll call vote (numeric)}
\item{V181}{Roll call vote (numeric)}
\item{V182}{Roll call vote (numeric)}
\item{V183}{Roll call vote (numeric)}
\item{V184}{Roll call vote (numeric)}
\item{V185}{Roll call vote (numeric)}
\item{V186}{Roll call vote (numeric)}
\item{V187}{Roll call vote (numeric)}
\item{V188}{Roll call vote (numeric)}
\item{V189}{Roll call vote (numeric)}
\item{V190}{Roll call vote (numeric)}
\item{V191}{Roll call vote (numeric)}
\item{V192}{Roll call vote (numeric)}
\item{V193}{Roll call vote (numeric)}
\item{V194}{Roll call vote (numeric)}
\item{V195}{Roll call vote (numeric)}
\item{V196}{Roll call vote (numeric)}
\item{V197}{Roll call vote (numeric)}
\item{V198}{Roll call vote (numeric)}
\item{V199}{Roll call vote (numeric)}
\item{V200}{Roll call vote (numeric)}
\item{V201}{Roll call vote (numeric)}
\item{V202}{Roll call vote (numeric)}
\item{V203}{Roll call vote (numeric)}
\item{V204}{Roll call vote (numeric)}
\item{V205}{Roll call vote (numeric)}
\item{V206}{Roll call vote (numeric)}
\item{V207}{Roll call vote (numeric)}
\item{V208}{Roll call vote (numeric)}
\item{V209}{Roll call vote (numeric)}
\item{V210}{Roll call vote (numeric)}
\item{V211}{Roll call vote (numeric)}
\item{V212}{Roll call vote (numeric)}
\item{V213}{Roll call vote (numeric)}
\item{V214}{Roll call vote (numeric)}
\item{V215}{Roll call vote (numeric)}
\item{V216}{Roll call vote (numeric)}
\item{V217}{Roll call vote (numeric)}
\item{V218}{Roll call vote (numeric)}
\item{V219}{Roll call vote (numeric)}
\item{V220}{Roll call vote (numeric)}
\item{V221}{Roll call vote (numeric)}
\item{V222}{Roll call vote (numeric)}
\item{V223}{Roll call vote (numeric)}
\item{V224}{Roll call vote (numeric)}
\item{V225}{Roll call vote (numeric)}
\item{V226}{Roll call vote (numeric)}
\item{V227}{Roll call vote (numeric)}
\item{V228}{Roll call vote (numeric)}
\item{V229}{Roll call vote (numeric)}
\item{V230}{Roll call vote (numeric)}
\item{V231}{Roll call vote (numeric)}
\item{V232}{Roll call vote (numeric)}
\item{V233}{Roll call vote (numeric)}
\item{V234}{Roll call vote (numeric)}
\item{V235}{Roll call vote (numeric)}
\item{V236}{Roll call vote (numeric)}
\item{V237}{Roll call vote (numeric)}
\item{V238}{Roll call vote (numeric)}
\item{V239}{Roll call vote (numeric)}
\item{V240}{Roll call vote (numeric)}
\item{V241}{Roll call vote (numeric)}
\item{V242}{Roll call vote (numeric)}
\item{V243}{Roll call vote (numeric)}
\item{V244}{Roll call vote (numeric)}
\item{V245}{Roll call vote (numeric)}
\item{V246}{Roll call vote (numeric)}
\item{V247}{Roll call vote (numeric)}
\item{V248}{Roll call vote (numeric)}
\item{V249}{Roll call vote (numeric)}
\item{V250}{Roll call vote (numeric)}
\item{V251}{Roll call vote (numeric)}
\item{V252}{Roll call vote (numeric)}
\item{V253}{Roll call vote (numeric)}
\item{V254}{Roll call vote (numeric)}
\item{V255}{Roll call vote (numeric)}
\item{V256}{Roll call vote (numeric)}
\item{V257}{Roll call vote (numeric)}
\item{V258}{Roll call vote (numeric)}
\item{V259}{Roll call vote (numeric)}
\item{V260}{Roll call vote (numeric)}
\item{V261}{Roll call vote (numeric)}
\item{V262}{Roll call vote (numeric)}
\item{V263}{Roll call vote (numeric)}
\item{V264}{Roll call vote (numeric)}
\item{V265}{Roll call vote (numeric)}
\item{V266}{Roll call vote (numeric)}
\item{V267}{Roll call vote (numeric)}
\item{V268}{Roll call vote (numeric)}
\item{V269}{Roll call vote (numeric)}
\item{V270}{Roll call vote (numeric)}
\item{V271}{Roll call vote (numeric)}
\item{V272}{Roll call vote (numeric)}
\item{V273}{Roll call vote (numeric)}
\item{V274}{Roll call vote (numeric)}
\item{V275}{Roll call vote (numeric)}
\item{V276}{Roll call vote (numeric)}
\item{V277}{Roll call vote (numeric)}
\item{V278}{Roll call vote (numeric)}
\item{V279}{Roll call vote (numeric)}
\item{V280}{Roll call vote (numeric)}
\item{V281}{Roll call vote (numeric)}
\item{V282}{Roll call vote (numeric)}
\item{V283}{Roll call vote (numeric)}
\item{V284}{Roll call vote (numeric)}
\item{V285}{Roll call vote (numeric)}
\item{V286}{Roll call vote (numeric)}
\item{V287}{Roll call vote (numeric)}
\item{V288}{Roll call vote (numeric)}
\item{V289}{Roll call vote (numeric)}
\item{V290}{Roll call vote (numeric)}
\item{V291}{Roll call vote (numeric)}
\item{V292}{Roll call vote (numeric)}
\item{V293}{Roll call vote (numeric)}
\item{V294}{Roll call vote (numeric)}
\item{V295}{Roll call vote (numeric)}
\item{V296}{Roll call vote (numeric)}
\item{V297}{Roll call vote (numeric)}
\item{V298}{Roll call vote (numeric)}
\item{V299}{Roll call vote (numeric)}
\item{V300}{Roll call vote (numeric)}
\item{V301}{Roll call vote (numeric)}
\item{V302}{Roll call vote (numeric)}
\item{V303}{Roll call vote (numeric)}
\item{V304}{Roll call vote (numeric)}
\item{V305}{Roll call vote (numeric)}
\item{V306}{Roll call vote (numeric)}
\item{V307}{Roll call vote (numeric)}
\item{V308}{Roll call vote (numeric)}
\item{V309}{Roll call vote (numeric)}
\item{V310}{Roll call vote (numeric)}
\item{V311}{Roll call vote (numeric)}
\item{V312}{Roll call vote (numeric)}
\item{V313}{Roll call vote (numeric)}
\item{V314}{Roll call vote (numeric)}
\item{V315}{Roll call vote (numeric)}
\item{V316}{Roll call vote (numeric)}
\item{V317}{Roll call vote (numeric)}
\item{V318}{Roll call vote (numeric)}
\item{V319}{Roll call vote (numeric)}
\item{V320}{Roll call vote (numeric)}
\item{V321}{Roll call vote (numeric)}
\item{V322}{Roll call vote (numeric)}
\item{V323}{Roll call vote (numeric)}
\item{V324}{Roll call vote (numeric)}
\item{V325}{Roll call vote (numeric)}
\item{V326}{Roll call vote (numeric)}
\item{V327}{Roll call vote (numeric)}
\item{V328}{Roll call vote (numeric)}
\item{V329}{Roll call vote (numeric)}
\item{V330}{Roll call vote (numeric)}
\item{V331}{Roll call vote (numeric)}
\item{V332}{Roll call vote (numeric)}
\item{V333}{Roll call vote (numeric)}
\item{V334}{Roll call vote (numeric)}
\item{V335}{Roll call vote (numeric)}
\item{V336}{Roll call vote (numeric)}
\item{V337}{Roll call vote (numeric)}
\item{V338}{Roll call vote (numeric)}
\item{V339}{Roll call vote (numeric)}
\item{V340}{Roll call vote (numeric)}
\item{V341}{Roll call vote (numeric)}
\item{V342}{Roll call vote (numeric)}
\item{V343}{Roll call vote (numeric)}
\item{V344}{Roll call vote (numeric)}
\item{V345}{Roll call vote (numeric)}
\item{V346}{Roll call vote (numeric)}
\item{V347}{Roll call vote (numeric)}
\item{V348}{Roll call vote (numeric)}
\item{V349}{Roll call vote (numeric)}
\item{V350}{Roll call vote (numeric)}
\item{V351}{Roll call vote (numeric)}
\item{V352}{Roll call vote (numeric)}
\item{V353}{Roll call vote (numeric)}
\item{V354}{Roll call vote (numeric)}
\item{V355}{Roll call vote (numeric)}
\item{V356}{Roll call vote (numeric)}
\item{V357}{Roll call vote (numeric)}
\item{V358}{Roll call vote (numeric)}
\item{V359}{Roll call vote (numeric)}
\item{V360}{Roll call vote (numeric)}
\item{V361}{Roll call vote (numeric)}
\item{V362}{Roll call vote (numeric)}
\item{V363}{Roll call vote (numeric)}
\item{V364}{Roll call vote (numeric)}
\item{V365}{Roll call vote (numeric)}
\item{V366}{Roll call vote (numeric)}
\item{V367}{Roll call vote (numeric)}
\item{V368}{Roll call vote (numeric)}
\item{V369}{Roll call vote (numeric)}
\item{V370}{Roll call vote (numeric)}
\item{V371}{Roll call vote (numeric)}
\item{V372}{Roll call vote (numeric)}
\item{V373}{Roll call vote (numeric)}
\item{V374}{Roll call vote (numeric)}
\item{V375}{Roll call vote (numeric)}
\item{V376}{Roll call vote (numeric)}
\item{V377}{Roll call vote (numeric)}
\item{V378}{Roll call vote (numeric)}
\item{V379}{Roll call vote (numeric)}
\item{V380}{Roll call vote (numeric)}
\item{V381}{Roll call vote (numeric)}
\item{V382}{Roll call vote (numeric)}
\item{V383}{Roll call vote (numeric)}
\item{V384}{Roll call vote (numeric)}
\item{V385}{Roll call vote (numeric)}
\item{V386}{Roll call vote (numeric)}
\item{V387}{Roll call vote (numeric)}
\item{V388}{Roll call vote (numeric)}
\item{V389}{Roll call vote (numeric)}
\item{V390}{Roll call vote (numeric)}
\item{V391}{Roll call vote (numeric)}
\item{V392}{Roll call vote (numeric)}
\item{V393}{Roll call vote (numeric)}
\item{V394}{Roll call vote (numeric)}
\item{V395}{Roll call vote (numeric)}
\item{V396}{Roll call vote (numeric)}
\item{V397}{Roll call vote (numeric)}
\item{V398}{Roll call vote (numeric)}
\item{V399}{Roll call vote (numeric)}
\item{V400}{Roll call vote (numeric)}
\item{V401}{Roll call vote (numeric)}
\item{V402}{Roll call vote (numeric)}
\item{V403}{Roll call vote (numeric)}
\item{V404}{Roll call vote (numeric)}
\item{V405}{Roll call vote (numeric)}
\item{V406}{Roll call vote (numeric)}
\item{V407}{Roll call vote (numeric)}
\item{V408}{Roll call vote (numeric)}
\item{V409}{Roll call vote (numeric)}
\item{V410}{Roll call vote (numeric)}
\item{V411}{Roll call vote (numeric)}
\item{V412}{Roll call vote (numeric)}
\item{V413}{Roll call vote (numeric)}
\item{V414}{Roll call vote (numeric)}
\item{V415}{Roll call vote (numeric)}
\item{V416}{Roll call vote (numeric)}
\item{V417}{Roll call vote (numeric)}
\item{V418}{Roll call vote (numeric)}
\item{V419}{Roll call vote (numeric)}
\item{V420}{Roll call vote (numeric)}
\item{V421}{Roll call vote (numeric)}
\item{V422}{Roll call vote (numeric)}
\item{V423}{Roll call vote (numeric)}
\item{V424}{Roll call vote (numeric)}
\item{V425}{Roll call vote (numeric)}
\item{V426}{Roll call vote (numeric)}
\item{V427}{Roll call vote (numeric)}
\item{V428}{Roll call vote (numeric)}
\item{V429}{Roll call vote (numeric)}
\item{V430}{Roll call vote (numeric)}
\item{V431}{Roll call vote (numeric)}
\item{V432}{Roll call vote (numeric)}
\item{V433}{Roll call vote (numeric)}
\item{V434}{Roll call vote (numeric)}
\item{V435}{Roll call vote (numeric)}
\item{V436}{Roll call vote (numeric)}
\item{V437}{Roll call vote (numeric)}
\item{V438}{Roll call vote (numeric)}
\item{V439}{Roll call vote (numeric)}
\item{V440}{Roll call vote (numeric)}
\item{V441}{Roll call vote (numeric)}
\item{V442}{Roll call vote (numeric)}
\item{V443}{Roll call vote (numeric)}
\item{V444}{Roll call vote (numeric)}
\item{V445}{Roll call vote (numeric)}
\item{V446}{Roll call vote (numeric)}
\item{V447}{Roll call vote (numeric)}
\item{V448}{Roll call vote (numeric)}
\item{V449}{Roll call vote (numeric)}
\item{V450}{Roll call vote (numeric)}
\item{V451}{Roll call vote (numeric)}
\item{V452}{Roll call vote (numeric)}
\item{V453}{Roll call vote (numeric)}
\item{V454}{Roll call vote (numeric)}
\item{V455}{Roll call vote (numeric)}
\item{V456}{Roll call vote (numeric)}
\item{V457}{Roll call vote (numeric)}
\item{V458}{Roll call vote (numeric)}
\item{V459}{Roll call vote (numeric)}
\item{V460}{Roll call vote (numeric)}
\item{V461}{Roll call vote (numeric)}
\item{V462}{Roll call vote (numeric)}
\item{V463}{Roll call vote (numeric)}
\item{V464}{Roll call vote (numeric)}
\item{V465}{Roll call vote (numeric)}
\item{V466}{Roll call vote (numeric)}
\item{V467}{Roll call vote (numeric)}
\item{V468}{Roll call vote (numeric)}
\item{V469}{Roll call vote (numeric)}
\item{V470}{Roll call vote (numeric)}
\item{V471}{Roll call vote (numeric)}
\item{V472}{Roll call vote (numeric)}
\item{V473}{Roll call vote (numeric)}
\item{V474}{Roll call vote (numeric)}
\item{V475}{Roll call vote (numeric)}
\item{V476}{Roll call vote (numeric)}
\item{V477}{Roll call vote (numeric)}
\item{V478}{Roll call vote (numeric)}
\item{V479}{Roll call vote (numeric)}
\item{V480}{Roll call vote (numeric)}
\item{V481}{Roll call vote (numeric)}
\item{V482}{Roll call vote (numeric)}
\item{V483}{Roll call vote (numeric)}
\item{V484}{Roll call vote (numeric)}
\item{V485}{Roll call vote (numeric)}
\item{V486}{Roll call vote (numeric)}
\item{V487}{Roll call vote (numeric)}
\item{V488}{Roll call vote (numeric)}
\item{V489}{Roll call vote (numeric)}
\item{V490}{Roll call vote (numeric)}
\item{V491}{Roll call vote (numeric)}
\item{V492}{Roll call vote (numeric)}
\item{V493}{Roll call vote (numeric)}
\item{V494}{Roll call vote (numeric)}
\item{V495}{Roll call vote (numeric)}
\item{V496}{Roll call vote (numeric)}
\item{V497}{Roll call vote (numeric)}
\item{V498}{Roll call vote (numeric)}
\item{V499}{Roll call vote (numeric)}
\item{V500}{Roll call vote (numeric)}
\item{V501}{Roll call vote (numeric)}
\item{V502}{Roll call vote (numeric)}
\item{V503}{Roll call vote (numeric)}
\item{V504}{Roll call vote (numeric)}
\item{V505}{Roll call vote (numeric)}
\item{V506}{Roll call vote (numeric)}
\item{V507}{Roll call vote (numeric)}
\item{V508}{Roll call vote (numeric)}
\item{V509}{Roll call vote (numeric)}
\item{V510}{Roll call vote (numeric)}
\item{V511}{Roll call vote (numeric)}
\item{V512}{Roll call vote (numeric)}
\item{V513}{Roll call vote (numeric)}
\item{V514}{Roll call vote (numeric)}
\item{V515}{Roll call vote (numeric)}
\item{V516}{Roll call vote (numeric)}
\item{V517}{Roll call vote (numeric)}
\item{V518}{Roll call vote (numeric)}
\item{V519}{Roll call vote (numeric)}
\item{V520}{Roll call vote (numeric)}
\item{V521}{Roll call vote (numeric)}
\item{V522}{Roll call vote (numeric)}
\item{V523}{Roll call vote (numeric)}
\item{V524}{Roll call vote (numeric)}
\item{V525}{Roll call vote (numeric)}
\item{V526}{Roll call vote (numeric)}
\item{V527}{Roll call vote (numeric)}
\item{V528}{Roll call vote (numeric)}
\item{V529}{Roll call vote (numeric)}
\item{V530}{Roll call vote (numeric)}
\item{V531}{Roll call vote (numeric)}
\item{V532}{Roll call vote (numeric)}
\item{V533}{Roll call vote (numeric)}
\item{V534}{Roll call vote (numeric)}
\item{V535}{Roll call vote (numeric)}
\item{V536}{Roll call vote (numeric)}
\item{V537}{Roll call vote (numeric)}
\item{V538}{Roll call vote (numeric)}
\item{V539}{Roll call vote (numeric)}
\item{V540}{Roll call vote (numeric)}
\item{V541}{Roll call vote (numeric)}
\item{V542}{Roll call vote (numeric)}
\item{V543}{Roll call vote (numeric)}
\item{V544}{Roll call vote (numeric)}
\item{V545}{Roll call vote (numeric)}
\item{V546}{Roll call vote (numeric)}
\item{V547}{Roll call vote (numeric)}
\item{V548}{Roll call vote (numeric)}
\item{V549}{Roll call vote (numeric)}
\item{V550}{Roll call vote (numeric)}
\item{V551}{Roll call vote (numeric)}
\item{V552}{Roll call vote (numeric)}
\item{V553}{Roll call vote (numeric)}
\item{V554}{Roll call vote (numeric)}
\item{V555}{Roll call vote (numeric)}
\item{V556}{Roll call vote (numeric)}
\item{V557}{Roll call vote (numeric)}
\item{V558}{Roll call vote (numeric)}
\item{V559}{Roll call vote (numeric)}
\item{V560}{Roll call vote (numeric)}
\item{V561}{Roll call vote (numeric)}
\item{V562}{Roll call vote (numeric)}
\item{V563}{Roll call vote (numeric)}
\item{V564}{Roll call vote (numeric)}
\item{V565}{Roll call vote (numeric)}
\item{V566}{Roll call vote (numeric)}
\item{V567}{Roll call vote (numeric)}
\item{V568}{Roll call vote (numeric)}
\item{V569}{Roll call vote (numeric)}
\item{V570}{Roll call vote (numeric)}
\item{V571}{Roll call vote (numeric)}
\item{V572}{Roll call vote (numeric)}
\item{V573}{Roll call vote (numeric)}
\item{V574}{Roll call vote (numeric)}
\item{V575}{Roll call vote (numeric)}
\item{V576}{Roll call vote (numeric)}
\item{V577}{Roll call vote (numeric)}
\item{V578}{Roll call vote (numeric)}
\item{V579}{Roll call vote (numeric)}
\item{V580}{Roll call vote (numeric)}
\item{V581}{Roll call vote (numeric)}
\item{V582}{Roll call vote (numeric)}
\item{V583}{Roll call vote (numeric)}
\item{V584}{Roll call vote (numeric)}
\item{V585}{Roll call vote (numeric)}
\item{V586}{Roll call vote (numeric)}
\item{V587}{Roll call vote (numeric)}
\item{V588}{Roll call vote (numeric)}
\item{V589}{Roll call vote (numeric)}
\item{V590}{Roll call vote (numeric)}
\item{V591}{Roll call vote (numeric)}
\item{V592}{Roll call vote (numeric)}
\item{V593}{Roll call vote (numeric)}
\item{V594}{Roll call vote (numeric)}
\item{V595}{Roll call vote (numeric)}
\item{V596}{Roll call vote (numeric)}
\item{V597}{Roll call vote (numeric)}
\item{V598}{Roll call vote (numeric)}
\item{V599}{Roll call vote (numeric)}
\item{V600}{Roll call vote (numeric)}
\item{V601}{Roll call vote (numeric)}
\item{V602}{Roll call vote (numeric)}
\item{V603}{Roll call vote (numeric)}
\item{V604}{Roll call vote (numeric)}
\item{V605}{Roll call vote (numeric)}
\item{V606}{Roll call vote (numeric)}
\item{V607}{Roll call vote (numeric)}
\item{V608}{Roll call vote (numeric)}
\item{V609}{Roll call vote (numeric)}
\item{V610}{Roll call vote (numeric)}
\item{V611}{Roll call vote (numeric)}
\item{V612}{Roll call vote (numeric)}
\item{V613}{Roll call vote (numeric)}
\item{V614}{Roll call vote (numeric)}
\item{V615}{Roll call vote (numeric)}
\item{V616}{Roll call vote (numeric)}
\item{V617}{Roll call vote (numeric)}
\item{V618}{Roll call vote (numeric)}
\item{V619}{Roll call vote (numeric)}
\item{V620}{Roll call vote (numeric)}
\item{V621}{Roll call vote (numeric)}
\item{V622}{Roll call vote (numeric)}
\item{V623}{Roll call vote (numeric)}
\item{V624}{Roll call vote (numeric)}
\item{V625}{Roll call vote (numeric)}
\item{V626}{Roll call vote (numeric)}
\item{V627}{Roll call vote (numeric)}
\item{V628}{Roll call vote (numeric)}
\item{V629}{Roll call vote (numeric)}
\item{V630}{Roll call vote (numeric)}
\item{V631}{Roll call vote (numeric)}
\item{V632}{Roll call vote (numeric)}
\item{V633}{Roll call vote (numeric)}
\item{V634}{Roll call vote (numeric)}
\item{V635}{Roll call vote (numeric)}
\item{V636}{Roll call vote (numeric)}
\item{V637}{Roll call vote (numeric)}
\item{V638}{Roll call vote (numeric)}
\item{V639}{Roll call vote (numeric)}
\item{V640}{Roll call vote (numeric)}
\item{V641}{Roll call vote (numeric)}
\item{V642}{Roll call vote (numeric)}
\item{V643}{Roll call vote (numeric)}
\item{V644}{Roll call vote (numeric)}
\item{V645}{Roll call vote (numeric)}
\item{V646}{Roll call vote (numeric)}
\item{V647}{Roll call vote (numeric)}
\item{V648}{Roll call vote (numeric)}
\item{V649}{Roll call vote (numeric)}
\item{V650}{Roll call vote (numeric)}
\item{V651}{Roll call vote (numeric)}
\item{V652}{Roll call vote (numeric)}
\item{V653}{Roll call vote (numeric)}
\item{V654}{Roll call vote (numeric)}
\item{V655}{Roll call vote (numeric)}
\item{V656}{Roll call vote (numeric)}
\item{V657}{Roll call vote (numeric)}
\item{V658}{Roll call vote (numeric)}
\item{V659}{Roll call vote (numeric)}
\item{V660}{Roll call vote (numeric)}
\item{V661}{Roll call vote (numeric)}
\item{V662}{Roll call vote (numeric)}
\item{V663}{Roll call vote (numeric)}
\item{V664}{Roll call vote (numeric)}
\item{V665}{Roll call vote (numeric)}
\item{V666}{Roll call vote (numeric)}
\item{V667}{Roll call vote (numeric)}
\item{V668}{Roll call vote (numeric)}
\item{V669}{Roll call vote (numeric)}
\item{V670}{Roll call vote (numeric)}
\item{V671}{Roll call vote (numeric)}
\item{V672}{Roll call vote (numeric)}
\item{V673}{Roll call vote (numeric)}
\item{V674}{Roll call vote (numeric)}
\item{V675}{Roll call vote (numeric)}
\item{V676}{Roll call vote (numeric)}
\item{V677}{Roll call vote (numeric)}
\item{V678}{Roll call vote (numeric)}
\item{V679}{Roll call vote (numeric)}
\item{V680}{Roll call vote (numeric)}
\item{V681}{Roll call vote (numeric)}
\item{V682}{Roll call vote (numeric)}
\item{V683}{Roll call vote (numeric)}
\item{V684}{Roll call vote (numeric)}
\item{V685}{Roll call vote (numeric)}
\item{V686}{Roll call vote (numeric)}
\item{V687}{Roll call vote (numeric)}
\item{V688}{Roll call vote (numeric)}
\item{V689}{Roll call vote (numeric)}
\item{V690}{Roll call vote (numeric)}
\item{V691}{Roll call vote (numeric)}
\item{V692}{Roll call vote (numeric)}
\item{V693}{Roll call vote (numeric)}
\item{V694}{Roll call vote (numeric)}
\item{V695}{Roll call vote (numeric)}
\item{V696}{Roll call vote (numeric)}
\item{V697}{Roll call vote (numeric)}
\item{V698}{Roll call vote (numeric)}
\item{V699}{Roll call vote (numeric)}
\item{V700}{Roll call vote (numeric)}
\item{V701}{Roll call vote (numeric)}
\item{V702}{Roll call vote (numeric)}
\item{V703}{Roll call vote (numeric)}
\item{V704}{Roll call vote (numeric)}
\item{V705}{Roll call vote (numeric)}
\item{V706}{Roll call vote (numeric)}
\item{V707}{Roll call vote (numeric)}
\item{V708}{Roll call vote (numeric)}
\item{V709}{Roll call vote (numeric)}
\item{V710}{Roll call vote (numeric)}
\item{V711}{Roll call vote (numeric)}
\item{V712}{Roll call vote (numeric)}
\item{V713}{Roll call vote (numeric)}
\item{V714}{Roll call vote (numeric)}
\item{V715}{Roll call vote (numeric)}
\item{V716}{Roll call vote (numeric)}
\item{V717}{Roll call vote (numeric)}
\item{V718}{Roll call vote (numeric)}
\item{V719}{Roll call vote (numeric)}
\item{V720}{Roll call vote (numeric)}
\item{V721}{Roll call vote (numeric)}
\item{V722}{Roll call vote (numeric)}
\item{V723}{Roll call vote (numeric)}
\item{V724}{Roll call vote (numeric)}
\item{V725}{Roll call vote (numeric)}
\item{V726}{Roll call vote (numeric)}
\item{V727}{Roll call vote (numeric)}
\item{V728}{Roll call vote (numeric)}
\item{V729}{Roll call vote (numeric)}
\item{V730}{Roll call vote (numeric)}
\item{V731}{Roll call vote (numeric)}
\item{V732}{Roll call vote (numeric)}
\item{V733}{Roll call vote (numeric)}
\item{V734}{Roll call vote (numeric)}
\item{V735}{Roll call vote (numeric)}
\item{V736}{Roll call vote (numeric)}
\item{V737}{Roll call vote (numeric)}
\item{V738}{Roll call vote (numeric)}
\item{V739}{Roll call vote (numeric)}
\item{V740}{Roll call vote (numeric)}
\item{V741}{Roll call vote (numeric)}
\item{V742}{Roll call vote (numeric)}
\item{V743}{Roll call vote (numeric)}
\item{V744}{Roll call vote (numeric)}
\item{V745}{Roll call vote (numeric)}
\item{V746}{Roll call vote (numeric)}
\item{V747}{Roll call vote (numeric)}
\item{V748}{Roll call vote (numeric)}
\item{V749}{Roll call vote (numeric)}
\item{V750}{Roll call vote (numeric)}
\item{V751}{Roll call vote (numeric)}
\item{V752}{Roll call vote (numeric)}
\item{V753}{Roll call vote (numeric)}
\item{V754}{Roll call vote (numeric)}
\item{V755}{Roll call vote (numeric)}
\item{V756}{Roll call vote (numeric)}
\item{V757}{Roll call vote (numeric)}
\item{V758}{Roll call vote (numeric)}
\item{V759}{Roll call vote (numeric)}
\item{V760}{Roll call vote (numeric)}
\item{V761}{Roll call vote (numeric)}
\item{V762}{Roll call vote (numeric)}
\item{V763}{Roll call vote (numeric)}
\item{V764}{Roll call vote (numeric)}
\item{V765}{Roll call vote (numeric)}
\item{V766}{Roll call vote (numeric)}
\item{V767}{Roll call vote (numeric)}
\item{V768}{Roll call vote (numeric)}
\item{V769}{Roll call vote (numeric)}
\item{V770}{Roll call vote (numeric)}
\item{V771}{Roll call vote (numeric)}
\item{V772}{Roll call vote (numeric)}
\item{V773}{Roll call vote (numeric)}
\item{V774}{Roll call vote (numeric)}
\item{V775}{Roll call vote (numeric)}
\item{V776}{Roll call vote (numeric)}
\item{V777}{Roll call vote (numeric)}
\item{V778}{Roll call vote (numeric)}
\item{V779}{Roll call vote (numeric)}
\item{V780}{Roll call vote (numeric)}
\item{V781}{Roll call vote (numeric)}
\item{V782}{Roll call vote (numeric)}
\item{V783}{Roll call vote (numeric)}
\item{V784}{Roll call vote (numeric)}
\item{V785}{Roll call vote (numeric)}
\item{V786}{Roll call vote (numeric)}
\item{V787}{Roll call vote (numeric)}
\item{V788}{Roll call vote (numeric)}
\item{V789}{Roll call vote (numeric)}
\item{V790}{Roll call vote (numeric)}
\item{V791}{Roll call vote (numeric)}
\item{V792}{Roll call vote (numeric)}
\item{V793}{Roll call vote (numeric)}
\item{V794}{Roll call vote (numeric)}
\item{V795}{Roll call vote (numeric)}
\item{V796}{Roll call vote (numeric)}
\item{V797}{Roll call vote (numeric)}
\item{V798}{Roll call vote (numeric)}
\item{V799}{Roll call vote (numeric)}
\item{V800}{Roll call vote (numeric)}
\item{V801}{Roll call vote (numeric)}
\item{V802}{Roll call vote (numeric)}
\item{V803}{Roll call vote (numeric)}
\item{V804}{Roll call vote (numeric)}
\item{V805}{Roll call vote (numeric)}
\item{V806}{Roll call vote (numeric)}
\item{V807}{Roll call vote (numeric)}
\item{V808}{Roll call vote (numeric)}
\item{V809}{Roll call vote (numeric)}
\item{V810}{Roll call vote (numeric)}
\item{V811}{Roll call vote (numeric)}
\item{V812}{Roll call vote (numeric)}
\item{V813}{Roll call vote (numeric)}
\item{V814}{Roll call vote (numeric)}
\item{V815}{Roll call vote (numeric)}
\item{V816}{Roll call vote (numeric)}
\item{V817}{Roll call vote (numeric)}
\item{V818}{Roll call vote (numeric)}
\item{V819}{Roll call vote (numeric)}
\item{V820}{Roll call vote (numeric)}
\item{V821}{Roll call vote (numeric)}
\item{V822}{Roll call vote (numeric)}
\item{V823}{Roll call vote (numeric)}
\item{V824}{Roll call vote (numeric)}
\item{V825}{Roll call vote (numeric)}
\item{V826}{Roll call vote (numeric)}
\item{V827}{Roll call vote (numeric)}
\item{V828}{Roll call vote (numeric)}
\item{V829}{Roll call vote (numeric)}
\item{V830}{Roll call vote (numeric)}
\item{V831}{Roll call vote (numeric)}
\item{V832}{Roll call vote (numeric)}
\item{V833}{Roll call vote (numeric)}
\item{V834}{Roll call vote (numeric)}
\item{V835}{Roll call vote (numeric)}
\item{V836}{Roll call vote (numeric)}
\item{V837}{Roll call vote (numeric)}
\item{V838}{Roll call vote (numeric)}
\item{V839}{Roll call vote (numeric)}
\item{V840}{Roll call vote (numeric)}
\item{V841}{Roll call vote (numeric)}
\item{V842}{Roll call vote (numeric)}
\item{V843}{Roll call vote (numeric)}
\item{V844}{Roll call vote (numeric)}
\item{V845}{Roll call vote (numeric)}
\item{V846}{Roll call vote (numeric)}
\item{V847}{Roll call vote (numeric)}
\item{V848}{Roll call vote (numeric)}
\item{V849}{Roll call vote (numeric)}
\item{V850}{Roll call vote (numeric)}
\item{V851}{Roll call vote (numeric)}
\item{V852}{Roll call vote (numeric)}
\item{V853}{Roll call vote (numeric)}
\item{V854}{Roll call vote (numeric)}
\item{V855}{Roll call vote (numeric)}
\item{V856}{Roll call vote (numeric)}
\item{V857}{Roll call vote (numeric)}
\item{V858}{Roll call vote (numeric)}
\item{V859}{Roll call vote (numeric)}
\item{V860}{Roll call vote (numeric)}
\item{V861}{Roll call vote (numeric)}
\item{V862}{Roll call vote (numeric)}
\item{V863}{Roll call vote (numeric)}
\item{V864}{Roll call vote (numeric)}
\item{V865}{Roll call vote (numeric)}
\item{V866}{Roll call vote (numeric)}
\item{V867}{Roll call vote (numeric)}
\item{V868}{Roll call vote (numeric)}
\item{V869}{Roll call vote (numeric)}
\item{V870}{Roll call vote (numeric)}
\item{V871}{Roll call vote (numeric)}
\item{V872}{Roll call vote (numeric)}
\item{V873}{Roll call vote (numeric)}
\item{V874}{Roll call vote (numeric)}
\item{V875}{Roll call vote (numeric)}
\item{V876}{Roll call vote (numeric)}
\item{V877}{Roll call vote (numeric)}
\item{V878}{Roll call vote (numeric)}
\item{V879}{Roll call vote (numeric)}
\item{V880}{Roll call vote (numeric)}
\item{V881}{Roll call vote (numeric)}
\item{V882}{Roll call vote (numeric)}
\item{V883}{Roll call vote (numeric)}
\item{V884}{Roll call vote (numeric)}
\item{V885}{Roll call vote (numeric)}
\item{V886}{Roll call vote (numeric)}
}
}
\source{
Hix, S., Noury, A., & Roland, G. (2006). Dimensions of Politics in the
European Parliament. \emph{American Journal of Political Science}, 50(2),
494-520. \doi{10.1111/j.1540-5907.2006.00198.x}

Original data: \url{http://personal.lse.ac.uk/hix/HixNouryRolandEPdata.htm}
}
\usage{
data(rcv_ep1)
}
\description{
Roll Call Voting Data from the First European Parliament (1979-1984)
#' @description Roll call voting data from the first elected European Parliament (1979-1984),
assembled by Hix, Noury, and Roland (2006). Contains voting records of
Members of the European Parliament (MEPs) on 886 roll call votes.
}
\details{
This dataset covers the first directly elected European Parliament
(1979-1984), a formative period for EU legislative politics. The first
five columns contain MEP identification and affiliation variables; the
remaining 886 columns represent roll call votes.

Each row represents one MEP, and each vote column (V1 through V886)
represents a specific legislative vote. The dataset is useful for analyzing
transnational party behavior, national vs. European party loyalty, and
coalition formation in the early European Parliament.
}
\examples{
\dontrun{
data(rcv_ep1)

# Dataset dimensions
dim(rcv_ep1)  # 410 MEPs, 891 columns

# View MEP information
head(rcv_ep1[, 1:5])

# Distribution by country
table(rcv_ep1$MS)

# Distribution by EP group
table(rcv_ep1$EPG)

# View first few votes
head(rcv_ep1[, 6:10])

# Check voting patterns
summary(rcv_ep1[, 6:891])
}

}
\references{
Hix, S., Noury, A., & Roland, G. (2006). Dimensions of Politics in the
European Parliament. \emph{American Journal of Political Science}, 50(2),
494-520.
}
\keyword{datasets}
