% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rcv_ep1}
\alias{rcv_ep1}
\title{Roll Call Voting Data from the First European Parliament (1979–1984)}
\format{
A data frame (or matrix) with the following variables:
\describe{
\item{MEPID}{Unique identifier for each Member of the European Parliament (MEP).}
\item{MEPNAME}{Name of the MEP.}
\item{MS}{Country (Member State) of the MEP.}
\item{NP}{National party affiliation of the MEP.}
\item{EPG}{European Parliament group (party) affiliation of the MEP.}
\item{V1}{Result of the first roll call vote (and so on for subsequent votes).}
\item{V2-V886}{Results of the roll call votes from the second to the 886th vote. Each column represents a specific vote, and the value indicates the MEP's vote on that issue.}
}
}
\source{
Data from Hix, Noury, and Roland (2006), available online at \url{http://personal.lse.ac.uk/hix/HixNouryRolandEPdata.htm}.
}
\usage{
data(rcv_ep1)
}
\description{
The \code{rcv_ep1} dataset contains roll call voting data from the first European Parliament (1979–1984), as assembled by Hix, Noury, and Roland (2006).
The dataset includes information on Members of the European Parliament (MEPs) and their votes on various issues.
}
\details{
The first five columns of the dataset are legislator-specific variables: MEPID (ID number), MEPNAME (name), MS (country), NP (national party affiliation), and EPG (European Parliament group affiliation).
The remaining columns (V1 to V886) represent the roll call votes. Each vote is represented by a column, and the value in each cell indicates how the MEP voted on that particular issue.

This dataset was used to analyze voting behavior in the European Parliament during its first term. The raw data can be accessed from \url{http://personal.lse.ac.uk/hix/HixNouryRolandEPdata.htm}.
}
\examples{
\dontrun{
data(rcv_ep1)
}

}
\keyword{datasets}
