\name{senate.90}
\alias{senate.90}
\docType{data}
\title{
90th Senate Agreement Scores
}
\description{
Agreement scores for the 90th US Senate (1967-1968).  
}
\usage{data(senate.90)}
\format{
A symmetric 102x102 matrix of agreement scores for the senators from the 90th senate.  The \sQuote{idvars} attribute of the matrix provides some data about the senators represented in the rows of the data with the following variables: 
  \describe{
    \item{\code{congress}}{Congress Number}
    \item{\code{id}}{Numeric Senator ID}
    \item{\code{statecode}}{State Code}
    \item{\code{statename}}{State Name}
    \item{\code{party}}{Party}
    \item{\code{election}}{Election}
    \item{\code{name}}{Senator name}
  }
}
\keyword{datasets}
