\name{start.mcmc.list}
\alias{start.mcmc.list}
\alias{end.mcmc.list}
\alias{window.mcmc.list}
\title{Time Functions for MCMC Objects}
\description{Time functions for MCMC objects. }

\usage{
\method{start}{mcmc.list}(x, ...)
\method{end}{mcmc.list}(x, ...)
\method{window}{mcmc.list}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{An object of class \code{mcmc.list}.}
  \item{...}{Other arguments to be passed down.}
}
\value{
A time start, end or window
}
