% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{start.mcmc.list}
\alias{start.mcmc.list}
\title{Get the Start Time of an MCMC List}
\usage{
\method{start}{mcmc.list}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mcmc.list}, which is a list of MCMC chains.}

\item{...}{Additional arguments (currently not used).}
}
\value{
The result of calling the \code{start} function on the first element of the MCMC list.
}
\description{
This function returns the start time or the first element of the first chain in an MCMC list.
}
\examples{
# Assuming `mcmc_list` is an object of class `mcmc.list`
# start_time <- start.mcmc.list(mcmc_list)

}
