\name{Sweden2010}
\alias{Sweden2010}
\docType{data}
\title{
Swedish Parliamentary Candidate Survey (2010)
}
\description{
2010 Survey of Swedish Parliamentary Candidates 
}
\usage{data(Sweden2010)}
\format{
  A data frame with 2751 observations on the following 56 variables.  Each of the policy statements is coded 1 (strongly disagree), 2 (disagree), 3 (agree), 4 (strongly agree), 8 (don't know)
  \describe{
    \item{\code{id}}{Candidate ID}
    \item{\code{elected}}{Whether the candidate was elected (1) or not (0)}
    \item{\code{party.name}}{Name of candidate's party}
    \item{\code{party.code}}{Candidate's party code}
    \item{\code{govt.party}}{Whether the candidate is in a government party (1) or not (0)}
    \item{\code{left.right.self.fivept}}{Candidate's left-right self-placement from (1=left to 5=right)}
    \item{\code{congestion.taxes}}{Congestion taxes should be introduced in more big cities}
    \item{\code{highspeed.trains}}{Sweden should invest in high-speed trains}
    \item{\code{hunt.wolves}}{Sweden should allow license hunt of wolves}
    \item{\code{nuclear.power}}{The nuclear power should be further expanded}
    \item{\code{gasoline.taxes}}{The tax on gasoline should be raised}
    \item{\code{museum.fees}}{The entrance fee for state-owned museums should be removed}
    \item{\code{online.piracy}}{It should be a criminal offense to download copyright protected material from the internet for personal use}
    \item{\code{state.TV}}{The fee for state-owned TV should be replaced with a mandatory tax}{No municipality should be allowed to refuse to take in refugees}
    \item{\code{refugee.cities}}{No municipality should be allowed to refuse to take in refugees}
    \item{\code{asylum.seekers}}{Fewer asylum seekers should be accepted into Sweden}
    \item{\code{refugee.healthcare}}{Refugees without documents should be able to receive free health care}
    \item{\code{teacher.veils}}{Teachers should not be allowed to wear face-covering veil in Swedish schools}
    \item{\code{paternal.leave}}{A bigger share of the parental leave should be reserved for the fathers}
    \item{\code{affirmative.action.universities}}{The use of affirmative action in university admission should be abolished}
    \item{\code{child.raising.allowance}}{The child-raising allowance should be abolished}
    \item{\code{property.taxes.wealthy}}{The property tax for detached houses with a taxable value above 4.5 millions should be raised}
    \item{\code{wealth.tax}}{The wealth tax should be reinstated}
    \item{\code{tax.wealthy}}{The tax rate for those with high incomes should be raised}
    \item{\code{tax.pensions}}{The same tax rate should be applied to pensions and wages}
    \item{\code{household.services.deduction}}{The tax deduction for household services [RUT] should be kept}
    \item{\code{work.income.tax}}{The tax on work-income should be lowered}
    \item{\code{sex.purchase}}{The sex purchase law should be abolished}
    \item{\code{DUI.penalty}}{The penalty for driving under intoxication should be increased}
    \item{\code{criminal.sentences}}{Violent crimes should lead to tougher sentences}
    \item{\code{wiretaps}}{The FRA law [a legislative package that authorizes the state to wiretap all telephone and Internet traffic that crosses Sweden's borders] should be torn up}
    \item{\code{retirement.age}}{Those who are 58 years old should be eligible for early retirement}
    \item{\code{health.insurance.time}}{The maximum time on health insurance should be limited}
    \item{\code{dental.insurance}}{Dental care should be included in health care insurance}
    \item{\code{competition.public.sector}}{There should be more competition in the public sector}
    \item{\code{mandatory.unemployment.insurance}}{Unemployment insurance should be mandatory}
    \item{\code{municipal.home.care}}{Municipal home-care service fees should be fully equalized across municipalities }
    \item{\code{circumcision}}{Swedish health care should only offer to carry out circumcision when it has a medical value}
    \item{\code{private.healthcare.profits}}{Government financed private health care service providers should be able to distribute profits to their owners}
    \item{\code{govt.alcohol.monopoly}}{The government monopoly on selling alcoholic beverages should be removed}
    \item{\code{employment.protection}}{It should be made more difficult for personnel agencies to circumvent employment protection legislation}
    \item{\code{sell.public.corporations}}{State-owned corporations should not be sold}
    \item{\code{Aghanistan.withdrawal}}{Swedish soldiers in Afghanistan should be brought immediately}
    \item{\code{exporting.arms}}{Swedish should cease exporting arms immediately}
    \item{\code{aid.undemocratic.countries}}{Foreign aid should not go to undemocratic countries}
    \item{\code{compulsory.military.service}}{The compulsory military service should be reinstated}
    \item{\code{leave.EU}}{Sweden should leave the European Union}
    \item{\code{transfer.students}}{It should be possible to transfer students with discipline problems from their school against the will of the student and their parents}
    \item{\code{local.control.education}}{Responsibility for primary and secondary education should be moved from municipalities to the national government}
    \item{\code{student.grades}}{Students should received grades from garde six and onwards}
    \item{\code{number.private.schools}}{The number of private schools should be limited}
    \item{\code{university.eligibility}}{All who pass secondary school programs should be eligible for university studies}
    \item{\code{criminalize.racist.organizations}}{It should be a criminal offense to be a member of a racist organisation}
    \item{\code{abolish.monarchy}}{Monarchy should be abolished}
    \item{\code{ballot.order}}{Voter control over the order of candidates on ballots in Swedish elections should be strengthened}
    \item{\code{referenda.elections}}{More referenda should be held in Sweden}
  }
}
\source{
Need source citation
}
\keyword{datasets}
