% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{window.mcmc.list}
\alias{window.mcmc.list}
\title{Apply Windowing to an MCMC List}
\usage{
\method{window}{mcmc.list}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mcmc.list}, which is a list of MCMC chains.}

\item{...}{Additional arguments passed to the \code{window.mcmc} function, such as \code{start}, \code{end}, and \code{thin}.}
}
\value{
An object of class \code{mcmc.list} with the same structure as the input but with each chain windowed according to the specified parameters.
}
\description{
This function applies the \code{window.mcmc} function to each chain in an MCMC list, allowing you to subset the chains based on the desired window.
}
\examples{
# Assuming `mcmc_list` is an object of class `mcmc.list`
# Subset the MCMC list to include only iterations between 500 and 1000
windowed_mcmc_list <- window.mcmc.list(mcmc_list, start = 500, end = 1000)

}
